Fill in missing rows using enter and exit functions

This function figures out which rows are missing in either state and applies the provided enter and exit functions to fill in the blanks and provide a 1-to-1 relation between the rows in from and to.

.complete_states(from, to, id, enter, exit, max_id)

Arguments

from, to
Data.frames to tween between

id
The name of the column that holds the matching id

enter, exit
functions to fill out missing rows in from and to respectively

Value

A list with the elements from and to holding the filled out versions of from and to