.get_last_frame {tweenr}	R Documentation

Helpers for working with tweened data

Description

These are internal helpers for extracting and inserting data into a data.frame of tweened states.

Usage

.get_last_frame(data)

.get_first_frame(data)

.with_prior_frames(prior, new_tween, nframes)

.with_later_frames(later, new_tween, nframes)

.has_frames(data)

Arguments

data, prior, later
A data.frame. If a .frame column exists it will be interpreted as a data.frame containing multiple states

new_tween
The result of a tweening

Value

A data.frame

[Package tweenr version 1.0.1 Index]