Transition in and out of events

This tweening function is a more powerful version of tween_appear(), with support for newer features such as enter/exits and tween phase identification. The tweener treats each row in the data as unique events in time, and creates frames with the correct events present at any given time.

tween_events(
  .data,
  ease,
  nframes,
  start,
  end = NULL,
  range = NULL,
  enter = NULL,
  exit = NULL,
  enter_length = 0,
  exit_length = 0
)

Arguments

.data

A data.frame with components at different stages

ease

The easing function to use. Either a single string or one for each column in the data set.

nframes

The number of frames to calculate for the tween

start, end

The start (and potential end) of the event encoded in the row, as unquoted expressions. Will be evaluated in the context of .data so can refer to columns in it. If end = NULL the event will be without extend and only visible in a single frame, unless enter and/or exit is given.

range

The range of time points to include in the tween. If NULL it will use the range of time

enter, exit

functions that calculate a start state for new observations that appear in to or an end state for observations that are not present in to. If NULL the new/old observations will not be part of the tween. The function gets a data.frame with either the start state of the exiting observations, or the end state of the entering observations and must return a modified version of that data.frame. See the Match, Enter, and Exit section for more information.

enter_length, exit_length

The lenght of the opening and closing transitions if enter and/or exit is given. Measured in the same units as time

Value

A data.frame with the same columns as .data along with .id giving the component id, .phase giving the state of each component in each frame, and .frame giving the frame membership of each row.

See also

Examples

d <- data.frame(
  x = runif(20),
  y = runif(20),
  time = runif(20),
  duration = runif(20, max = 0.1)
)
from_left <- function(x) {
  x$x <- -0.5
  x
}
to_right <- function(x) {
  x$x <- 1.5
  x
}

tween_events(d, 'cubic-in-out', 50, start = time, end = time + duration,
             enter = from_left, exit = to_right, enter_length = 0.1,
             exit_length = 0.05)
#>                x          y       time    duration .id .phase .frame
#> 119 -0.500000000 0.21095708 0.02398098 0.072015348  14  enter      1
#> 35  -0.500000000 0.02228596 0.04371294 0.098423430   5  enter      2
#> 120 -0.408775217 0.21095708 0.02398098 0.072015348  14  enter      2
#> 36  -0.464116482 0.02228596 0.04371294 0.098423430   5  enter      3
#> 121  0.229798265 0.21095708 0.02398098 0.072015348  14  enter      3
#> 37  -0.212931853 0.02228596 0.04371294 0.098423430   5  enter      4
#> 122  0.868371747 0.21095708 0.02398098 0.072015348  14  enter      4
#> 38   0.038252776 0.02228596 0.04371294 0.098423430   5  enter      5
#> 73  -0.500000000 0.59308743 0.11763830 0.020662604   9  enter      5
#> 123  0.959596530 0.21095708 0.02398098 0.072015348  14 static      5
#> 147 -0.500000000 0.33837736 0.12862960 0.073066090  17  enter      5
#> 39   0.074136295 0.02228596 0.04371294 0.098423430   5 static      6
#> 74  -0.437250324 0.59308743 0.11763830 0.020662604   9  enter      6
#> 124  0.959596530 0.21095708 0.02398098 0.072015348  14 static      6
#> 148 -0.460529646 0.33837736 0.12862960 0.073066090  17  enter      6
#> 40   0.074136295 0.02228596 0.04371294 0.098423430   5 static      7
#> 75   0.001997407 0.59308743 0.11763830 0.020662604   9  enter      7
#> 125  0.959596530 0.21095708 0.02398098 0.072015348  14 static      7
#> 149 -0.184237167 0.33837736 0.12862960 0.073066090  17  enter      7
#> 41   0.074136295 0.02228596 0.04371294 0.098423430   5 static      8
#> 76   0.441245138 0.59308743 0.11763830 0.020662604   9  enter      8
#> 126  0.959596530 0.21095708 0.02398098 0.072015348  14    raw      8
#> 150  0.417685733 0.33837736 0.12862960 0.073066090  17  enter      8
#> 158 -0.500000000 0.20541942 0.18202781 0.023581285  18  enter      8
#> 42   0.074136295 0.02228596 0.04371294 0.098423430   5 static      9
#> 77   0.503994813 0.59308743 0.11763830 0.020662604   9 static      9
#> 127  1.229798265 0.21095708 0.02398098 0.072015348  14   exit      9
#> 151  0.693978211 0.33837736 0.12862960 0.073066090  17  enter      9
#> 159 -0.462742144 0.20541942 0.18202781 0.023581285  18  enter      9
#> 43   0.074136295 0.02228596 0.04371294 0.098423430   5    raw     10
#> 46  -0.500000000 0.24330205 0.23831459 0.066831222   6  enter     10
#> 78   0.503994813 0.59308743 0.11763830 0.020662604   9    raw     10
#> 109 -0.500000000 0.90268608 0.22884401 0.068729136  13  enter     10
#> 128  1.500000000 0.21095708 0.02398098 0.072015348  14   exit     10
#> 152  0.733448565 0.33837736 0.12862960 0.073066090  17 static     10
#> 160 -0.201937154 0.20541942 0.18202781 0.023581285  18  enter     10
#> 44   0.787068147 0.02228596 0.04371294 0.098423430   5   exit     11
#> 47  -0.448858622 0.24330205 0.23831459 0.066831222   6  enter     11
#> 79   1.001997407 0.59308743 0.11763830 0.020662604   9   exit     11
#> 110 -0.438881337 0.90268608 0.22884401 0.068729136  13  enter     11
#> 153  0.733448565 0.33837736 0.12862960 0.073066090  17 static     11
#> 161  0.058867836 0.20541942 0.18202781 0.023581285  18  enter     11
#> 45   1.500000000 0.02228596 0.04371294 0.098423430   5   exit     12
#> 48  -0.090868972 0.24330205 0.23831459 0.066831222   6  enter     12
#> 80   1.500000000 0.59308743 0.11763830 0.020662604   9   exit     12
#> 111 -0.011050692 0.90268608 0.22884401 0.068729136  13  enter     12
#> 154  0.733448565 0.33837736 0.12862960 0.073066090  17 static     12
#> 162  0.096125692 0.20541942 0.18202781 0.023581285  18 static     12
#> 49   0.267120677 0.24330205 0.23831459 0.066831222   6  enter     13
#> 112  0.416779952 0.90268608 0.22884401 0.068729136  13  enter     13
#> 137 -0.500000000 0.30910952 0.32178244 0.054808992  16  enter     13
#> 155  0.733448565 0.33837736 0.12862960 0.073066090  17    raw     13
#> 163  0.096125692 0.20541942 0.18202781 0.023581285  18    raw     13
#> 50   0.318262056 0.24330205 0.23831459 0.066831222   6 static     14
#> 113  0.477898615 0.90268608 0.22884401 0.068729136  13 static     14
#> 138 -0.458486529 0.30910952 0.32178244 0.054808992  16  enter     14
#> 156  1.116724283 0.33837736 0.12862960 0.073066090  17   exit     14
#> 164  0.798062846 0.20541942 0.18202781 0.023581285  18   exit     14
#> 51   0.318262056 0.24330205 0.23831459 0.066831222   6 static     15
#> 114  0.477898615 0.90268608 0.22884401 0.068729136  13 static     15
#> 139 -0.167892230 0.30910952 0.32178244 0.054808992  16  enter     15
#> 157  1.500000000 0.33837736 0.12862960 0.073066090  17   exit     15
#> 165  1.500000000 0.20541942 0.18202781 0.023581285  18   exit     15
#> 8   -0.500000000 0.11112984 0.38218939 0.061043941   2  enter     16
#> 52   0.318262056 0.24330205 0.23831459 0.066831222   6 static     16
#> 115  0.477898615 0.90268608 0.22884401 0.068729136  13 static     16
#> 129 -0.500000000 0.88724044 0.37494594 0.031198547  15  enter     16
#> 140  0.465188206 0.30910952 0.32178244 0.054808992  16  enter     16
#> 9   -0.466178755 0.11112984 0.38218939 0.061043941   2  enter     17
#> 53   0.318262056 0.24330205 0.23831459 0.066831222   6    raw     17
#> 116  0.477898615 0.90268608 0.22884401 0.068729136  13    raw     17
#> 130 -0.445287910 0.88724044 0.37494594 0.031198547  15  enter     17
#> 141  0.755782505 0.30910952 0.32178244 0.054808992  16  enter     17
#> 10  -0.229430040 0.11112984 0.38218939 0.061043941   2  enter     18
#> 54   0.909131028 0.24330205 0.23831459 0.066831222   6   exit     18
#> 65  -0.500000000 0.74608345 0.43688645 0.006062876   8  enter     18
#> 117  0.988949308 0.90268608 0.22884401 0.068729136  13   exit     18
#> 131 -0.062303277 0.88724044 0.37494594 0.031198547  15  enter     18
#> 142  0.797295976 0.30910952 0.32178244 0.054808992  16 static     18
#> 174 -0.500000000 0.47571161 0.43045295 0.048877007  20  enter     18
#> 11   0.007318675 0.11112984 0.38218939 0.061043941   2  enter     19
#> 55   1.500000000 0.24330205 0.23831459 0.066831222   6   exit     19
#> 66  -0.469496457 0.74608345 0.43688645 0.006062876   8  enter     19
#> 118  1.500000000 0.90268608 0.22884401 0.068729136  13   exit     19
#> 132  0.320681355 0.88724044 0.37494594 0.031198547  15  enter     19
#> 143  0.797295976 0.30910952 0.32178244 0.054808992  16 static     19
#> 175 -0.428325875 0.47571161 0.43045295 0.048877007  20  enter     19
#> 12   0.041139920 0.11112984 0.38218939 0.061043941   2 static     20
#> 67  -0.255971658 0.74608345 0.43688645 0.006062876   8  enter     20
#> 133  0.375393445 0.88724044 0.37494594 0.031198547  15 static     20
#> 144  0.797295976 0.30910952 0.32178244 0.054808992  16    raw     20
#> 176  0.073393000 0.47571161 0.43045295 0.048877007  20  enter     20
#> 13   0.041139920 0.11112984 0.38218939 0.061043941   2 static     21
#> 68   0.209207370 0.74608345 0.43688645 0.006062876   8  enter     21
#> 134  0.375393445 0.88724044 0.37494594 0.031198547  15    raw     21
#> 145  1.148647988 0.30910952 0.32178244 0.054808992  16   exit     21
#> 177  0.575111875 0.47571161 0.43045295 0.048877007  20  enter     21
#> 14   0.041139920 0.11112984 0.38218939 0.061043941   2 static     22
#> 69   0.422732169 0.74608345 0.43688645 0.006062876   8  enter     22
#> 92  -0.500000000 0.28514394 0.52445587 0.009480935  11  enter     22
#> 135  0.937696723 0.88724044 0.37494594 0.031198547  15   exit     22
#> 146  1.500000000 0.30910952 0.32178244 0.054808992  16   exit     22
#> 178  0.646786000 0.47571161 0.43045295 0.048877007  20 static     22
#> 15   0.041139920 0.11112984 0.38218939 0.061043941   2    raw     23
#> 70   0.453235712 0.74608345 0.43688645 0.006062876   8    raw     23
#> 81  -0.500000000 0.05633473 0.54183985 0.097378625  10  enter     23
#> 93  -0.440171906 0.28514394 0.52445587 0.009480935  11  enter     23
#> 136  1.500000000 0.88724044 0.37494594 0.031198547  15   exit     23
#> 179  0.646786000 0.47571161 0.43045295 0.048877007  20 static     23
#> 16   0.770569960 0.11112984 0.38218939 0.061043941   2   exit     24
#> 71   0.976617856 0.74608345 0.43688645 0.006062876   8   exit     24
#> 82  -0.446034842 0.05633473 0.54183985 0.097378625  10  enter     24
#> 94  -0.021375251 0.28514394 0.52445587 0.009480935  11  enter     24
#> 180  0.646786000 0.47571161 0.43045295 0.048877007  20    raw     24
#> 17   1.500000000 0.11112984 0.38218939 0.061043941   2   exit     25
#> 18  -0.500000000 0.85229086 0.60724524 0.012420505   3  enter     25
#> 72   1.500000000 0.74608345 0.43688645 0.006062876   8   exit     25
#> 83  -0.068278737 0.05633473 0.54183985 0.097378625  10  enter     25
#> 95   0.397421405 0.28514394 0.52445587 0.009480935  11  enter     25
#> 181  1.073393000 0.47571161 0.43045295 0.048877007  20   exit     25
#> 19  -0.458930233 0.85229086 0.60724524 0.012420505   3  enter     26
#> 84   0.309477368 0.05633473 0.54183985 0.097378625  10  enter     26
#> 96   0.457249499 0.28514394 0.52445587 0.009480935  11 static     26
#> 182  1.500000000 0.47571161 0.43045295 0.048877007  20   exit     26
#> 20  -0.171441866 0.85229086 0.60724524 0.012420505   3  enter     27
#> 85   0.363442526 0.05633473 0.54183985 0.097378625  10 static     27
#> 97   0.457249499 0.28514394 0.52445587 0.009480935  11    raw     27
#> 21   0.454872077 0.85229086 0.60724524 0.012420505   3  enter     28
#> 26  -0.500000000 0.52933187 0.66016210 0.053824663   4  enter     28
#> 86   0.363442526 0.05633473 0.54183985 0.097378625  10 static     28
#> 98   0.978624749 0.28514394 0.52445587 0.009480935  11   exit     28
#> 22   0.742360444 0.85229086 0.60724524 0.012420505   3  enter     29
#> 27  -0.440810251 0.52933187 0.66016210 0.053824663   4  enter     29
#> 87   0.363442526 0.05633473 0.54183985 0.097378625  10 static     29
#> 99   1.500000000 0.28514394 0.52445587 0.009480935  11   exit     29
#> 1   -0.500000000 0.36202673 0.71125475 0.002084538   1  enter     30
#> 23   0.783430211 0.85229086 0.60724524 0.012420505   3    raw     30
#> 28  -0.026482007 0.52933187 0.66016210 0.053824663   4  enter     30
#> 56  -0.500000000 0.59604943 0.72197279 0.032280196   7  enter     30
#> 88   0.363442526 0.05633473 0.54183985 0.097378625  10 static     30
#> 2   -0.418363280 0.36202673 0.71125475 0.002084538   1  enter     31
#> 24   1.141715106 0.85229086 0.60724524 0.012420505   3   exit     31
#> 29   0.387846238 0.52933187 0.66016210 0.053824663   4  enter     31
#> 57  -0.408290180 0.59604943 0.72197279 0.032280196   7  enter     31
#> 89   0.363442526 0.05633473 0.54183985 0.097378625  10    raw     31
#> 3    0.153093761 0.36202673 0.71125475 0.002084538   1  enter     32
#> 25   1.500000000 0.85229086 0.60724524 0.012420505   3   exit     32
#> 30   0.447035987 0.52933187 0.66016210 0.053824663   4 static     32
#> 58   0.233678559 0.59604943 0.72197279 0.032280196   7  enter     32
#> 90   0.931721263 0.05633473 0.54183985 0.097378625  10   exit     32
#> 4    0.724550802 0.36202673 0.71125475 0.002084538   1  enter     33
#> 31   0.447035987 0.52933187 0.66016210 0.053824663   4 static     33
#> 59   0.875647298 0.59604943 0.72197279 0.032280196   7  enter     33
#> 91   1.500000000 0.05633473 0.54183985 0.097378625  10   exit     33
#> 166 -0.500000000 0.14347877 0.78463630 0.019186806  19  enter     33
#> 5    0.806187522 0.36202673 0.71125475 0.002084538   1    raw     34
#> 32   0.447035987 0.52933187 0.66016210 0.053824663   4    raw     34
#> 60   0.967357118 0.59604943 0.72197279 0.032280196   7 static     34
#> 167 -0.451580793 0.14347877 0.78463630 0.019186806  19  enter     34
#> 6    1.153093761 0.36202673 0.71125475 0.002084538   1   exit     35
#> 33   0.973517993 0.52933187 0.66016210 0.053824663   4   exit     35
#> 61   0.967357118 0.59604943 0.72197279 0.032280196   7 static     35
#> 168 -0.112646346 0.14347877 0.78463630 0.019186806  19  enter     35
#> 7    1.500000000 0.36202673 0.71125475 0.002084538   1   exit     36
#> 34   1.500000000 0.52933187 0.66016210 0.053824663   4   exit     36
#> 62   0.967357118 0.59604943 0.72197279 0.032280196   7    raw     36
#> 169  0.226288101 0.14347877 0.78463630 0.019186806  19  enter     36
#> 63   1.233678559 0.59604943 0.72197279 0.032280196   7   exit     37
#> 170  0.274707308 0.14347877 0.78463630 0.019186806  19 static     37
#> 64   1.500000000 0.59604943 0.72197279 0.032280196   7   exit     38
#> 171  0.274707308 0.14347877 0.78463630 0.019186806  19    raw     38
#> 172  0.887353654 0.14347877 0.78463630 0.019186806  19   exit     39
#> 173  1.500000000 0.14347877 0.78463630 0.019186806  19   exit     40
#> 100 -0.500000000 0.16477394 0.99995099 0.042767192  12  enter     42
#> 101 -0.427878479 0.16477394 0.99995099 0.042767192  12  enter     43
#> 102  0.076972168 0.16477394 0.99995099 0.042767192  12  enter     44
#> 103  0.581822815 0.16477394 0.99995099 0.042767192  12  enter     45
#> 104  0.653944336 0.16477394 0.99995099 0.042767192  12 static     46
#> 105  0.653944336 0.16477394 0.99995099 0.042767192  12 static     47
#> 106  0.653944336 0.16477394 0.99995099 0.042767192  12    raw     48
#> 107  1.076972168 0.16477394 0.99995099 0.042767192  12   exit     49
#> 108  1.500000000 0.16477394 0.99995099 0.042767192  12   exit     50