Get a specific position between two states

This tween allows you to query a specific postion between two states rather than generate evenly spaced states. It can work with either data.frames or single vectors and each row/element can have its own position and easing.

tween_at(from, to, at, ease)

Arguments

from, to

A data.frame or vector of the same type. If either is of length/nrow 1 it will get repeated to match the length of the other

at

A numeric between 0 and 1 recycled to match the nrow/length of from

ease

A character vector giving valid easing functions. Recycled to match the ncol of from

Value

If from/to is a data.frame then a data.frame with the same columns. If from/to is a vector then a vector.

Examples

tween_at(mtcars[1:6, ], mtcars[6:1, ], runif(6), 'cubic-in-out')
#>        mpg      cyl     disp       hp     drat       wt     qsec        vs
#> 1 18.20318 6.000000 222.6874 105.1779 2.800560 3.430114 20.08622 0.9644211
#> 2 18.70679 7.994093 359.4093 174.8080 3.152215 3.438331 17.02000 0.0000000
#> 3 22.03155 5.097787 190.3340 102.3312 3.427352 2.811260 19.06558 1.0000000
#> 4 21.47716 5.889767 249.7325 109.0630 3.122440 3.165671 19.39425 1.0000000
#> 5 19.44400 7.353041 295.3041 153.9738 3.392609 3.257234 17.02000 0.0000000
#> 6 20.26835 6.000000 176.3991 108.7385 3.612385 2.831927 17.40863 0.2522939
#>            am     gear     carb
#> 1 0.035578938 3.035579 1.106737
#> 2 0.002953658 3.002954 2.005907
#> 3 0.451106664 3.451107 1.000000
#> 4 0.055116511 3.055117 1.000000
#> 5 0.323479298 3.323479 2.646959
#> 6 0.747706088 3.747706 3.243118