Get a specific position between two states

This tween allows you to query a specific postion between two states rather than generate evenly spaced states. It can work with either data.frames or single vectors and each row/element can have its own position and easing.

tween_at(from, to, at, ease)

Arguments

from, to

A data.frame or vector of the same type. If either is of length/nrow 1 it will get repeated to match the length of the other

at

A numeric between 0 and 1 recycled to match the nrow/length of from

ease

A character vector giving valid easing functions. Recycled to match the ncol of from

Value

If from/to is a data.frame then a data.frame with the same columns. If from/to is a vector then a vector.

Examples

tween_at(mtcars[1:6, ], mtcars[6:1, ], runif(6), 'cubic-in-out')
#>        mpg      cyl     disp       hp     drat       wt     qsec         vs
#> 1 20.99601 6.000000 160.0893 109.9931 3.898433 2.621155 16.46517 0.00137454
#> 2 20.98063 6.016846 161.6846 110.5475 3.893683 2.879759 17.02000 0.00000000
#> 3 21.41032 5.985256 256.8942 109.8747 3.085677 3.208402 19.43388 1.00000000
#> 4 22.11542 4.977973 181.3480 101.3128 3.473480 2.757643 19.01586 1.00000000
#> 5 20.26702 6.637378 223.7378 130.7148 3.660983 3.055059 17.02000 0.00000000
#> 6 18.42368 6.000000 217.7452 105.5581 2.887238 3.366246 19.80034 0.88838769
#>           am     gear     carb
#> 1 0.99862546 3.998625 3.995876
#> 2 0.99157722 3.991577 3.983154
#> 3 0.00737218 3.007372 1.000000
#> 4 0.51101344 3.511013 1.000000
#> 5 0.68131099 3.681311 3.362622
#> 6 0.11161231 3.111612 1.334837