Get a specific position between two states

This tween allows you to query a specific postion between two states rather than generate evenly spaced states. It can work with either data.frames or single vectors and each row/element can have its own position and easing.

tween_at(from, to, at, ease)

Arguments

from, to
A data.frame or vector of the same type. If either is of length/nrow 1 it will get repeated to match the length of the other

at
A numeric between 0 and 1 recycled to match the nrow/length of from

ease
A character vector giving valid easing functions. Recycled to match the ncol of from

Value

If from/to is a data.frame then a data.frame with the same columns. If from/to is a vector then a vector.

Examples

tween_at(mtcars[1:6, ], mtcars[6:1, ], runif(6), 'cubic-in-out')

#>        mpg      cyl     disp       hp     drat       wt     qsec         vs
#> 1 20.16513 6.000000 178.7125 108.5606 3.571811 2.861824 17.54245 0.28788514
#> 2 20.22084 6.677530 227.7530 132.0197 3.645926 3.066402 17.02000 0.00000000
#> 3 22.76603 4.048530 111.6397  93.4125 3.831316 2.341717 18.63014 1.00000000
#> 4 21.40269 5.996155 257.7117 109.9673 3.081480 3.213280 19.43840 1.00000000
#> 5 18.70005 7.999955 359.9955 174.9985 3.150017 3.439987 17.02000 0.00000000
#> 6 20.76703 6.000000 165.2218 109.5983 3.808418 2.687481 16.76206 0.08033492
#>             am     gear     carb
#> 1 7.121149e-01 3.712115 3.136345
#> 2 6.612349e-01 3.661235 3.322470
#> 3 9.757352e-01 3.975735 1.000000
#> 4 1.922275e-03 3.001922 1.000000
#> 5 2.244497e-05 3.000022 2.000045
#> 6 9.196651e-01 3.919665 3.758995