Get a specific position between two states

This tween allows you to query a specific postion between two states rather than generate evenly spaced states. It can work with either data.frames or single vectors and each row/element can have its own position and easing.

tween_at(from, to, at, ease)

Arguments

from, to

A data.frame or vector of the same type. If either is of length/nrow 1 it will get repeated to match the length of the other

at

A numeric between 0 and 1 recycled to match the nrow/length of from

ease

A character vector giving valid easing functions. Recycled to match the ncol of from

Value

If from/to is a data.frame then a data.frame with the same columns. If from/to is a vector then a vector.

Examples

tween_at(mtcars[1:6, ], mtcars[6:1, ], runif(6), 'cubic-in-out')
#>        mpg      cyl     disp       hp     drat       wt     qsec        vs
#> 1 18.24010 6.000000 221.8599 105.2415 2.815073 3.419420 20.03836 0.9516906
#> 2 20.97054 6.025617 162.5617 110.8325 3.890394 2.882237 17.02000 0.0000000
#> 3 21.52145 5.826505 244.9878 108.5253 3.146796 3.137361 19.36800 1.0000000
#> 4 21.40116 5.998345 257.8759 109.9859 3.080637 3.214260 19.43931 1.0000000
#> 5 18.72551 7.977817 357.7817 174.2791 3.158319 3.433733 17.02000 0.0000000
#> 6 18.12265 6.000000 224.4924 105.0390 2.768903 3.453440 20.19064 0.9921908
#>             am     gear     carb
#> 1 0.0483094280 3.048309 1.144928
#> 2 0.9871916747 3.987192 3.974383
#> 3 0.0867477183 3.086748 1.000000
#> 4 0.0008273677 3.000827 1.000000
#> 5 0.0110914040 3.011091 2.022183
#> 6 0.0078092435 3.007809 1.023428