Get a specific position between two states

This tween allows you to query a specific postion between two states rather than generate evenly spaced states. It can work with either data.frames or single vectors and each row/element can have its own position and easing.

tween_at(from, to, at, ease)

Arguments

from, to

A data.frame or vector of the same type. If either is of length/nrow 1 it will get repeated to match the length of the other

at

A numeric between 0 and 1 recycled to match the nrow/length of from

ease

A character vector giving valid easing functions. Recycled to match the ncol of from

Value

If from/to is a data.frame then a data.frame with the same columns. If from/to is a vector then a vector.

Examples

tween_at(mtcars[1:6, ], mtcars[6:1, ], runif(6), 'cubic-in-out')
#>        mpg      cyl     disp       hp     drat       wt     qsec        vs
#> 1 19.02874 6.000000 204.1834 106.6013 3.125091 3.190985 19.01584 0.6797445
#> 2 18.74715 7.959002 355.9002 173.6676 3.165374 3.428418 17.02000 0.0000000
#> 3 21.80754 5.417804 214.3353 105.0513 3.304146 2.954467 19.19839 1.0000000
#> 4 22.03013 5.099817 190.4863 102.3484 3.426570 2.812168 19.06642 1.0000000
#> 5 19.82984 7.017533 261.7533 143.0698 3.518425 3.162453 17.02000 0.0000000
#> 6 20.62687 6.000000 168.3633 109.3567 3.753320 2.728080 16.94379 0.1286669
#>           am     gear     carb
#> 1 0.32025552 3.320256 1.960767
#> 2 0.02049916 3.020499 2.040998
#> 3 0.29109820 3.291098 1.000000
#> 4 0.45009146 3.450091 1.000000
#> 5 0.49123365 3.491234 2.982467
#> 6 0.87133312 3.871333 3.613999