Transition in and out of events

This tweening function is a more powerful version of tween_appear(), with support for newer features such as enter/exits and tween phase identification. The tweener treats each row in the data as unique events in time, and creates frames with the correct events present at any given time.

tween_events(
  .data,
  ease,
  nframes,
  start,
  end = NULL,
  range = NULL,
  enter = NULL,
  exit = NULL,
  enter_length = 0,
  exit_length = 0
)

Arguments

.data

A data.frame with components at different stages

ease

The easing function to use. Either a single string or one for each column in the data set.

nframes

The number of frames to calculate for the tween

start, end

The start (and potential end) of the event encoded in the row, as unquoted expressions. Will be evaluated in the context of .data so can refer to columns in it. If end = NULL the event will be without extend and only visible in a single frame, unless enter and/or exit is given.

range

The range of time points to include in the tween. If NULL it will use the range of time

enter

functions that calculate a start state for new observations that appear in to or an end state for observations that are not present in to. If NULL the new/old observations will not be part of the tween. The function gets a data.frame with either the start state of the exiting observations, or the end state of the entering observations and must return a modified version of that data.frame. See the Match, Enter, and Exit section for more information.

exit

functions that calculate a start state for new observations that appear in to or an end state for observations that are not present in to. If NULL the new/old observations will not be part of the tween. The function gets a data.frame with either the start state of the exiting observations, or the end state of the entering observations and must return a modified version of that data.frame. See the Match, Enter, and Exit section for more information.

enter_length

The lenght of the opening and closing transitions if enter and/or exit is given. Measured in the same units as time

exit_length

The lenght of the opening and closing transitions if enter and/or exit is given. Measured in the same units as time

Value

A data.frame with the same columns as .data along with .id giving the component id, .phase giving the state of each component in each frame, and .frame giving the frame membership of each row.

See also

Examples

d <- data.frame(
  x = runif(20),
  y = runif(20),
  time = runif(20),
  duration = runif(20, max = 0.1)
)
from_left <- function(x) {
  x$x <- -0.5
  x
}
to_right <- function(x) {
  x$x <- 1.5
  x
}

tween_events(d, 'cubic-in-out', 50, start = time, end = time + duration,
             enter = from_left, exit = to_right, enter_length = 0.1,
             exit_length = 0.05)
#>               x          y       time    duration .id .phase .frame
#> 155 -0.50000000 0.34680064 0.03690889 0.094270142  18  enter      1
#> 128 -0.50000000 0.34095137 0.06532939 0.035082030  15  enter      2
#> 156 -0.45272849 0.34680064 0.03690889 0.094270142  18  enter      2
#> 99  -0.50000000 0.20578233 0.08620800 0.038630515  12  enter      3
#> 129 -0.45682074 0.34095137 0.06532939 0.035082030  15  enter      3
#> 157 -0.12182795 0.34680064 0.03690889 0.094270142  18  enter      3
#> 100 -0.42484351 0.20578233 0.08620800 0.038630515  12  enter      4
#> 130 -0.15456593 0.34095137 0.06532939 0.035082030  15  enter      4
#> 158  0.20907260 0.34680064 0.03690889 0.094270142  18  enter      4
#> 101  0.10125190 0.20578233 0.08620800 0.038630515  12  enter      5
#> 131  0.14768889 0.34095137 0.06532939 0.035082030  15  enter      5
#> 159  0.25634410 0.34680064 0.03690889 0.094270142  18 static      5
#> 166 -0.50000000 0.68480356 0.12704018 0.088773059  19  enter      5
#> 38  -0.50000000 0.83502867 0.14964807 0.001907543   5  enter      6
#> 102  0.62734731 0.20578233 0.08620800 0.038630515  12  enter      6
#> 132  0.19086815 0.34095137 0.06532939 0.035082030  15 static      6
#> 160  0.25634410 0.34680064 0.03690889 0.094270142  18 static      6
#> 167 -0.43609883 0.68480356 0.12704018 0.088773059  19  enter      6
#> 39  -0.44896357 0.83502867 0.14964807 0.001907543   5  enter      7
#> 103  0.70250380 0.20578233 0.08620800 0.038630515  12 static      7
#> 133  0.19086815 0.34095137 0.06532939 0.035082030  15 static      7
#> 161  0.25634410 0.34680064 0.03690889 0.094270142  18 static      7
#> 168  0.01120937 0.68480356 0.12704018 0.088773059  19  enter      7
#> 40  -0.09170858 0.83502867 0.14964807 0.001907543   5  enter      8
#> 104  0.70250380 0.20578233 0.08620800 0.038630515  12 static      8
#> 108 -0.50000000 0.10678549 0.20336838 0.039068028  13  enter      8
#> 134  0.19086815 0.34095137 0.06532939 0.035082030  15    raw      8
#> 162  0.25634410 0.34680064 0.03690889 0.094270142  18 static      8
#> 169  0.45851757 0.68480356 0.12704018 0.088773059  19  enter      8
#> 41   0.26554641 0.83502867 0.14964807 0.001907543   5  enter      9
#> 105  0.70250380 0.20578233 0.08620800 0.038630515  12    raw      9
#> 109 -0.43436134 0.10678549 0.20336838 0.039068028  13  enter      9
#> 135  0.84543407 0.34095137 0.06532939 0.035082030  15   exit      9
#> 163  0.25634410 0.34680064 0.03690889 0.094270142  18    raw      9
#> 170  0.52241874 0.68480356 0.12704018 0.088773059  19 static      9
#> 42   0.31658284 0.83502867 0.14964807 0.001907543   5    raw     10
#> 106  1.10125190 0.20578233 0.08620800 0.038630515  12   exit     10
#> 110  0.02510932 0.10678549 0.20336838 0.039068028  13  enter     10
#> 136  1.50000000 0.34095137 0.06532939 0.035082030  15   exit     10
#> 164  0.87817205 0.34680064 0.03690889 0.094270142  18   exit     10
#> 171  0.52241874 0.68480356 0.12704018 0.088773059  19 static     10
#> 43   0.90829142 0.83502867 0.14964807 0.001907543   5   exit     11
#> 107  1.50000000 0.20578233 0.08620800 0.038630515  12   exit     11
#> 111  0.48457997 0.10678549 0.20336838 0.039068028  13  enter     11
#> 165  1.50000000 0.34680064 0.03690889 0.094270142  18   exit     11
#> 172  0.52241874 0.68480356 0.12704018 0.088773059  19 static     11
#> 44   1.50000000 0.83502867 0.14964807 0.001907543   5   exit     12
#> 112  0.55021863 0.10678549 0.20336838 0.039068028  13 static     12
#> 173  0.52241874 0.68480356 0.12704018 0.088773059  19 static     12
#> 61  -0.50000000 0.35600469 0.30749381 0.061284459   8  enter     13
#> 71  -0.50000000 0.02266115 0.30702645 0.033745463   9  enter     13
#> 113  0.55021863 0.10678549 0.20336838 0.039068028  13 static     13
#> 137 -0.50000000 0.93236644 0.32695617 0.010190330  16  enter     13
#> 174  0.52241874 0.68480356 0.12704018 0.088773059  19    raw     13
#> 62  -0.41654544 0.35600469 0.30749381 0.061284459   8  enter     14
#> 72  -0.40850095 0.02266115 0.30702645 0.033745463   9  enter     14
#> 114  0.55021863 0.10678549 0.20336838 0.039068028  13    raw     14
#> 138 -0.47782944 0.93236644 0.32695617 0.010190330  16  enter     14
#> 175  1.01120937 0.68480356 0.12704018 0.088773059  19   exit     14
#> 1   -0.50000000 0.67999756 0.35862014 0.026998859   1  enter     15
#> 63   0.16763644 0.35600469 0.30749381 0.061284459   8  enter     15
#> 73   0.23199244 0.02266115 0.30702645 0.033745463   9  enter     15
#> 115  1.02510932 0.10678549 0.20336838 0.039068028  13   exit     15
#> 139 -0.32263552 0.93236644 0.32695617 0.010190330  16  enter     15
#> 176  1.50000000 0.68480356 0.12704018 0.088773059  19   exit     15
#> 2   -0.41310709 0.67999756 0.35862014 0.026998859   1  enter     16
#> 64   0.75181833 0.35600469 0.30749381 0.061284459   8  enter     16
#> 74   0.87248582 0.02266115 0.30702645 0.033745463   9  enter     16
#> 116  1.50000000 0.10678549 0.20336838 0.039068028  13   exit     16
#> 140  0.01546553 0.93236644 0.32695617 0.010190330  16  enter     16
#> 3    0.19514331 0.67999756 0.35862014 0.026998859   1  enter     17
#> 65   0.83527288 0.35600469 0.30749381 0.061284459   8 static     17
#> 75   0.96398488 0.02266115 0.30702645 0.033745463   9 static     17
#> 141  0.17065946 0.93236644 0.32695617 0.010190330  16  enter     17
#> 4    0.80339370 0.67999756 0.35862014 0.026998859   1  enter     18
#> 66   0.83527288 0.35600469 0.30749381 0.061284459   8 static     18
#> 76   0.96398488 0.02266115 0.30702645 0.033745463   9    raw     18
#> 142  0.19283002 0.93236644 0.32695617 0.010190330  16    raw     18
#> 5    0.89028661 0.67999756 0.35862014 0.026998859   1 static     19
#> 67   0.83527288 0.35600469 0.30749381 0.061284459   8    raw     19
#> 77   1.23199244 0.02266115 0.30702645 0.033745463   9   exit     19
#> 143  0.84641501 0.93236644 0.32695617 0.010190330  16   exit     19
#> 6    0.89028661 0.67999756 0.35862014 0.026998859   1    raw     20
#> 68   0.93375097 0.35600469 0.30749381 0.061284459   8   exit     20
#> 78   1.50000000 0.02266115 0.30702645 0.033745463   9   exit     20
#> 144  1.50000000 0.93236644 0.32695617 0.010190330  16   exit     20
#> 7    1.19514331 0.67999756 0.35862014 0.026998859   1   exit     21
#> 45  -0.50000000 0.49856452 0.49961322 0.042655511   6  enter     21
#> 54  -0.50000000 0.33262318 0.50322941 0.001654428   7  enter     21
#> 69   1.40152191 0.35600469 0.30749381 0.061284459   8   exit     21
#> 8    1.50000000 0.67999756 0.35862014 0.026998859   1   exit     22
#> 20  -0.50000000 0.21966989 0.52377071 0.017158516   3  enter     22
#> 28  -0.50000000 0.93785525 0.52928933 0.062967270   4  enter     22
#> 46  -0.44925375 0.49856452 0.49961322 0.042655511   6  enter     22
#> 55  -0.46276826 0.33262318 0.50322941 0.001654428   7  enter     22
#> 70   1.50000000 0.35600469 0.30749381 0.061284459   8   exit     22
#> 21  -0.43891372 0.21966989 0.52377071 0.017158516   3  enter     23
#> 29  -0.43093318 0.93785525 0.52928933 0.062967270   4  enter     23
#> 47  -0.09402998 0.49856452 0.49961322 0.042655511   6  enter     23
#> 56  -0.20214611 0.33262318 0.50322941 0.001654428   7  enter     23
#> 88  -0.50000000 0.99148400 0.55924831 0.066867294  11  enter     23
#> 22  -0.01130973 0.21966989 0.52377071 0.017158516   3  enter     24
#> 30   0.05253457 0.93785525 0.52928933 0.062967270   4  enter     24
#> 48   0.26119379 0.49856452 0.49961322 0.042655511   6  enter     24
#> 57   0.05847605 0.33262318 0.50322941 0.001654428   7  enter     24
#> 89  -0.46481849 0.99148400 0.55924831 0.066867294  11  enter     24
#> 23   0.41629425 0.21966989 0.52377071 0.017158516   3  enter     25
#> 31   0.53600232 0.93785525 0.52928933 0.062967270   4  enter     25
#> 49   0.31194004 0.49856452 0.49961322 0.042655511   6 static     25
#> 58   0.09570778 0.33262318 0.50322941 0.001654428   7    raw     25
#> 90  -0.21854793 0.99148400 0.55924831 0.066867294  11  enter     25
#> 145 -0.50000000 0.71110255 0.60962920 0.041941086  17  enter     25
#> 24   0.47738053 0.21966989 0.52377071 0.017158516   3 static     26
#> 32   0.60506915 0.93785525 0.52928933 0.062967270   4 static     26
#> 50   0.31194004 0.49856452 0.49961322 0.042655511   6 static     26
#> 59   0.79785389 0.33262318 0.50322941 0.001654428   7   exit     26
#> 79  -0.50000000 0.78947534 0.61815799 0.038379653  10  enter     26
#> 91   0.31797009 0.99148400 0.55924831 0.066867294  11  enter     26
#> 146 -0.46611008 0.71110255 0.60962920 0.041941086  17  enter     26
#> 25   0.47738053 0.21966989 0.52377071 0.017158516   3    raw     27
#> 33   0.60506915 0.93785525 0.52928933 0.062967270   4 static     27
#> 51   0.31194004 0.49856452 0.49961322 0.042655511   6    raw     27
#> 60   1.50000000 0.33262318 0.50322941 0.001654428   7   exit     27
#> 80  -0.43448717 0.78947534 0.61815799 0.038379653  10  enter     27
#> 92   0.56424065 0.99148400 0.55924831 0.066867294  11  enter     27
#> 117 -0.50000000 0.83135566 0.64077211 0.085202701  14  enter     27
#> 147 -0.22888062 0.71110255 0.60962920 0.041941086  17  enter     27
#> 26   0.98869027 0.21966989 0.52377071 0.017158516   3   exit     28
#> 34   0.60506915 0.93785525 0.52928933 0.062967270   4 static     28
#> 52   0.90597002 0.49856452 0.49961322 0.042655511   6   exit     28
#> 81   0.02410266 0.78947534 0.61815799 0.038379653  10  enter     28
#> 93   0.59942216 0.99148400 0.55924831 0.066867294  11 static     28
#> 118 -0.45253698 0.83135566 0.64077211 0.085202701  14  enter     28
#> 148  0.28794070 0.71110255 0.60962920 0.041941086  17  enter     28
#> 27   1.50000000 0.21966989 0.52377071 0.017158516   3   exit     29
#> 35   0.60506915 0.93785525 0.52928933 0.062967270   4    raw     29
#> 53   1.50000000 0.49856452 0.49961322 0.042655511   6   exit     29
#> 82   0.48269248 0.78947534 0.61815799 0.038379653  10  enter     29
#> 94   0.59942216 0.99148400 0.55924831 0.066867294  11 static     29
#> 119 -0.12029583 0.83135566 0.64077211 0.085202701  14  enter     29
#> 149  0.52517015 0.71110255 0.60962920 0.041941086  17  enter     29
#> 36   1.05253457 0.93785525 0.52928933 0.062967270   4   exit     30
#> 83   0.54820532 0.78947534 0.61815799 0.038379653  10 static     30
#> 95   0.59942216 0.99148400 0.55924831 0.066867294  11    raw     30
#> 120  0.21194531 0.83135566 0.64077211 0.085202701  14  enter     30
#> 150  0.55906008 0.71110255 0.60962920 0.041941086  17 static     30
#> 9   -0.50000000 0.13290852 0.74382621 0.082531787   2  enter     31
#> 37   1.50000000 0.93785525 0.52928933 0.062967270   4   exit     31
#> 84   0.54820532 0.78947534 0.61815799 0.038379653  10 static     31
#> 96   0.73284110 0.99148400 0.55924831 0.066867294  11   exit     31
#> 121  0.25940833 0.83135566 0.64077211 0.085202701  14 static     31
#> 151  0.55906008 0.71110255 0.60962920 0.041941086  17 static     31
#> 10  -0.44360880 0.13290852 0.74382621 0.082531787   2  enter     32
#> 85   0.54820532 0.78947534 0.61815799 0.038379653  10    raw     32
#> 97   1.36658106 0.99148400 0.55924831 0.066867294  11   exit     32
#> 122  0.25940833 0.83135566 0.64077211 0.085202701  14 static     32
#> 152  0.55906008 0.71110255 0.60962920 0.041941086  17    raw     32
#> 11  -0.04887039 0.13290852 0.74382621 0.082531787   2  enter     33
#> 86   1.02410266 0.78947534 0.61815799 0.038379653  10   exit     33
#> 98   1.50000000 0.99148400 0.55924831 0.066867294  11   exit     33
#> 123  0.25940833 0.83135566 0.64077211 0.085202701  14 static     33
#> 153  1.02953004 0.71110255 0.60962920 0.041941086  17   exit     33
#> 12   0.34586802 0.13290852 0.74382621 0.082531787   2  enter     34
#> 87   1.50000000 0.78947534 0.61815799 0.038379653  10   exit     34
#> 124  0.25940833 0.83135566 0.64077211 0.085202701  14 static     34
#> 154  1.50000000 0.71110255 0.60962920 0.041941086  17   exit     34
#> 13   0.40225922 0.13290852 0.74382621 0.082531787   2 static     35
#> 125  0.25940833 0.83135566 0.64077211 0.085202701  14    raw     35
#> 14   0.40225922 0.13290852 0.74382621 0.082531787   2 static     36
#> 126  0.87970417 0.83135566 0.64077211 0.085202701  14   exit     36
#> 15   0.40225922 0.13290852 0.74382621 0.082531787   2 static     37
#> 127  1.50000000 0.83135566 0.64077211 0.085202701  14   exit     37
#> 16   0.40225922 0.13290852 0.74382621 0.082531787   2 static     38
#> 17   0.40225922 0.13290852 0.74382621 0.082531787   2    raw     39
#> 177 -0.50000000 0.77965262 0.93498610 0.099813373  20  enter     39
#> 18   0.95112961 0.13290852 0.74382621 0.082531787   2   exit     40
#> 178 -0.46689630 0.77965262 0.93498610 0.099813373  20  enter     40
#> 19   1.50000000 0.13290852 0.74382621 0.082531787   2   exit     41
#> 179 -0.23517036 0.77965262 0.93498610 0.099813373  20  enter     41
#> 180  0.26966113 0.77965262 0.93498610 0.099813373  20  enter     42
#> 181  0.50138707 0.77965262 0.93498610 0.099813373  20  enter     43
#> 182  0.53449077 0.77965262 0.93498610 0.099813373  20 static     44
#> 183  0.53449077 0.77965262 0.93498610 0.099813373  20 static     45
#> 184  0.53449077 0.77965262 0.93498610 0.099813373  20 static     46
#> 185  0.53449077 0.77965262 0.93498610 0.099813373  20 static     47
#> 186  0.53449077 0.77965262 0.93498610 0.099813373  20    raw     48
#> 187  1.01724539 0.77965262 0.93498610 0.099813373  20   exit     49
#> 188  1.50000000 0.77965262 0.93498610 0.099813373  20   exit     50