Transition in and out of events

This tweening function is a more powerful version of tween_appear(), with support for newer features such as enter/exits and tween phase identification. The tweener treats each row in the data as unique events in time, and creates frames with the correct events present at any given time.

tween_events(
  .data,
  ease,
  nframes,
  start,
  end = NULL,
  range = NULL,
  enter = NULL,
  exit = NULL,
  enter_length = 0,
  exit_length = 0
)

Arguments

.data

A data.frame with components at different stages

ease

The easing function to use. Either a single string or one for each column in the data set.

nframes

The number of frames to calculate for the tween

start, end

The start (and potential end) of the event encoded in the row, as unquoted expressions. Will be evaluated in the context of .data so can refer to columns in it. If end = NULL the event will be without extend and only visible in a single frame, unless enter and/or exit is given.

range

The range of time points to include in the tween. If NULL it will use the range of time

enter, exit

functions that calculate a start state for new observations that appear in to or an end state for observations that are not present in to. If NULL the new/old observations will not be part of the tween. The function gets a data.frame with either the start state of the exiting observations, or the end state of the entering observations and must return a modified version of that data.frame. See the Match, Enter, and Exit section for more information.

enter_length, exit_length

The lenght of the opening and closing transitions if enter and/or exit is given. Measured in the same units as time

Value

A data.frame with the same columns as .data along with .id giving the component id, .phase giving the state of each component in each frame, and .frame giving the frame membership of each row.

See also

Examples

d <- data.frame(
  x = runif(20),
  y = runif(20),
  time = runif(20),
  duration = runif(20, max = 0.1)
)
from_left <- function(x) {
  x$x <- -0.5
  x
}
to_right <- function(x) {
  x$x <- 1.5
  x
}

tween_events(d, 'cubic-in-out', 50, start = time, end = time + duration,
             enter = from_left, exit = to_right, enter_length = 0.1,
             exit_length = 0.05)
#>                x          y        time     duration .id .phase .frame
#> 1   -0.500000000 0.56205709 0.003555029 8.878840e-02   1  enter      1
#> 2   -0.418269110 0.56205709 0.003555029 8.878840e-02   1  enter      2
#> 13  -0.500000000 0.93877804 0.032425741 4.396675e-02   2  enter      2
#> 3    0.153847120 0.56205709 0.003555029 8.878840e-02   1  enter      3
#> 14  -0.479673666 0.93877804 0.032425741 4.396675e-02   2  enter      3
#> 4    0.725963349 0.56205709 0.003555029 8.878840e-02   1  enter      4
#> 15  -0.337389328 0.93877804 0.032425741 4.396675e-02   2  enter      4
#> 5    0.807694239 0.56205709 0.003555029 8.878840e-02   1 static      5
#> 16  -0.027412734 0.93877804 0.032425741 4.396675e-02   2  enter      5
#> 74  -0.500000000 0.60186210 0.089546438 9.214369e-02   9  enter      5
#> 6    0.807694239 0.56205709 0.003555029 8.878840e-02   1 static      6
#> 17   0.114871604 0.93877804 0.032425741 4.396675e-02   2  enter      6
#> 75  -0.410600015 0.60186210 0.089546438 9.214369e-02   9  enter      6
#> 187 -0.500000000 0.85209805 0.124553506 6.898404e-02  20  enter      6
#> 7    0.807694239 0.56205709 0.003555029 8.878840e-02   1 static      7
#> 18   0.135197938 0.93877804 0.032425741 4.396675e-02   2 static      7
#> 76   0.215199882 0.60186210 0.089546438 9.214369e-02   9  enter      7
#> 188 -0.471375912 0.85209805 0.124553506 6.898404e-02  20  enter      7
#> 8    0.807694239 0.56205709 0.003555029 8.878840e-02   1 static      8
#> 19   0.135197938 0.93877804 0.032425741 4.396675e-02   2 static      8
#> 77   0.840999779 0.60186210 0.089546438 9.214369e-02   9  enter      8
#> 115 -0.500000000 0.72713374 0.171076011 9.709459e-02  13  enter      8
#> 189 -0.271007298 0.85209805 0.124553506 6.898404e-02  20  enter      8
#> 9    0.807694239 0.56205709 0.003555029 8.878840e-02   1    raw      9
#> 20   0.135197938 0.93877804 0.032425741 4.396675e-02   2    raw      9
#> 78   0.930399764 0.60186210 0.089546438 9.214369e-02   9 static      9
#> 116 -0.476612111 0.72713374 0.171076011 9.709459e-02  13  enter      9
#> 190  0.165510039 0.85209805 0.124553506 6.898404e-02  20  enter      9
#> 10   0.910258056 0.56205709 0.003555029 8.878840e-02   1   exit     10
#> 21   0.817598969 0.93877804 0.032425741 4.396675e-02   2   exit     10
#> 79   0.930399764 0.60186210 0.089546438 9.214369e-02   9 static     10
#> 117 -0.312896885 0.72713374 0.171076011 9.709459e-02  13  enter     10
#> 191  0.365878653 0.85209805 0.124553506 6.898404e-02  20  enter     10
#> 11   1.397436184 0.56205709 0.003555029 8.878840e-02   1   exit     11
#> 22   1.500000000 0.93877804 0.032425741 4.396675e-02   2   exit     11
#> 80   0.930399764 0.60186210 0.089546438 9.214369e-02   9 static     11
#> 118  0.043768428 0.72713374 0.171076011 9.709459e-02  13  enter     11
#> 177 -0.500000000 0.16386655 0.225910313 6.503250e-02  19  enter     11
#> 192  0.394502740 0.85209805 0.124553506 6.898404e-02  20 static     11
#> 12   1.500000000 0.56205709 0.003555029 8.878840e-02   1   exit     12
#> 81   0.930399764 0.60186210 0.089546438 9.214369e-02   9 static     12
#> 119  0.207483654 0.72713374 0.171076011 9.709459e-02  13  enter     12
#> 156 -0.500000000 0.12499100 0.242176785 5.388201e-02  17  enter     12
#> 178 -0.451728346 0.16386655 0.225910313 6.503250e-02  19  enter     12
#> 193  0.394502740 0.85209805 0.124553506 6.898404e-02  20 static     12
#> 82   0.930399764 0.60186210 0.089546438 9.214369e-02   9    raw     13
#> 120  0.230871543 0.72713374 0.171076011 9.709459e-02  13 static     13
#> 157 -0.452496577 0.12499100 0.242176785 5.388201e-02  17  enter     13
#> 179 -0.113826770 0.16386655 0.225910313 6.503250e-02  19  enter     13
#> 194  0.394502740 0.85209805 0.124553506 6.898404e-02  20 static     13
#> 83   1.215199882 0.60186210 0.089546438 9.214369e-02   9   exit     14
#> 121  0.230871543 0.72713374 0.171076011 9.709459e-02  13 static     14
#> 158 -0.119972620 0.12499100 0.242176785 5.388201e-02  17  enter     14
#> 180  0.224074806 0.16386655 0.225910313 6.503250e-02  19  enter     14
#> 195  0.394502740 0.85209805 0.124553506 6.898404e-02  20    raw     14
#> 84   1.500000000 0.60186210 0.089546438 9.214369e-02   9   exit     15
#> 122  0.230871543 0.72713374 0.171076011 9.709459e-02  13 static     15
#> 159  0.212551338 0.12499100 0.242176785 5.388201e-02  17  enter     15
#> 181  0.272346459 0.16386655 0.225910313 6.503250e-02  19 static     15
#> 196  0.947251370 0.85209805 0.124553506 6.898404e-02  20   exit     15
#> 123  0.230871543 0.72713374 0.171076011 9.709459e-02  13 static     16
#> 136 -0.500000000 0.52951134 0.346693913 1.087099e-02  15  enter     16
#> 160  0.260054760 0.12499100 0.242176785 5.388201e-02  17 static     16
#> 182  0.272346459 0.16386655 0.225910313 6.503250e-02  19 static     16
#> 197  1.500000000 0.85209805 0.124553506 6.898404e-02  20   exit     16
#> 124  0.230871543 0.72713374 0.171076011 9.709459e-02  13    raw     17
#> 137 -0.456430665 0.52951134 0.346693913 1.087099e-02  15  enter     17
#> 161  0.260054760 0.12499100 0.242176785 5.388201e-02  17 static     17
#> 183  0.272346459 0.16386655 0.225910313 6.503250e-02  19 static     17
#> 125  0.865435772 0.72713374 0.171076011 9.709459e-02  13   exit     18
#> 138 -0.151445322 0.52951134 0.346693913 1.087099e-02  15  enter     18
#> 162  0.260054760 0.12499100 0.242176785 5.388201e-02  17    raw     18
#> 184  0.272346459 0.16386655 0.225910313 6.503250e-02  19    raw     18
#> 126  1.500000000 0.72713374 0.171076011 9.709459e-02  13   exit     19
#> 139  0.512987034 0.52951134 0.346693913 1.087099e-02  15  enter     19
#> 163  0.880027380 0.12499100 0.242176785 5.388201e-02  17   exit     19
#> 185  0.886173230 0.16386655 0.225910313 6.503250e-02  19   exit     19
#> 23  -0.500000000 0.63277503 0.445004227 1.330576e-02   3  enter     20
#> 104 -0.500000000 0.85939037 0.432509275 8.655477e-02  12  enter     20
#> 140  0.817972377 0.52951134 0.346693913 1.087099e-02  15  enter     20
#> 144 -0.500000000 0.55042508 0.443084809 9.886215e-02  16  enter     20
#> 164  1.500000000 0.12499100 0.242176785 5.388201e-02  17   exit     20
#> 186  1.500000000 0.16386655 0.225910313 6.503250e-02  19   exit     20
#> 24  -0.455535993 0.63277503 0.445004227 1.330576e-02   3  enter     21
#> 105 -0.440884827 0.85939037 0.432509275 8.655477e-02  12  enter     21
#> 141  0.861541712 0.52951134 0.346693913 1.087099e-02  15    raw     21
#> 145 -0.466168335 0.55042508 0.443084809 9.886215e-02  16  enter     21
#> 25  -0.144287947 0.63277503 0.445004227 1.330576e-02   3  enter     22
#> 85  -0.500000000 0.10121720 0.490894560 4.512529e-02  10  enter     22
#> 106 -0.027078616 0.85939037 0.432509275 8.655477e-02  12  enter     22
#> 142  1.180770856 0.52951134 0.346693913 1.087099e-02  15   exit     22
#> 146 -0.229346677 0.55042508 0.443084809 9.886215e-02  16  enter     22
#> 26   0.533788155 0.63277503 0.445004227 1.330576e-02   3  enter     23
#> 41  -0.500000000 0.10776236 0.495491741 2.844252e-02   5  enter     23
#> 86  -0.482411301 0.10121720 0.490894560 4.512529e-02  10  enter     23
#> 107  0.386727596 0.85939037 0.432509275 8.655477e-02  12  enter     23
#> 143  1.500000000 0.52951134 0.346693913 1.087099e-02  15   exit     23
#> 147  0.286586221 0.55042508 0.443084809 9.886215e-02  16  enter     23
#> 27   0.845036202 0.63277503 0.445004227 1.330576e-02   3  enter     24
#> 42  -0.449938420 0.10776236 0.495491741 2.844252e-02   5  enter     24
#> 87  -0.359290407 0.10121720 0.490894560 4.512529e-02  10  enter     24
#> 108  0.445842769 0.85939037 0.432509275 8.655477e-02  12 static     24
#> 148  0.523407879 0.55042508 0.443084809 9.886215e-02  16  enter     24
#> 28   0.889500209 0.63277503 0.445004227 1.330576e-02   3    raw     25
#> 43  -0.099507363 0.10776236 0.495491741 2.844252e-02   5  enter     25
#> 88  -0.091062744 0.10121720 0.490894560 4.512529e-02  10  enter     25
#> 109  0.445842769 0.85939037 0.432509275 8.655477e-02  12 static     25
#> 149  0.557239544 0.55042508 0.443084809 9.886215e-02  16 static     25
#> 29   0.979944622 0.63277503 0.445004227 1.330576e-02   3   exit     26
#> 32  -0.500000000 0.95898469 0.576323347 2.632709e-02   4  enter     26
#> 44   0.250923695 0.10776236 0.495491741 2.844252e-02   5  enter     26
#> 89   0.032058150 0.10121720 0.490894560 4.512529e-02  10  enter     26
#> 110  0.445842769 0.85939037 0.432509275 8.655477e-02  12 static     26
#> 150  0.557239544 0.55042508 0.443084809 9.886215e-02  16 static     26
#> 165 -0.500000000 0.34042784 0.563982989 9.597635e-02  18  enter     26
#> 30   1.409555586 0.63277503 0.445004227 1.330576e-02   3   exit     27
#> 33  -0.478566777 0.95898469 0.576323347 2.632709e-02   4  enter     27
#> 45   0.300985275 0.10776236 0.495491741 2.844252e-02   5 static     27
#> 50  -0.500000000 0.05873208 0.589306734 1.960772e-05   6  enter     27
#> 90   0.049646849 0.10121720 0.490894560 4.512529e-02  10 static     27
#> 111  0.445842769 0.85939037 0.432509275 8.655477e-02  12 static     27
#> 151  0.557239544 0.55042508 0.443084809 9.886215e-02  16 static     27
#> 166 -0.429065451 0.34042784 0.563982989 9.597635e-02  18  enter     27
#> 31   1.500000000 0.63277503 0.445004227 1.330576e-02   3   exit     28
#> 34  -0.328534213 0.95898469 0.576323347 2.632709e-02   4  enter     28
#> 46   0.300985275 0.10776236 0.495491741 2.844252e-02   5    raw     28
#> 51  -0.434831919 0.05873208 0.589306734 1.960772e-05   6  enter     28
#> 91   0.049646849 0.10121720 0.490894560 4.512529e-02  10 static     28
#> 112  0.445842769 0.85939037 0.432509275 8.655477e-02  12    raw     28
#> 152  0.557239544 0.55042508 0.443084809 9.886215e-02  16 static     28
#> 167  0.067476392 0.34042784 0.563982989 9.597635e-02  18  enter     28
#> 35  -0.001677558 0.95898469 0.576323347 2.632709e-02   4  enter     29
#> 47   0.478617086 0.10776236 0.495491741 2.844252e-02   5   exit     29
#> 52   0.021344644 0.05873208 0.589306734 1.960772e-05   6  enter     29
#> 92   0.049646849 0.10121720 0.490894560 4.512529e-02  10    raw     29
#> 113  0.972921384 0.85939037 0.432509275 8.655477e-02  12   exit     29
#> 153  0.557239544 0.55042508 0.443084809 9.886215e-02  16    raw     29
#> 168  0.564018236 0.34042784 0.563982989 9.597635e-02  18  enter     29
#> 36   0.148355005 0.95898469 0.576323347 2.632709e-02   4  enter     30
#> 48   1.322368189 0.10776236 0.495491741 2.844252e-02   5   exit     30
#> 53   0.477521208 0.05873208 0.589306734 1.960772e-05   6  enter     30
#> 93   0.774823425 0.10121720 0.490894560 4.512529e-02  10   exit     30
#> 114  1.500000000 0.85939037 0.432509275 8.655477e-02  12   exit     30
#> 127 -0.500000000 0.81238950 0.668456473 2.148653e-02  14  enter     30
#> 154  1.028619772 0.55042508 0.443084809 9.886215e-02  16   exit     30
#> 169  0.634952785 0.34042784 0.563982989 9.597635e-02  18 static     30
#> 37   0.169788229 0.95898469 0.576323347 2.632709e-02   4 static     31
#> 49   1.500000000 0.10776236 0.495491741 2.844252e-02   5   exit     31
#> 54   0.542689289 0.05873208 0.589306734 1.960772e-05   6    raw     31
#> 94   1.500000000 0.10121720 0.490894560 4.512529e-02  10   exit     31
#> 128 -0.471158819 0.81238950 0.668456473 2.148653e-02  14  enter     31
#> 155  1.500000000 0.55042508 0.443084809 9.886215e-02  16   exit     31
#> 170  0.634952785 0.34042784 0.563982989 9.597635e-02  18 static     31
#> 38   0.169788229 0.95898469 0.576323347 2.632709e-02   4    raw     32
#> 55   1.021344644 0.05873208 0.589306734 1.960772e-05   6   exit     32
#> 129 -0.269270556 0.81238950 0.668456473 2.148653e-02  14  enter     32
#> 171  0.634952785 0.34042784 0.563982989 9.597635e-02  18 static     32
#> 39   0.834894114 0.95898469 0.576323347 2.632709e-02   4   exit     33
#> 56   1.500000000 0.05873208 0.589306734 1.960772e-05   6   exit     33
#> 130  0.170557447 0.81238950 0.668456473 2.148653e-02  14  enter     33
#> 172  0.634952785 0.34042784 0.563982989 9.597635e-02  18 static     33
#> 40   1.500000000 0.95898469 0.576323347 2.632709e-02   4   exit     34
#> 131  0.372445711 0.81238950 0.668456473 2.148653e-02  14  enter     34
#> 173  0.634952785 0.34042784 0.563982989 9.597635e-02  18    raw     34
#> 132  0.401286891 0.81238950 0.668456473 2.148653e-02  14 static     35
#> 174  0.763107928 0.34042784 0.563982989 9.597635e-02  18   exit     35
#> 133  0.401286891 0.81238950 0.668456473 2.148653e-02  14    raw     36
#> 175  1.371844857 0.34042784 0.563982989 9.597635e-02  18   exit     36
#> 134  0.950643446 0.81238950 0.668456473 2.148653e-02  14   exit     37
#> 176  1.500000000 0.34042784 0.563982989 9.597635e-02  18   exit     37
#> 57  -0.500000000 0.84551495 0.852085546 3.071225e-02   7  enter     38
#> 95  -0.500000000 0.63218233 0.834070492 4.728269e-02  11  enter     38
#> 135  1.500000000 0.81238950 0.668456473 2.148653e-02  14   exit     38
#> 58  -0.480011908 0.84551495 0.852085546 3.071225e-02   7  enter     39
#> 96  -0.455156496 0.63218233 0.834070492 4.728269e-02  11  enter     39
#> 59  -0.340095266 0.84551495 0.852085546 3.071225e-02   7  enter     40
#> 97  -0.141251969 0.63218233 0.834070492 4.728269e-02  11  enter     40
#> 60  -0.035276868 0.84551495 0.852085546 3.071225e-02   7  enter     41
#> 98   0.172652557 0.63218233 0.834070492 4.728269e-02  11  enter     41
#> 61   0.104639774 0.84551495 0.852085546 3.071225e-02   7  enter     42
#> 99   0.217496061 0.63218233 0.834070492 4.728269e-02  11 static     42
#> 62   0.124627866 0.84551495 0.852085546 3.071225e-02   7 static     43
#> 66  -0.500000000 0.59699891 0.965256968 1.109575e-03   8  enter     43
#> 100  0.217496061 0.63218233 0.834070492 4.728269e-02  11 static     43
#> 63   0.124627866 0.84551495 0.852085546 3.071225e-02   7    raw     44
#> 67  -0.478608100 0.59699891 0.965256968 1.109575e-03   8  enter     44
#> 101  0.217496061 0.63218233 0.834070492 4.728269e-02  11    raw     44
#> 64   0.812313933 0.84551495 0.852085546 3.071225e-02   7   exit     45
#> 68  -0.328864798 0.59699891 0.965256968 1.109575e-03   8  enter     45
#> 102  0.858748031 0.63218233 0.834070492 4.728269e-02  11   exit     45
#> 65   1.500000000 0.84551495 0.852085546 3.071225e-02   7   exit     46
#> 69  -0.002638320 0.59699891 0.965256968 1.109575e-03   8  enter     46
#> 103  1.500000000 0.63218233 0.834070492 4.728269e-02  11   exit     46
#> 70   0.147104981 0.59699891 0.965256968 1.109575e-03   8  enter     47
#> 71   0.168496881 0.59699891 0.965256968 1.109575e-03   8    raw     48
#> 72   0.834248441 0.59699891 0.965256968 1.109575e-03   8   exit     49
#> 73   1.500000000 0.59699891 0.965256968 1.109575e-03   8   exit     50