Transition in and out of events

This tweening function is a more powerful version of tween_appear(), with support for newer features such as enter/exits and tween phase identification. The tweener treats each row in the data as unique events in time, and creates frames with the correct events present at any given time.

tween_events(
  .data,
  ease,
  nframes,
  start,
  end = NULL,
  range = NULL,
  enter = NULL,
  exit = NULL,
  enter_length = 0,
  exit_length = 0
)

Arguments

.data

A data.frame with components at different stages

ease

The easing function to use. Either a single string or one for each column in the data set.

nframes

The number of frames to calculate for the tween

start, end

The start (and potential end) of the event encoded in the row, as unquoted expressions. Will be evaluated in the context of .data so can refer to columns in it. If end = NULL the event will be without extend and only visible in a single frame, unless enter and/or exit is given.

range

The range of time points to include in the tween. If NULL it will use the range of time

enter, exit

functions that calculate a start state for new observations that appear in to or an end state for observations that are not present in to. If NULL the new/old observations will not be part of the tween. The function gets a data.frame with either the start state of the exiting observations, or the end state of the entering observations and must return a modified version of that data.frame. See the Match, Enter, and Exit section for more information.

enter_length, exit_length

The lenght of the opening and closing transitions if enter and/or exit is given. Measured in the same units as time

Value

A data.frame with the same columns as .data along with .id giving the component id, .phase giving the state of each component in each frame, and .frame giving the frame membership of each row.

See also

Examples

d <- data.frame(
  x = runif(20),
  y = runif(20),
  time = runif(20),
  duration = runif(20, max = 0.1)
)
from_left <- function(x) {
  x$x <- -0.5
  x
}
to_right <- function(x) {
  x$x <- 1.5
  x
}

tween_events(d, 'cubic-in-out', 50, start = time, end = time + duration,
             enter = from_left, exit = to_right, enter_length = 0.1,
             exit_length = 0.05)
#>                x          y        time    duration .id .phase .frame
#> 126 -0.500000000 0.64392772 0.002352421 0.019040156  15  enter      1
#> 127 -0.433532807 0.64392772 0.002352421 0.019040156  15  enter      2
#> 15  -0.500000000 0.26310165 0.058185157 0.054115373   3  enter      3
#> 36  -0.500000000 0.20997695 0.061274857 0.001936218   5  enter      3
#> 128  0.031737546 0.64392772 0.002352421 0.019040156  15  enter      3
#> 16  -0.453155792 0.26310165 0.058185157 0.054115373   3  enter      4
#> 37  -0.479751275 0.20997695 0.061274857 0.001936218   5  enter      4
#> 62  -0.500000000 0.56271849 0.077582508 0.035029097   8  enter      4
#> 129  0.497007899 0.64392772 0.002352421 0.019040156  15  enter      4
#> 17  -0.125246333 0.26310165 0.058185157 0.054115373   3  enter      5
#> 38  -0.338010200 0.20997695 0.061274857 0.001936218   5  enter      5
#> 63  -0.460361779 0.56271849 0.077582508 0.035029097   8  enter      5
#> 130  0.563475092 0.64392772 0.002352421 0.019040156  15 static      5
#> 18   0.202663125 0.26310165 0.058185157 0.054115373   3  enter      6
#> 39  -0.029217143 0.20997695 0.061274857 0.001936218   5  enter      6
#> 64  -0.182894232 0.56271849 0.077582508 0.035029097   8  enter      6
#> 131  0.563475092 0.64392772 0.002352421 0.019040156  15    raw      6
#> 153 -0.500000000 0.14763045 0.133805839 0.036657255  18  enter      6
#> 19   0.249507334 0.26310165 0.058185157 0.054115373   3 static      7
#> 40   0.112523932 0.20997695 0.061274857 0.001936218   5  enter      7
#> 65   0.094573316 0.56271849 0.077582508 0.035029097   8  enter      7
#> 132  1.031737546 0.64392772 0.002352421 0.019040156  15   exit      7
#> 154 -0.476747681 0.14763045 0.133805839 0.036657255  18  enter      7
#> 20   0.249507334 0.26310165 0.058185157 0.054115373   3 static      8
#> 41   0.132772657 0.20997695 0.061274857 0.001936218   5    raw      8
#> 66   0.134211537 0.56271849 0.077582508 0.035029097   8 static      8
#> 133  1.500000000 0.64392772 0.002352421 0.019040156  15   exit      8
#> 155 -0.313981447 0.14763045 0.133805839 0.036657255  18  enter      8
#> 21   0.249507334 0.26310165 0.058185157 0.054115373   3 static      9
#> 42   0.816386328 0.20997695 0.061274857 0.001936218   5   exit      9
#> 67   0.134211537 0.56271849 0.077582508 0.035029097   8 static      9
#> 156  0.040616420 0.14763045 0.133805839 0.036657255  18  enter      9
#> 1   -0.500000000 0.51568876 0.206431749 0.008355361   1  enter     10
#> 22   0.249507334 0.26310165 0.058185157 0.054115373   3    raw     10
#> 43   1.500000000 0.20997695 0.061274857 0.001936218   5   exit     10
#> 68   0.134211537 0.56271849 0.077582508 0.035029097   8    raw     10
#> 79  -0.500000000 0.74244253 0.229574004 0.050344603  10  enter     10
#> 157  0.203382654 0.14763045 0.133805839 0.036657255  18  enter     10
#> 2   -0.465354832 0.51568876 0.206431749 0.008355361   1  enter     11
#> 23   0.874753667 0.26310165 0.058185157 0.054115373   3   exit     11
#> 69   0.817105768 0.56271849 0.077582508 0.035029097   8   exit     11
#> 80  -0.477862530 0.74244253 0.229574004 0.050344603  10  enter     11
#> 158  0.226634973 0.14763045 0.133805839 0.036657255  18 static     11
#> 3   -0.222838652 0.51568876 0.206431749 0.008355361   1  enter     12
#> 24   1.500000000 0.26310165 0.058185157 0.054115373   3   exit     12
#> 70   1.500000000 0.56271849 0.077582508 0.035029097   8   exit     12
#> 81  -0.322900238 0.74244253 0.229574004 0.050344603  10  enter     12
#> 159  0.226634973 0.14763045 0.133805839 0.036657255  18    raw     12
#> 4    0.019677527 0.51568876 0.206431749 0.008355361   1  enter     13
#> 82   0.014696182 0.74244253 0.229574004 0.050344603  10  enter     13
#> 98  -0.500000000 0.73784142 0.288189583 0.080478723  12  enter     13
#> 160  0.863317487 0.14763045 0.133805839 0.036657255  18   exit     13
#> 5    0.054322696 0.51568876 0.206431749 0.008355361   1    raw     14
#> 83   0.169658473 0.74244253 0.229574004 0.050344603  10  enter     14
#> 99  -0.439456148 0.73784142 0.288189583 0.080478723  12  enter     14
#> 161  1.500000000 0.14763045 0.133805839 0.036657255  18   exit     14
#> 6    0.777161348 0.51568876 0.206431749 0.008355361   1   exit     15
#> 8   -0.500000000 0.33729809 0.332769852 0.008486850   2  enter     15
#> 84   0.191795944 0.74244253 0.229574004 0.050344603  10 static     15
#> 100 -0.015649183 0.73784142 0.288189583 0.080478723  12  enter     15
#> 109 -0.500000000 0.26828783 0.328708843 0.001882263  13  enter     15
#> 7    1.500000000 0.51568876 0.206431749 0.008355361   1   exit     16
#> 9   -0.433688851 0.33729809 0.332769852 0.008486850   2  enter     16
#> 85   0.191795944 0.74244253 0.229574004 0.050344603  10 static     16
#> 101  0.408157781 0.73784142 0.288189583 0.080478723  12  enter     16
#> 110 -0.440798094 0.26828783 0.328708843 0.001882263  13  enter     16
#> 116 -0.500000000 0.65191481 0.359571712 0.059291766  14  enter     16
#> 10   0.030489192 0.33729809 0.332769852 0.008486850   2  enter     17
#> 86   0.191795944 0.74244253 0.229574004 0.050344603  10    raw     17
#> 102  0.468701633 0.73784142 0.288189583 0.080478723  12 static     17
#> 111 -0.026384752 0.26828783 0.328708843 0.001882263  13  enter     17
#> 117 -0.454651249 0.65191481 0.359571712 0.059291766  14  enter     17
#> 11   0.494667235 0.33729809 0.332769852 0.008486850   2  enter     18
#> 87   0.845897972 0.74244253 0.229574004 0.050344603  10   exit     18
#> 103  0.468701633 0.73784142 0.288189583 0.080478723  12 static     18
#> 112  0.388028589 0.26828783 0.328708843 0.001882263  13  enter     18
#> 118 -0.137209989 0.65191481 0.359571712 0.059291766  14  enter     18
#> 12   0.560978384 0.33729809 0.332769852 0.008486850   2    raw     19
#> 88   1.500000000 0.74244253 0.229574004 0.050344603  10   exit     19
#> 104  0.468701633 0.73784142 0.288189583 0.080478723  12 static     19
#> 113  0.447230495 0.26828783 0.328708843 0.001882263  13    raw     19
#> 119  0.180231271 0.65191481 0.359571712 0.059291766  14  enter     19
#> 13   1.030489192 0.33729809 0.332769852 0.008486850   2   exit     20
#> 105  0.468701633 0.73784142 0.288189583 0.080478723  12    raw     20
#> 114  0.973615248 0.26828783 0.328708843 0.001882263  13   exit     20
#> 120  0.225580022 0.65191481 0.359571712 0.059291766  14 static     20
#> 14   1.500000000 0.33729809 0.332769852 0.008486850   2   exit     21
#> 106  0.621486576 0.73784142 0.288189583 0.080478723  12   exit     21
#> 115  1.500000000 0.26828783 0.328708843 0.001882263  13   exit     21
#> 121  0.225580022 0.65191481 0.359571712 0.059291766  14 static     21
#> 107  1.347215057 0.73784142 0.288189583 0.080478723  12   exit     22
#> 122  0.225580022 0.65191481 0.359571712 0.059291766  14 static     22
#> 108  1.500000000 0.73784142 0.288189583 0.080478723  12   exit     23
#> 123  0.225580022 0.65191481 0.359571712 0.059291766  14    raw     23
#> 142 -0.500000000 0.18088269 0.531812161 0.076185702  17  enter     23
#> 25  -0.500000000 0.47262700 0.559185639 0.072822845   4  enter     24
#> 124  0.862790011 0.65191481 0.359571712 0.059291766  14   exit     24
#> 143 -0.412514564 0.18088269 0.531812161 0.076185702  17  enter     24
#> 26  -0.437963721 0.47262700 0.559185639 0.072822845   4  enter     25
#> 125  1.500000000 0.65191481 0.359571712 0.059291766  14   exit     25
#> 144  0.199883485 0.18088269 0.531812161 0.076185702  17  enter     25
#> 27  -0.003709766 0.47262700 0.559185639 0.072822845   4  enter     26
#> 145  0.812281535 0.18088269 0.531812161 0.076185702  17  enter     26
#> 28   0.430544189 0.47262700 0.559185639 0.072822845   4  enter     27
#> 146  0.899766971 0.18088269 0.531812161 0.076185702  17 static     27
#> 29   0.492580469 0.47262700 0.559185639 0.072822845   4 static     28
#> 147  0.899766971 0.18088269 0.531812161 0.076185702  17 static     28
#> 30   0.492580469 0.47262700 0.559185639 0.072822845   4 static     29
#> 148  0.899766971 0.18088269 0.531812161 0.076185702  17 static     29
#> 31   0.492580469 0.47262700 0.559185639 0.072822845   4 static     30
#> 149  0.899766971 0.18088269 0.531812161 0.076185702  17    raw     30
#> 171 -0.500000000 0.16123886 0.705251220 0.044602923  20  enter     30
#> 32   0.492580469 0.47262700 0.559185639 0.072822845   4    raw     31
#> 150  0.988690383 0.18088269 0.531812161 0.076185702  17   exit     31
#> 172 -0.438077093 0.16123886 0.705251220 0.044602923  20  enter     31
#> 33   0.641827807 0.47262700 0.559185639 0.072822845   4   exit     32
#> 151  1.411076588 0.18088269 0.531812161 0.076185702  17   exit     32
#> 173 -0.004616740 0.16123886 0.705251220 0.044602923  20  enter     32
#> 34   1.350752662 0.47262700 0.559185639 0.072822845   4   exit     33
#> 134 -0.500000000 0.46304983 0.763503714 0.026032240  16  enter     33
#> 152  1.500000000 0.18088269 0.531812161 0.076185702  17   exit     33
#> 162 -0.500000000 0.60738385 0.778759911 0.040924736  19  enter     33
#> 174  0.428843612 0.16123886 0.705251220 0.044602923  20  enter     33
#> 35   1.500000000 0.47262700 0.559185639 0.072822845   4   exit     34
#> 54  -0.500000000 0.05740801 0.787206781 0.030659756   7  enter     34
#> 135 -0.413317752 0.46304983 0.763503714 0.026032240  16  enter     34
#> 163 -0.457035060 0.60738385 0.778759911 0.040924736  19  enter     34
#> 175  0.490766519 0.16123886 0.705251220 0.044602923  20 static     34
#> 55  -0.428392665 0.05740801 0.787206781 0.030659756   7  enter     35
#> 136  0.193457988 0.46304983 0.763503714 0.026032240  16  enter     35
#> 164 -0.156280482 0.60738385 0.778759911 0.040924736  19  enter     35
#> 176  0.490766519 0.16123886 0.705251220 0.044602923  20 static     35
#> 56   0.072858683 0.05740801 0.787206781 0.030659756   7  enter     36
#> 137  0.800233727 0.46304983 0.763503714 0.026032240  16  enter     36
#> 165  0.498934850 0.60738385 0.778759911 0.040924736  19  enter     36
#> 177  0.490766519 0.16123886 0.705251220 0.044602923  20    raw     36
#> 57   0.574110031 0.05740801 0.787206781 0.030659756   7  enter     37
#> 89  -0.500000000 0.88585018 0.867465737 0.033842302  11  enter     37
#> 138  0.886915976 0.46304983 0.763503714 0.026032240  16 static     37
#> 166  0.799689429 0.60738385 0.778759911 0.040924736  19  enter     37
#> 178  0.995383260 0.16123886 0.705251220 0.044602923  20   exit     37
#> 58   0.645717366 0.05740801 0.787206781 0.030659756   7 static     38
#> 90  -0.409900620 0.88585018 0.867465737 0.033842302  11  enter     38
#> 139  0.886915976 0.46304983 0.763503714 0.026032240  16    raw     38
#> 167  0.842654368 0.60738385 0.778759911 0.040924736  19 static     38
#> 179  1.500000000 0.16123886 0.705251220 0.044602923  20   exit     38
#> 59   0.645717366 0.05740801 0.787206781 0.030659756   7    raw     39
#> 71  -0.500000000 0.29681741 0.914333480 0.027583672   9  enter     39
#> 91   0.220795043 0.88585018 0.867465737 0.033842302  11  enter     39
#> 140  1.193457988 0.46304983 0.763503714 0.026032240  16   exit     39
#> 168  0.842654368 0.60738385 0.778759911 0.040924736  19    raw     39
#> 60   1.072858683 0.05740801 0.787206781 0.030659756   7   exit     40
#> 72  -0.409395546 0.29681741 0.914333480 0.027583672   9  enter     40
#> 92   0.851490705 0.88585018 0.867465737 0.033842302  11  enter     40
#> 141  1.500000000 0.46304983 0.763503714 0.026032240  16   exit     40
#> 169  1.171327184 0.60738385 0.778759911 0.040924736  19   exit     40
#> 44  -0.500000000 0.85216158 0.956857463 0.070459388   6  enter     41
#> 61   1.500000000 0.05740801 0.787206781 0.030659756   7   exit     41
#> 73   0.224835631 0.29681741 0.914333480 0.027583672   9  enter     41
#> 93   0.941590085 0.88585018 0.867465737 0.033842302  11 static     41
#> 170  1.500000000 0.60738385 0.778759911 0.040924736  19   exit     41
#> 45  -0.443192469 0.85216158 0.956857463 0.070459388   6  enter     42
#> 74   0.859066808 0.29681741 0.914333480 0.027583672   9  enter     42
#> 94   0.941590085 0.88585018 0.867465737 0.033842302  11 static     42
#> 46  -0.045539750 0.85216158 0.956857463 0.070459388   6  enter     43
#> 75   0.949671262 0.29681741 0.914333480 0.027583672   9 static     43
#> 95   0.941590085 0.88585018 0.867465737 0.033842302  11    raw     43
#> 47   0.352112969 0.85216158 0.956857463 0.070459388   6  enter     44
#> 76   0.949671262 0.29681741 0.914333480 0.027583672   9    raw     44
#> 96   1.220795043 0.88585018 0.867465737 0.033842302  11   exit     44
#> 48   0.408920500 0.85216158 0.956857463 0.070459388   6 static     45
#> 77   1.224835631 0.29681741 0.914333480 0.027583672   9   exit     45
#> 97   1.500000000 0.88585018 0.867465737 0.033842302  11   exit     45
#> 49   0.408920500 0.85216158 0.956857463 0.070459388   6 static     46
#> 78   1.500000000 0.29681741 0.914333480 0.027583672   9   exit     46
#> 50   0.408920500 0.85216158 0.956857463 0.070459388   6 static     47
#> 51   0.408920500 0.85216158 0.956857463 0.070459388   6    raw     48
#> 52   0.954460250 0.85216158 0.956857463 0.070459388   6   exit     49
#> 53   1.500000000 0.85216158 0.956857463 0.070459388   6   exit     50