Transition in and out of events

This tweening function is a more powerful version of tween_appear(), with support for newer features such as enter/exits and tween phase identification. The tweener treats each row in the data as unique events in time, and creates frames with the correct events present at any given time.

tween_events(
  .data,
  ease,
  nframes,
  start,
  end = NULL,
  range = NULL,
  enter = NULL,
  exit = NULL,
  enter_length = 0,
  exit_length = 0
)

Arguments

.data

A data.frame with components at different stages

ease

The easing function to use. Either a single string or one for each column in the data set.

nframes

The number of frames to calculate for the tween

start, end

The start (and potential end) of the event encoded in the row, as unquoted expressions. Will be evaluated in the context of .data so can refer to columns in it. If end = NULL the event will be without extend and only visible in a single frame, unless enter and/or exit is given.

range

The range of time points to include in the tween. If NULL it will use the range of time

enter

functions that calculate a start state for new observations that appear in to or an end state for observations that are not present in to. If NULL the new/old observations will not be part of the tween. The function gets a data.frame with either the start state of the exiting observations, or the end state of the entering observations and must return a modified version of that data.frame. See the Match, Enter, and Exit section for more information.

exit

functions that calculate a start state for new observations that appear in to or an end state for observations that are not present in to. If NULL the new/old observations will not be part of the tween. The function gets a data.frame with either the start state of the exiting observations, or the end state of the entering observations and must return a modified version of that data.frame. See the Match, Enter, and Exit section for more information.

enter_length

The lenght of the opening and closing transitions if enter and/or exit is given. Measured in the same units as time

exit_length

The lenght of the opening and closing transitions if enter and/or exit is given. Measured in the same units as time

Value

A data.frame with the same columns as .data along with .id giving the component id, .phase giving the state of each component in each frame, and .frame giving the frame membership of each row.

See also

Examples

d <- data.frame(
  x = runif(20),
  y = runif(20),
  time = runif(20),
  duration = runif(20, max = 0.1)
)
from_left <- function(x) {
  x$x <- -0.5
  x
}
to_right <- function(x) {
  x$x <- 1.5
  x
}

tween_events(d, 'cubic-in-out', 50, start = time, end = time + duration,
             enter = from_left, exit = to_right, enter_length = 0.1,
             exit_length = 0.05)
#>               x           y       time    duration .id .phase .frame
#> 69  -0.50000000 0.252730283 0.08758807 0.021286654   8  enter      1
#> 70  -0.45274200 0.252730283 0.08758807 0.021286654   8  enter      2
#> 71  -0.12193598 0.252730283 0.08758807 0.021286654   8  enter      3
#> 78  -0.50000000 0.341305613 0.12683495 0.011096029   9  enter      3
#> 187 -0.50000000 0.407255717 0.12345674 0.085029565  20  enter      3
#> 72   0.59874856 0.252730283 0.08758807 0.021286654   8  enter      4
#> 79  -0.45696584 0.341305613 0.12683495 0.011096029   9  enter      4
#> 128 -0.50000000 0.255323277 0.14099474 0.042323682  14  enter      4
#> 188 -0.46749787 0.407255717 0.12345674 0.085029565  20  enter      4
#> 73   0.92955458 0.252730283 0.08758807 0.021286654   8  enter      5
#> 80  -0.15572673 0.341305613 0.12683495 0.011096029   9  enter      5
#> 129 -0.44924835 0.255323277 0.14099474 0.042323682  14  enter      5
#> 189 -0.23998300 0.407255717 0.12345674 0.085029565  20  enter      5
#> 74   0.97681259 0.252730283 0.08758807 0.021286654   8 static      6
#> 81   0.50054418 0.341305613 0.12683495 0.011096029   9  enter      6
#> 130 -0.09398683 0.255323277 0.14099474 0.042323682  14  enter      6
#> 190  0.25567441 0.407255717 0.12345674 0.085029565  20  enter      6
#> 75   0.97681259 0.252730283 0.08758807 0.021286654   8    raw      7
#> 82   0.80178329 0.341305613 0.12683495 0.011096029   9  enter      7
#> 131  0.26127469 0.255323277 0.14099474 0.042323682  14  enter      7
#> 191  0.48318928 0.407255717 0.12345674 0.085029565  20  enter      7
#> 76   1.23840629 0.252730283 0.08758807 0.021286654   8   exit      8
#> 83   0.84481745 0.341305613 0.12683495 0.011096029   9    raw      8
#> 132  0.31202634 0.255323277 0.14099474 0.042323682  14 static      8
#> 192  0.51569141 0.407255717 0.12345674 0.085029565  20 static      8
#> 77   1.50000000 0.252730283 0.08758807 0.021286654   8   exit      9
#> 84   0.94188153 0.341305613 0.12683495 0.011096029   9   exit      9
#> 133  0.31202634 0.255323277 0.14099474 0.042323682  14 static      9
#> 193  0.51569141 0.407255717 0.12345674 0.085029565  20 static      9
#> 59  -0.50000000 0.637680070 0.28508297 0.030654481   7  enter     10
#> 85   1.40293592 0.341305613 0.12683495 0.011096029   9   exit     10
#> 134  0.31202634 0.255323277 0.14099474 0.042323682  14    raw     10
#> 194  0.51569141 0.407255717 0.12345674 0.085029565  20 static     10
#> 60  -0.46433329 0.637680070 0.28508297 0.030654481   7  enter     11
#> 86   1.50000000 0.341305613 0.12683495 0.011096029   9   exit     11
#> 98  -0.50000000 0.045832326 0.29272116 0.050245588  11  enter     11
#> 135  0.48802244 0.255323277 0.14099474 0.042323682  14   exit     11
#> 146 -0.50000000 0.724826409 0.29011459 0.013777174  16  enter     11
#> 195  0.51569141 0.407255717 0.12345674 0.085029565  20 static     11
#> 61  -0.21466635 0.637680070 0.28508297 0.030654481   7  enter     12
#> 99  -0.47359579 0.045832326 0.29272116 0.050245588  11  enter     12
#> 136  1.32400390 0.255323277 0.14099474 0.042323682  14   exit     12
#> 147 -0.46619076 0.724826409 0.29011459 0.013777174  16  enter     12
#> 196  0.51569141 0.407255717 0.12345674 0.085029565  20    raw     12
#> 62   0.32925093 0.637680070 0.28508297 0.030654481   7  enter     13
#> 100 -0.28876633 0.045832326 0.29272116 0.050245588  11  enter     13
#> 108 -0.50000000 0.156039551 0.33550895 0.087817678  12  enter     13
#> 137  1.50000000 0.255323277 0.14099474 0.042323682  14   exit     13
#> 148 -0.22952605 0.724826409 0.29011459 0.013777174  16  enter     13
#> 197  1.00784571 0.407255717 0.12345674 0.085029565  20   exit     13
#> 38  -0.50000000 0.308170997 0.36090770 0.063045681   5  enter     14
#> 63   0.57891787 0.637680070 0.28508297 0.030654481   7  enter     14
#> 101  0.11389786 0.045832326 0.29272116 0.050245588  11  enter     14
#> 109 -0.46853011 0.156039551 0.33550895 0.087817678  12  enter     14
#> 149  0.28606491 0.724826409 0.29011459 0.013777174  16  enter     14
#> 198  1.50000000 0.407255717 0.12345674 0.085029565  20   exit     14
#> 39  -0.47196127 0.308170997 0.36090770 0.063045681   5  enter     15
#> 64   0.61458458 0.637680070 0.28508297 0.030654481   7 static     15
#> 102  0.29872732 0.045832326 0.29272116 0.050245588  11  enter     15
#> 110 -0.24824085 0.156039551 0.33550895 0.087817678  12  enter     15
#> 150  0.52272961 0.724826409 0.29011459 0.013777174  16  enter     15
#> 40  -0.27569018 0.308170997 0.36090770 0.063045681   5  enter     16
#> 65   0.61458458 0.637680070 0.28508297 0.030654481   7 static     16
#> 103  0.32513153 0.045832326 0.29272116 0.050245588  11 static     16
#> 111  0.23167503 0.156039551 0.33550895 0.087817678  12  enter     16
#> 151  0.55653885 0.724826409 0.29011459 0.013777174  16    raw     16
#> 28  -0.50000000 0.009342814 0.41945364 0.044849282   4  enter     17
#> 41   0.15190041 0.308170997 0.36090770 0.063045681   5  enter     17
#> 66   0.61458458 0.637680070 0.28508297 0.030654481   7    raw     17
#> 87  -0.50000000 0.675146163 0.42001838 0.065088298  10  enter     17
#> 104  0.32513153 0.045832326 0.29272116 0.050245588  11 static     17
#> 112  0.45196429 0.156039551 0.33550895 0.087817678  12  enter     17
#> 152  0.69631088 0.724826409 0.29011459 0.013777174  16   exit     17
#> 29  -0.45724851 0.009342814 0.41945364 0.044849282   4  enter     18
#> 42   0.34817150 0.308170997 0.36090770 0.063045681   5  enter     18
#> 67   1.05729229 0.637680070 0.28508297 0.030654481   7   exit     18
#> 88  -0.46749487 0.675146163 0.42001838 0.065088298  10  enter     18
#> 105  0.32513153 0.045832326 0.29272116 0.050245588  11    raw     18
#> 113  0.48343418 0.156039551 0.33550895 0.087817678  12 static     18
#> 153  1.36022798 0.724826409 0.29011459 0.013777174  16   exit     18
#> 175 -0.50000000 0.294995994 0.44105865 0.070204246  19  enter     18
#> 30  -0.15798811 0.009342814 0.41945364 0.044849282   4  enter     19
#> 43   0.37621022 0.308170997 0.36090770 0.063045681   5 static     19
#> 68   1.50000000 0.637680070 0.28508297 0.030654481   7   exit     19
#> 89  -0.23995893 0.675146163 0.42001838 0.065088298  10  enter     19
#> 106  0.91256576 0.045832326 0.29272116 0.050245588  11   exit     19
#> 114  0.48343418 0.156039551 0.33550895 0.087817678  12 static     19
#> 154  1.50000000 0.724826409 0.29011459 0.013777174  16   exit     19
#> 176 -0.46521759 0.294995994 0.44105865 0.070204246  19  enter     19
#> 9   -0.50000000 0.721926169 0.47620709 0.013991969   2  enter     20
#> 18  -0.50000000 0.702665466 0.48619613 0.025429391   3  enter     20
#> 31   0.49397204 0.009342814 0.41945364 0.044849282   4  enter     20
#> 44   0.37621022 0.308170997 0.36090770 0.063045681   5 static     20
#> 90   0.25574435 0.675146163 0.42001838 0.065088298  10  enter     20
#> 107  1.50000000 0.045832326 0.29272116 0.050245588  11   exit     20
#> 115  0.48343418 0.156039551 0.33550895 0.087817678  12 static     20
#> 177 -0.22174071 0.294995994 0.44105865 0.070204246  19  enter     20
#> 10  -0.44918290 0.721926169 0.47620709 0.013991969   2  enter     21
#> 19  -0.47563259 0.702665466 0.48619613 0.025429391   3  enter     21
#> 32   0.79323244 0.009342814 0.41945364 0.044849282   4  enter     21
#> 45   0.37621022 0.308170997 0.36090770 0.063045681   5 static     21
#> 91   0.48328028 0.675146163 0.42001838 0.065088298  10  enter     21
#> 116  0.48343418 0.156039551 0.33550895 0.087817678  12 static     21
#> 178  0.30869106 0.294995994 0.44105865 0.070204246  19  enter     21
#> 11  -0.09346318 0.721926169 0.47620709 0.013991969   2  enter     22
#> 20  -0.30506070 0.702665466 0.48619613 0.025429391   3  enter     22
#> 33   0.83598393 0.009342814 0.41945364 0.044849282   4 static     22
#> 46   0.37621022 0.308170997 0.36090770 0.063045681   5    raw     22
#> 92   0.51578541 0.675146163 0.42001838 0.065088298  10 static     22
#> 117  0.48343418 0.156039551 0.33550895 0.087817678  12    raw     22
#> 179  0.55216794 0.294995994 0.44105865 0.070204246  19  enter     22
#> 12   0.26225653 0.721926169 0.47620709 0.013991969   2  enter     23
#> 21   0.06654235 0.702665466 0.48619613 0.025429391   3  enter     23
#> 34   0.83598393 0.009342814 0.41945364 0.044849282   4 static     23
#> 47   0.93810511 0.308170997 0.36090770 0.063045681   5   exit     23
#> 93   0.51578541 0.675146163 0.42001838 0.065088298  10 static     23
#> 118  0.99171709 0.156039551 0.33550895 0.087817678  12   exit     23
#> 180  0.58695035 0.294995994 0.44105865 0.070204246  19 static     23
#> 13   0.31307363 0.721926169 0.47620709 0.013991969   2 static     24
#> 22   0.23711424 0.702665466 0.48619613 0.025429391   3  enter     24
#> 35   0.83598393 0.009342814 0.41945364 0.044849282   4    raw     24
#> 48   1.50000000 0.308170997 0.36090770 0.063045681   5   exit     24
#> 49  -0.50000000 0.190109426 0.55691015 0.046012455   6  enter     24
#> 94   0.51578541 0.675146163 0.42001838 0.065088298  10 static     24
#> 119  1.50000000 0.156039551 0.33550895 0.087817678  12   exit     24
#> 120 -0.50000000 0.006968677 0.56462305 0.006780084  13  enter     24
#> 181  0.58695035 0.294995994 0.44105865 0.070204246  19 static     24
#> 14   0.31307363 0.721926169 0.47620709 0.013991969   2    raw     25
#> 23   0.26148166 0.702665466 0.48619613 0.025429391   3 static     25
#> 36   1.16799196 0.009342814 0.41945364 0.044849282   4   exit     25
#> 50  -0.43021334 0.190109426 0.55691015 0.046012455   6  enter     25
#> 95   0.51578541 0.675146163 0.42001838 0.065088298  10    raw     25
#> 121 -0.45341017 0.006968677 0.56462305 0.006780084  13  enter     25
#> 182  0.58695035 0.294995994 0.44105865 0.070204246  19 static     25
#> 15   0.48891458 0.721926169 0.47620709 0.013991969   2   exit     26
#> 24   0.26148166 0.702665466 0.48619613 0.025429391   3    raw     26
#> 37   1.50000000 0.009342814 0.41945364 0.044849282   4   exit     26
#> 51   0.05829328 0.190109426 0.55691015 0.046012455   6  enter     26
#> 96   1.00789271 0.675146163 0.42001838 0.065088298  10   exit     26
#> 122 -0.12728135 0.006968677 0.56462305 0.006780084  13  enter     26
#> 164 -0.50000000 0.207324513 0.61024175 0.062231453  18  enter     26
#> 183  0.58695035 0.294995994 0.44105865 0.070204246  19    raw     26
#> 16   1.32415906 0.721926169 0.47620709 0.013991969   2   exit     27
#> 25   0.44496586 0.702665466 0.48619613 0.025429391   3   exit     27
#> 52   0.54679989 0.190109426 0.55691015 0.046012455   6  enter     27
#> 97   1.50000000 0.675146163 0.42001838 0.065088298  10   exit     27
#> 123  0.58321358 0.006968677 0.56462305 0.006780084  13  enter     27
#> 165 -0.45855713 0.207324513 0.61024175 0.062231453  18  enter     27
#> 184  0.72221696 0.294995994 0.44105865 0.070204246  19   exit     27
#> 17   1.50000000 0.721926169 0.47620709 0.013991969   2   exit     28
#> 26   1.31651580 0.702665466 0.48619613 0.025429391   3   exit     28
#> 53   0.61658655 0.190109426 0.55691015 0.046012455   6 static     28
#> 124  0.90934241 0.006968677 0.56462305 0.006780084  13  enter     28
#> 166 -0.16845703 0.207324513 0.61024175 0.062231453  18  enter     28
#> 185  1.36473339 0.294995994 0.44105865 0.070204246  19   exit     28
#> 27   1.50000000 0.702665466 0.48619613 0.025429391   3   exit     29
#> 54   0.61658655 0.190109426 0.55691015 0.046012455   6 static     29
#> 125  0.95593224 0.006968677 0.56462305 0.006780084  13    raw     29
#> 167  0.46354675 0.207324513 0.61024175 0.062231453  18  enter     29
#> 186  1.50000000 0.294995994 0.44105865 0.070204246  19   exit     29
#> 55   0.61658655 0.190109426 0.55691015 0.046012455   6 static     30
#> 126  1.22796612 0.006968677 0.56462305 0.006780084  13   exit     30
#> 168  0.75364684 0.207324513 0.61024175 0.062231453  18  enter     30
#> 56   0.61658655 0.190109426 0.55691015 0.046012455   6    raw     31
#> 127  1.50000000 0.006968677 0.56462305 0.006780084  13   exit     31
#> 169  0.79508971 0.207324513 0.61024175 0.062231453  18 static     31
#> 1   -0.50000000 0.869365572 0.73098980 0.005864858   1  enter     32
#> 57   1.05829328 0.190109426 0.55691015 0.046012455   6   exit     32
#> 170  0.79508971 0.207324513 0.61024175 0.062231453  18 static     32
#> 2   -0.47991996 0.869365572 0.73098980 0.005864858   1  enter     33
#> 58   1.50000000 0.190109426 0.55691015 0.046012455   6   exit     33
#> 171  0.79508971 0.207324513 0.61024175 0.062231453  18 static     33
#> 3   -0.33935971 0.869365572 0.73098980 0.005864858   1  enter     34
#> 172  0.79508971 0.207324513 0.61024175 0.062231453  18    raw     34
#> 4   -0.03313915 0.869365572 0.73098980 0.005864858   1  enter     35
#> 155 -0.50000000 0.041300160 0.80053347 0.012401836  17  enter     35
#> 173  1.14754486 0.207324513 0.61024175 0.062231453  18   exit     35
#> 5    0.10742111 0.869365572 0.73098980 0.005864858   1  enter     36
#> 156 -0.47613812 0.041300160 0.80053347 0.012401836  17  enter     36
#> 174  1.50000000 0.207324513 0.61024175 0.062231453  18   exit     36
#> 6    0.12750115 0.869365572 0.73098980 0.005864858   1    raw     37
#> 157 -0.30910498 0.041300160 0.80053347 0.012401836  17  enter     37
#> 7    0.81375057 0.869365572 0.73098980 0.005864858   1   exit     38
#> 158  0.05478865 0.041300160 0.80053347 0.012401836  17  enter     38
#> 8    1.50000000 0.869365572 0.73098980 0.005864858   1   exit     39
#> 159  0.22182179 0.041300160 0.80053347 0.012401836  17  enter     39
#> 160  0.24568367 0.041300160 0.80053347 0.012401836  17 static     40
#> 161  0.24568367 0.041300160 0.80053347 0.012401836  17    raw     41
#> 162  0.87284183 0.041300160 0.80053347 0.012401836  17   exit     42
#> 138 -0.50000000 0.171277873 0.95516476 0.003808899  15  enter     43
#> 163  1.50000000 0.041300160 0.80053347 0.012401836  17   exit     43
#> 139 -0.43027187 0.171277873 0.95516476 0.003808899  15  enter     44
#> 140  0.05782505 0.171277873 0.95516476 0.003808899  15  enter     45
#> 141  0.54592196 0.171277873 0.95516476 0.003808899  15  enter     46
#> 142  0.61565010 0.171277873 0.95516476 0.003808899  15 static     47
#> 143  0.61565010 0.171277873 0.95516476 0.003808899  15    raw     48
#> 144  1.05782505 0.171277873 0.95516476 0.003808899  15   exit     49
#> 145  1.50000000 0.171277873 0.95516476 0.003808899  15   exit     50