Transition in and out of events

This tweening function is a more powerful version of tween_appear(), with support for newer features such as enter/exits and tween phase identification. The tweener treats each row in the data as unique events in time, and creates frames with the correct events present at any given time.

tween_events(
  .data,
  ease,
  nframes,
  start,
  end = NULL,
  range = NULL,
  enter = NULL,
  exit = NULL,
  enter_length = 0,
  exit_length = 0
)

Arguments

.data

A data.frame with components at different stages

ease

The easing function to use. Either a single string or one for each column in the data set.

nframes

The number of frames to calculate for the tween

start, end

The start (and potential end) of the event encoded in the row, as unquoted expressions. Will be evaluated in the context of .data so can refer to columns in it. If end = NULL the event will be without extend and only visible in a single frame, unless enter and/or exit is given.

range

The range of time points to include in the tween. If NULL it will use the range of time

enter, exit

functions that calculate a start state for new observations that appear in to or an end state for observations that are not present in to. If NULL the new/old observations will not be part of the tween. The function gets a data.frame with either the start state of the exiting observations, or the end state of the entering observations and must return a modified version of that data.frame. See the Match, Enter, and Exit section for more information.

enter_length, exit_length

The lenght of the opening and closing transitions if enter and/or exit is given. Measured in the same units as time

Value

A data.frame with the same columns as .data along with .id giving the component id, .phase giving the state of each component in each frame, and .frame giving the frame membership of each row.

See also

Examples

d <- data.frame(
  x = runif(20),
  y = runif(20),
  time = runif(20),
  duration = runif(20, max = 0.1)
)
from_left <- function(x) {
  x$x <- -0.5
  x
}
to_right <- function(x) {
  x$x <- 1.5
  x
}

tween_events(d, 'cubic-in-out', 50, start = time, end = time + duration,
             enter = from_left, exit = to_right, enter_length = 0.1,
             exit_length = 0.05)
#>               x          y        time    duration .id .phase .frame
#> 1   -0.50000000 0.22369852 0.014098640 0.017011686   1  enter      1
#> 28  -0.50000000 0.36702648 0.006086655 0.055636500   4  enter      1
#> 140 -0.50000000 0.88061954 0.010749855 0.049096655  16  enter      1
#> 2   -0.46581137 0.22369852 0.014098640 0.017011686   1  enter      2
#> 29  -0.43478130 0.36702648 0.006086655 0.055636500   4  enter      2
#> 141 -0.42452895 0.88061954 0.010749855 0.049096655  16  enter      2
#> 3   -0.22649097 0.22369852 0.014098640 0.017011686   1  enter      3
#> 30   0.02174962 0.36702648 0.006086655 0.055636500   4  enter      3
#> 142  0.10376837 0.88061954 0.010749855 0.049096655  16  enter      3
#> 4    0.29488561 0.22369852 0.014098640 0.017011686   1  enter      4
#> 31   0.47828054 0.36702648 0.006086655 0.055636500   4  enter      4
#> 85  -0.50000000 0.26033049 0.073680921 0.009143642  10  enter      4
#> 143  0.63206569 0.88061954 0.010749855 0.049096655  16  enter      4
#> 5    0.53420601 0.22369852 0.014098640 0.017011686   1  enter      5
#> 32   0.54349925 0.36702648 0.006086655 0.055636500   4 static      5
#> 86  -0.40627469 0.26033049 0.073680921 0.009143642  10  enter      5
#> 144  0.70753674 0.88061954 0.010749855 0.049096655  16 static      5
#> 6    0.56839463 0.22369852 0.014098640 0.017011686   1    raw      6
#> 33   0.54349925 0.36702648 0.006086655 0.055636500   4 static      6
#> 87   0.24980247 0.26033049 0.073680921 0.009143642  10  enter      6
#> 145  0.70753674 0.88061954 0.010749855 0.049096655  16 static      6
#> 7    1.03419732 0.22369852 0.014098640 0.017011686   1   exit      7
#> 34   0.54349925 0.36702648 0.006086655 0.055636500   4 static      7
#> 88   0.90587964 0.26033049 0.073680921 0.009143642  10  enter      7
#> 146  0.70753674 0.88061954 0.010749855 0.049096655  16 static      7
#> 8    1.50000000 0.22369852 0.014098640 0.017011686   1   exit      8
#> 35   0.54349925 0.36702648 0.006086655 0.055636500   4    raw      8
#> 89   0.99960495 0.26033049 0.073680921 0.009143642  10 static      8
#> 147  0.70753674 0.88061954 0.010749855 0.049096655  16    raw      8
#> 36   1.02174962 0.36702648 0.006086655 0.055636500   4   exit      9
#> 90   0.99960495 0.26033049 0.073680921 0.009143642  10    raw      9
#> 148  1.10376837 0.88061954 0.010749855 0.049096655  16   exit      9
#> 37   1.50000000 0.36702648 0.006086655 0.055636500   4   exit     10
#> 91   1.24980247 0.26033049 0.073680921 0.009143642  10   exit     10
#> 149  1.50000000 0.88061954 0.010749855 0.049096655  16   exit     10
#> 92   1.50000000 0.26033049 0.073680921 0.009143642  10   exit     11
#> 76  -0.50000000 0.89912618 0.292433564 0.040977085   9  enter     13
#> 77  -0.45565367 0.89912618 0.292433564 0.040977085   9  enter     14
#> 78  -0.14522938 0.89912618 0.292433564 0.040977085   9  enter     15
#> 79   0.16519491 0.89912618 0.292433564 0.040977085   9  enter     16
#> 80   0.20954123 0.89912618 0.292433564 0.040977085   9 static     17
#> 81   0.20954123 0.89912618 0.292433564 0.040977085   9 static     18
#> 82   0.20954123 0.89912618 0.292433564 0.040977085   9    raw     19
#> 121 -0.50000000 0.01720221 0.426247393 0.080602921  14  enter     19
#> 83   0.85477062 0.89912618 0.292433564 0.040977085   9   exit     20
#> 122 -0.45170935 0.01720221 0.426247393 0.080602921  14  enter     20
#> 84   1.50000000 0.89912618 0.292433564 0.040977085   9   exit     21
#> 123 -0.11367477 0.01720221 0.426247393 0.080602921  14  enter     21
#> 124  0.22435980 0.01720221 0.426247393 0.080602921  14  enter     22
#> 47  -0.50000000 0.46363016 0.526965949 0.054201570   6  enter     23
#> 125  0.27265045 0.01720221 0.426247393 0.080602921  14 static     23
#> 175 -0.50000000 0.95900889 0.533913606 0.060180188  20  enter     23
#> 48  -0.41007262 0.46363016 0.526965949 0.054201570   6  enter     24
#> 126  0.27265045 0.01720221 0.426247393 0.080602921  14 static     24
#> 150 -0.50000000 0.39014530 0.540012266 0.033185250  17  enter     24
#> 176 -0.45978124 0.95900889 0.533913606 0.060180188  20  enter     24
#> 49   0.21941907 0.46363016 0.526965949 0.054201570   6  enter     25
#> 127  0.27265045 0.01720221 0.426247393 0.080602921  14 static     25
#> 151 -0.46474294 0.39014530 0.540012266 0.033185250  17  enter     25
#> 177 -0.17824994 0.95900889 0.533913606 0.060180188  20  enter     25
#> 50   0.84891075 0.46363016 0.526965949 0.054201570   6  enter     26
#> 128  0.27265045 0.01720221 0.426247393 0.080602921  14 static     26
#> 152 -0.21794355 0.39014530 0.540012266 0.033185250  17  enter     26
#> 178  0.43508611 0.95900889 0.533913606 0.060180188  20  enter     26
#> 51   0.93883813 0.46363016 0.526965949 0.054201570   6 static     27
#> 129  0.27265045 0.01720221 0.426247393 0.080602921  14    raw     27
#> 153  0.02885584 0.39014530 0.540012266 0.033185250  17  enter     27
#> 179  0.71661741 0.95900889 0.533913606 0.060180188  20  enter     27
#> 52   0.93883813 0.46363016 0.526965949 0.054201570   6 static     28
#> 130  0.88632523 0.01720221 0.426247393 0.080602921  14   exit     28
#> 154  0.06411290 0.39014530 0.540012266 0.033185250  17 static     28
#> 180  0.75683617 0.95900889 0.533913606 0.060180188  20 static     28
#> 38  -0.50000000 0.81912892 0.670848205 0.044649266   5  enter     29
#> 53   0.93883813 0.46363016 0.526965949 0.054201570   6 static     29
#> 57  -0.50000000 0.55009286 0.667798111 0.029869597   7  enter     29
#> 66  -0.50000000 0.33741755 0.660251011 0.061442820   8  enter     29
#> 113 -0.50000000 0.94333165 0.666746517 0.020160991  13  enter     29
#> 131  1.50000000 0.01720221 0.426247393 0.080602921  14   exit     29
#> 155  0.06411290 0.39014530 0.540012266 0.033185250  17    raw     29
#> 181  0.75683617 0.95900889 0.533913606 0.060180188  20 static     29
#> 39  -0.43236641 0.81912892 0.670848205 0.044649266   5  enter     30
#> 54   0.93883813 0.46363016 0.526965949 0.054201570   6    raw     30
#> 58  -0.42023098 0.55009286 0.667798111 0.029869597   7  enter     30
#> 67  -0.46036634 0.33741755 0.660251011 0.061442820   8  enter     30
#> 102 -0.50000000 0.72700733 0.682949542 0.095902759  12  enter     30
#> 114 -0.42707673 0.94333165 0.666746517 0.020160991  13  enter     30
#> 156  0.78205645 0.39014530 0.540012266 0.033185250  17   exit     30
#> 182  0.75683617 0.95900889 0.533913606 0.060180188  20    raw     30
#> 40   0.04106872 0.81912892 0.670848205 0.044649266   5  enter     31
#> 55   1.21941907 0.46363016 0.526965949 0.054201570   6   exit     31
#> 59   0.13815214 0.55009286 0.667798111 0.029869597   7  enter     31
#> 68  -0.18293071 0.33741755 0.660251011 0.061442820   8  enter     31
#> 103 -0.42627503 0.72700733 0.682949542 0.095902759  12  enter     31
#> 115  0.08338618 0.94333165 0.666746517 0.020160991  13  enter     31
#> 157  1.50000000 0.39014530 0.540012266 0.033185250  17   exit     31
#> 183  1.12841809 0.95900889 0.533913606 0.060180188  20   exit     31
#> 41   0.51450386 0.81912892 0.670848205 0.044649266   5  enter     32
#> 56   1.50000000 0.46363016 0.526965949 0.054201570   6   exit     32
#> 60   0.69653527 0.55009286 0.667798111 0.029869597   7  enter     32
#> 69   0.09450492 0.33741755 0.660251011 0.061442820   8  enter     32
#> 104  0.08979978 0.72700733 0.682949542 0.095902759  12  enter     32
#> 116  0.59384908 0.94333165 0.666746517 0.020160991  13  enter     32
#> 184  1.50000000 0.95900889 0.533913606 0.060180188  20   exit     32
#> 42   0.58213745 0.81912892 0.670848205 0.044649266   5 static     33
#> 61   0.77630429 0.55009286 0.667798111 0.029869597   7 static     33
#> 70   0.13413859 0.33741755 0.660251011 0.061442820   8 static     33
#> 105  0.60587459 0.72700733 0.682949542 0.095902759  12  enter     33
#> 117  0.66677235 0.94333165 0.666746517 0.020160991  13 static     33
#> 43   0.58213745 0.81912892 0.670848205 0.044649266   5 static     34
#> 62   0.77630429 0.55009286 0.667798111 0.029869597   7 static     34
#> 71   0.13413859 0.33741755 0.660251011 0.061442820   8 static     34
#> 106  0.67959956 0.72700733 0.682949542 0.095902759  12 static     34
#> 118  0.66677235 0.94333165 0.666746517 0.020160991  13    raw     34
#> 158 -0.50000000 0.22306353 0.793169803 0.024060011  18  enter     34
#> 167 -0.50000000 0.41325562 0.783561200 0.022625757  19  enter     34
#> 9   -0.50000000 0.16465645 0.808036254 0.018201484   2  enter     35
#> 44   0.58213745 0.81912892 0.670848205 0.044649266   5    raw     35
#> 63   0.77630429 0.55009286 0.667798111 0.029869597   7    raw     35
#> 72   0.13413859 0.33741755 0.660251011 0.061442820   8 static     35
#> 107  0.67959956 0.72700733 0.682949542 0.095902759  12 static     35
#> 119  1.08338618 0.94333165 0.666746517 0.020160991  13   exit     35
#> 159 -0.45235789 0.22306353 0.793169803 0.024060011  18  enter     35
#> 168 -0.40847446 0.41325562 0.783561200 0.022625757  19  enter     35
#> 10  -0.43502858 0.16465645 0.808036254 0.018201484   2  enter     36
#> 45   1.04106872 0.81912892 0.670848205 0.044649266   5   exit     36
#> 64   1.13815214 0.55009286 0.667798111 0.029869597   7   exit     36
#> 73   0.13413859 0.33741755 0.660251011 0.061442820   8    raw     36
#> 108  0.67959956 0.72700733 0.682949542 0.095902759  12 static     36
#> 120  1.50000000 0.94333165 0.666746517 0.020160991  13   exit     36
#> 160 -0.11886314 0.22306353 0.793169803 0.024060011  18  enter     36
#> 169  0.23220430 0.41325562 0.783561200 0.022625757  19  enter     36
#> 11   0.01977134 0.16465645 0.808036254 0.018201484   2  enter     37
#> 46   1.50000000 0.81912892 0.670848205 0.044649266   5   exit     37
#> 65   1.50000000 0.55009286 0.667798111 0.029869597   7   exit     37
#> 74   0.81706929 0.33741755 0.660251011 0.061442820   8   exit     37
#> 93  -0.50000000 0.14402975 0.844570943 0.036298348  11  enter     37
#> 109  0.67959956 0.72700733 0.682949542 0.095902759  12 static     37
#> 161  0.60767899 0.22306353 0.793169803 0.024060011  18  enter     37
#> 170  0.87288307 0.41325562 0.783561200 0.022625757  19  enter     37
#> 12   0.47457126 0.16465645 0.808036254 0.018201484   2  enter     38
#> 75   1.50000000 0.33741755 0.660251011 0.061442820   8   exit     38
#> 94  -0.46397059 0.14402975 0.844570943 0.036298348  11  enter     38
#> 110  0.67959956 0.72700733 0.682949542 0.095902759  12    raw     38
#> 162  0.94117373 0.22306353 0.793169803 0.024060011  18  enter     38
#> 171  0.96440861 0.41325562 0.783561200 0.022625757  19 static     38
#> 13   0.53954267 0.16465645 0.808036254 0.018201484   2 static     39
#> 95  -0.21176474 0.14402975 0.844570943 0.036298348  11  enter     39
#> 111  1.08979978 0.72700733 0.682949542 0.095902759  12   exit     39
#> 163  0.98881584 0.22306353 0.793169803 0.024060011  18 static     39
#> 172  0.96440861 0.41325562 0.783561200 0.022625757  19    raw     39
#> 14   0.53954267 0.16465645 0.808036254 0.018201484   2    raw     40
#> 17  -0.50000000 0.66641463 0.928605680 0.081658729   3  enter     40
#> 96   0.04044112 0.14402975 0.844570943 0.036298348  11  enter     40
#> 112  1.50000000 0.72700733 0.682949542 0.095902759  12   exit     40
#> 164  0.98881584 0.22306353 0.793169803 0.024060011  18    raw     40
#> 173  1.23220430 0.41325562 0.783561200 0.022625757  19   exit     40
#> 15   1.01977134 0.16465645 0.808036254 0.018201484   2   exit     41
#> 18  -0.41442765 0.66641463 0.928605680 0.081658729   3  enter     41
#> 97   0.07647052 0.14402975 0.844570943 0.036298348  11 static     41
#> 132 -0.50000000 0.39267707 0.950042350 0.010369353  15  enter     41
#> 165  1.24440792 0.22306353 0.793169803 0.024060011  18   exit     41
#> 174  1.50000000 0.41325562 0.783561200 0.022625757  19   exit     41
#> 16   1.50000000 0.16465645 0.808036254 0.018201484   2   exit     42
#> 19   0.18457883 0.66641463 0.928605680 0.081658729   3  enter     42
#> 98   0.07647052 0.14402975 0.844570943 0.036298348  11    raw     42
#> 133 -0.46147345 0.39267707 0.950042350 0.010369353  15  enter     42
#> 166  1.50000000 0.22306353 0.793169803 0.024060011  18   exit     42
#> 20   0.78358531 0.66641463 0.928605680 0.081658729   3  enter     43
#> 99   0.28736378 0.14402975 0.844570943 0.036298348  11   exit     43
#> 134 -0.19178762 0.39267707 0.950042350 0.010369353  15  enter     43
#> 21   0.86915766 0.66641463 0.928605680 0.081658729   3 static     44
#> 100  1.28910674 0.14402975 0.844570943 0.036298348  11   exit     44
#> 135  0.07789821 0.39267707 0.950042350 0.010369353  15  enter     44
#> 22   0.86915766 0.66641463 0.928605680 0.081658729   3 static     45
#> 101  1.50000000 0.14402975 0.844570943 0.036298348  11   exit     45
#> 136  0.11642476 0.39267707 0.950042350 0.010369353  15 static     45
#> 23   0.86915766 0.66641463 0.928605680 0.081658729   3 static     46
#> 137  0.11642476 0.39267707 0.950042350 0.010369353  15    raw     46
#> 24   0.86915766 0.66641463 0.928605680 0.081658729   3 static     47
#> 138  0.80821238 0.39267707 0.950042350 0.010369353  15   exit     47
#> 25   0.86915766 0.66641463 0.928605680 0.081658729   3    raw     48
#> 139  1.50000000 0.39267707 0.950042350 0.010369353  15   exit     48
#> 26   1.18457883 0.66641463 0.928605680 0.081658729   3   exit     49
#> 27   1.50000000 0.66641463 0.928605680 0.081658729   3   exit     50