Transition in and out of events

This tweening function is a more powerful version of tween_appear(), with support for newer features such as enter/exits and tween phase identification. The tweener treats each row in the data as unique events in time, and creates frames with the correct events present at any given time.

tween_events(
  .data,
  ease,
  nframes,
  start,
  end = NULL,
  range = NULL,
  enter = NULL,
  exit = NULL,
  enter_length = 0,
  exit_length = 0
)

Arguments

.data

A data.frame with components at different stages

ease

The easing function to use. Either a single string or one for each column in the data set.

nframes

The number of frames to calculate for the tween

start, end

The start (and potential end) of the event encoded in the row, as unquoted expressions. Will be evaluated in the context of .data so can refer to columns in it. If end = NULL the event will be without extend and only visible in a single frame, unless enter and/or exit is given.

range

The range of time points to include in the tween. If NULL it will use the range of time

enter, exit

functions that calculate a start state for new observations that appear in to or an end state for observations that are not present in to. If NULL the new/old observations will not be part of the tween. The function gets a data.frame with either the start state of the exiting observations, or the end state of the entering observations and must return a modified version of that data.frame. See the Match, Enter, and Exit section for more information.

enter_length, exit_length

The lenght of the opening and closing transitions if enter and/or exit is given. Measured in the same units as time

Value

A data.frame with the same columns as .data along with .id giving the component id, .phase giving the state of each component in each frame, and .frame giving the frame membership of each row.

See also

Examples

d <- data.frame(
  x = runif(20),
  y = runif(20),
  time = runif(20),
  duration = runif(20, max = 0.1)
)
from_left <- function(x) {
  x$x <- -0.5
  x
}
to_right <- function(x) {
  x$x <- 1.5
  x
}

tween_events(d, 'cubic-in-out', 50, start = time, end = time + duration,
             enter = from_left, exit = to_right, enter_length = 0.1,
             exit_length = 0.05)
#>                x            y       time    duration .id .phase .frame
#> 112 -0.500000000 0.8967243955 0.02954076 0.040435478  13  enter      1
#> 113 -0.477330089 0.8967243955 0.02954076 0.040435478  13  enter      2
#> 114 -0.318640715 0.8967243955 0.02954076 0.040435478  13  enter      3
#> 94  -0.500000000 0.7886738486 0.10084201 0.035132090  11  enter      4
#> 115  0.027075423 0.8967243955 0.02954076 0.040435478  13  enter      4
#> 95  -0.472014967 0.7886738486 0.10084201 0.035132090  11  enter      5
#> 116  0.185764798 0.8967243955 0.02954076 0.040435478  13  enter      5
#> 96  -0.276119740 0.7886738486 0.10084201 0.035132090  11  enter      6
#> 117  0.208434708 0.8967243955 0.02954076 0.040435478  13 static      6
#> 97   0.150652006 0.7886738486 0.10084201 0.035132090  11  enter      7
#> 118  0.208434708 0.8967243955 0.02954076 0.040435478  13    raw      7
#> 98   0.346547233 0.7886738486 0.10084201 0.035132090  11  enter      8
#> 119  0.399777715 0.8967243955 0.02954076 0.040435478  13   exit      8
#> 9   -0.500000000 0.3985216257 0.19840863 0.005263014   2  enter      9
#> 99   0.374532266 0.7886738486 0.10084201 0.035132090  11 static      9
#> 120  1.308656994 0.8967243955 0.02954076 0.040435478  13   exit      9
#> 10  -0.417885417 0.3985216257 0.19840863 0.005263014   2  enter     10
#> 74  -0.500000000 0.0895739065 0.23523691 0.042302567   9  enter     10
#> 100  0.374532266 0.7886738486 0.10084201 0.035132090  11    raw     10
#> 104 -0.500000000 0.7694666549 0.23048492 0.007305671  12  enter     10
#> 121  1.500000000 0.8967243955 0.02954076 0.040435478  13   exit     10
#> 151 -0.500000000 0.8193608329 0.23131222 0.030041935  17  enter     10
#> 11   0.156916666 0.3985216257 0.19840863 0.005263014   2  enter     11
#> 75  -0.478065207 0.0895739065 0.23523691 0.042302567   9  enter     11
#> 101  0.541268226 0.7886738486 0.10084201 0.035132090  11   exit     11
#> 105 -0.478037383 0.7694666549 0.23048492 0.007305671  12  enter     11
#> 152 -0.465507616 0.8193608329 0.23131222 0.030041935  17  enter     11
#> 12   0.731718749 0.3985216257 0.19840863 0.005263014   2  enter     12
#> 76  -0.324521657 0.0895739065 0.23523691 0.042302567   9  enter     12
#> 102  1.333264039 0.7886738486 0.10084201 0.035132090  11   exit     12
#> 106 -0.324299064 0.7694666549 0.23048492 0.007305671  12  enter     12
#> 153 -0.224060925 0.8193608329 0.23131222 0.030041935  17  enter     12
#> 13   0.813833332 0.3985216257 0.19840863 0.005263014   2    raw     13
#> 77   0.009983935 0.0895739065 0.23523691 0.042302567   9  enter     13
#> 103  1.500000000 0.7886738486 0.10084201 0.035132090  11   exit     13
#> 107  0.010630846 0.7694666549 0.23048492 0.007305671  12  enter     13
#> 154  0.301947937 0.8193608329 0.23131222 0.030041935  17  enter     13
#> 14   0.915487653 0.3985216257 0.19840863 0.005263014   2   exit     14
#> 78   0.163527485 0.0895739065 0.23523691 0.042302567   9  enter     14
#> 108  0.164369165 0.7694666549 0.23048492 0.007305671  12  enter     14
#> 155  0.543394628 0.8193608329 0.23131222 0.030041935  17  enter     14
#> 15   1.398345679 0.3985216257 0.19840863 0.005263014   2   exit     15
#> 79   0.185462278 0.0895739065 0.23523691 0.042302567   9 static     15
#> 109  0.186331782 0.7694666549 0.23048492 0.007305671  12    raw     15
#> 156  0.577887012 0.8193608329 0.23131222 0.030041935  17 static     15
#> 1   -0.500000000 0.2463958098 0.35310088 0.012489818   1  enter     16
#> 16   1.500000000 0.3985216257 0.19840863 0.005263014   2   exit     16
#> 80   0.185462278 0.0895739065 0.23523691 0.042302567   9 static     16
#> 84  -0.500000000 0.1052521600 0.36808534 0.046760149  10  enter     16
#> 110  0.843165891 0.7694666549 0.23048492 0.007305671  12   exit     16
#> 122 -0.500000000 0.5378663200 0.36233559 0.010359605  14  enter     16
#> 157  0.577887012 0.8193608329 0.23131222 0.030041935  17    raw     16
#> 2   -0.431667379 0.2463958098 0.35310088 0.012489818   1  enter     17
#> 81   0.185462278 0.0895739065 0.23523691 0.042302567   9    raw     17
#> 85  -0.455952566 0.1052521600 0.36808534 0.046760149  10  enter     17
#> 111  1.500000000 0.7694666549 0.23048492 0.007305671  12   exit     17
#> 123 -0.482830321 0.5378663200 0.36233559 0.010359605  14  enter     17
#> 158  1.038943506 0.8193608329 0.23131222 0.030041935  17   exit     17
#> 3    0.046660966 0.2463958098 0.35310088 0.012489818   1  enter     18
#> 82   0.842731139 0.0895739065 0.23523691 0.042302567   9   exit     18
#> 86  -0.147620527 0.1052521600 0.36808534 0.046760149  10  enter     18
#> 124 -0.362642565 0.5378663200 0.36233559 0.010359605  14  enter     18
#> 159  1.500000000 0.8193608329 0.23131222 0.030041935  17   exit     18
#> 4    0.524989312 0.2463958098 0.35310088 0.012489818   1  enter     19
#> 83   1.500000000 0.0895739065 0.23523691 0.042302567   9   exit     19
#> 87   0.524102844 0.1052521600 0.36808534 0.046760149  10  enter     19
#> 125 -0.100804954 0.5378663200 0.36233559 0.010359605  14  enter     19
#> 5    0.593321932 0.2463958098 0.35310088 0.012489818   1 static     20
#> 88   0.832434883 0.1052521600 0.36808534 0.046760149  10  enter     20
#> 126  0.019382802 0.5378663200 0.36233559 0.010359605  14  enter     20
#> 6    0.593321932 0.2463958098 0.35310088 0.012489818   1    raw     21
#> 89   0.876482317 0.1052521600 0.36808534 0.046760149  10 static     21
#> 127  0.036552482 0.5378663200 0.36233559 0.010359605  14    raw     21
#> 7    1.046660966 0.2463958098 0.35310088 0.012489818   1   exit     22
#> 43  -0.500000000 0.0005360937 0.48175362 0.053483600   6  enter     22
#> 90   0.876482317 0.1052521600 0.36808534 0.046760149  10 static     22
#> 128  0.768276241 0.5378663200 0.36233559 0.010359605  14   exit     22
#> 8    1.500000000 0.2463958098 0.35310088 0.012489818   1   exit     23
#> 35  -0.500000000 0.6999949529 0.50591472 0.007447603   5  enter     23
#> 44  -0.461650707 0.0005360937 0.48175362 0.053483600   6  enter     23
#> 91   0.876482317 0.1052521600 0.36808534 0.046760149  10    raw     23
#> 129  1.500000000 0.5378663200 0.36233559 0.010359605  14   exit     23
#> 36  -0.434552111 0.6999949529 0.50591472 0.007447603   5  enter     24
#> 45  -0.193205658 0.0005360937 0.48175362 0.053483600   6  enter     24
#> 92   1.188241158 0.1052521600 0.36808534 0.046760149  10   exit     24
#> 37   0.023583115 0.6999949529 0.50591472 0.007447603   5  enter     25
#> 46   0.075239391 0.0005360937 0.48175362 0.053483600   6  enter     25
#> 93   1.500000000 0.1052521600 0.36808534 0.046760149  10   exit     25
#> 130 -0.500000000 0.3037101415 0.55556479 0.074563833  15  enter     25
#> 38   0.481718340 0.6999949529 0.50591472 0.007447603   5  enter     26
#> 47   0.113588684 0.0005360937 0.48175362 0.053483600   6 static     26
#> 131 -0.455493672 0.3037101415 0.55556479 0.074563833  15  enter     26
#> 39   0.547166229 0.6999949529 0.50591472 0.007447603   5 static     27
#> 48   0.113588684 0.0005360937 0.48175362 0.053483600   6 static     27
#> 132 -0.143949373 0.3037101415 0.55556479 0.074563833  15  enter     27
#> 40   0.547166229 0.6999949529 0.50591472 0.007447603   5    raw     28
#> 49   0.113588684 0.0005360937 0.48175362 0.053483600   6 static     28
#> 133  0.534772133 0.3037101415 0.55556479 0.074563833  15  enter     28
#> 41   1.023583115 0.6999949529 0.50591472 0.007447603   5   exit     29
#> 50   0.113588684 0.0005360937 0.48175362 0.053483600   6    raw     29
#> 134  0.846316432 0.3037101415 0.55556479 0.074563833  15  enter     29
#> 26  -0.500000000 0.2218419353 0.66447404 0.035749231   4  enter     30
#> 42   1.500000000 0.6999949529 0.50591472 0.007447603   5   exit     30
#> 51   0.806794342 0.0005360937 0.48175362 0.053483600   6   exit     30
#> 135  0.890822760 0.3037101415 0.55556479 0.074563833  15 static     30
#> 179 -0.500000000 0.3600092146 0.67551345 0.089175946  20  enter     30
#> 27  -0.412968370 0.2218419353 0.66447404 0.035749231   4  enter     31
#> 52   1.500000000 0.0005360937 0.48175362 0.053483600   6   exit     31
#> 136  0.890822760 0.3037101415 0.55556479 0.074563833  15 static     31
#> 141 -0.500000000 0.2147969797 0.67831326 0.053841486  16  enter     31
#> 180 -0.469458503 0.3600092146 0.67551345 0.089175946  20  enter     31
#> 28   0.196253036 0.2218419353 0.66447404 0.035749231   4  enter     32
#> 137  0.890822760 0.3037101415 0.55556479 0.074563833  15 static     32
#> 142 -0.431970188 0.2147969797 0.67831326 0.053841486  16  enter     32
#> 181 -0.255668023 0.3600092146 0.67551345 0.089175946  20  enter     32
#> 29   0.805474443 0.2218419353 0.66447404 0.035749231   4  enter     33
#> 138  0.890822760 0.3037101415 0.55556479 0.074563833  15    raw     33
#> 143  0.044238495 0.2147969797 0.67831326 0.053841486  16  enter     33
#> 182  0.210089809 0.3600092146 0.67551345 0.089175946  20  enter     33
#> 30   0.892506073 0.2218419353 0.66447404 0.035749231   4 static     34
#> 139  1.195411380 0.3037101415 0.55556479 0.074563833  15   exit     34
#> 144  0.520447179 0.2147969797 0.67831326 0.053841486  16  enter     34
#> 183  0.423880289 0.3600092146 0.67551345 0.089175946  20  enter     34
#> 31   0.892506073 0.2218419353 0.66447404 0.035749231   4 static     35
#> 140  1.500000000 0.3037101415 0.55556479 0.074563833  15   exit     35
#> 145  0.588476991 0.2147969797 0.67831326 0.053841486  16 static     35
#> 184  0.454421786 0.3600092146 0.67551345 0.089175946  20 static     35
#> 32   0.892506073 0.2218419353 0.66447404 0.035749231   4    raw     36
#> 146  0.588476991 0.2147969797 0.67831326 0.053841486  16 static     36
#> 160 -0.500000000 0.3667614984 0.79239809 0.053369588  18  enter     36
#> 185  0.454421786 0.3600092146 0.67551345 0.089175946  20 static     36
#> 33   1.196253036 0.2218419353 0.66447404 0.035749231   4   exit     37
#> 147  0.588476991 0.2147969797 0.67831326 0.053841486  16 static     37
#> 161 -0.447451182 0.3667614984 0.79239809 0.053369588  18  enter     37
#> 186  0.454421786 0.3600092146 0.67551345 0.089175946  20 static     37
#> 17  -0.500000000 0.7552454052 0.83057943 0.026780652   3  enter     38
#> 34   1.500000000 0.2218419353 0.66447404 0.035749231   4   exit     38
#> 148  0.588476991 0.2147969797 0.67831326 0.053841486  16    raw     38
#> 162 -0.079609453 0.3667614984 0.79239809 0.053369588  18  enter     38
#> 187  0.454421786 0.3600092146 0.67551345 0.089175946  20 static     38
#> 18  -0.429617470 0.7552454052 0.83057943 0.026780652   3  enter     39
#> 149  1.044238495 0.2147969797 0.67831326 0.053841486  16   exit     39
#> 163  0.288232275 0.3667614984 0.79239809 0.053369588  18  enter     39
#> 188  0.454421786 0.3600092146 0.67551345 0.089175946  20    raw     39
#> 19   0.063060237 0.7552454052 0.83057943 0.026780652   3  enter     40
#> 53  -0.500000000 0.8068563868 0.89560784 0.043131437   7  enter     40
#> 63  -0.500000000 0.2884054463 0.89276196 0.062064298   8  enter     40
#> 150  1.500000000 0.2147969797 0.67831326 0.053841486  16   exit     40
#> 164  0.340781093 0.3667614984 0.79239809 0.053369588  18 static     40
#> 170 -0.500000000 0.4238340650 0.89055622 0.021907877  19  enter     40
#> 189  0.977210893 0.3600092146 0.67551345 0.089175946  20   exit     40
#> 20   0.555737945 0.7552454052 0.83057943 0.026780652   3  enter     41
#> 54  -0.463528620 0.8068563868 0.89560784 0.043131437   7  enter     41
#> 64  -0.482404214 0.2884054463 0.89276196 0.062064298   8  enter     41
#> 165  0.340781093 0.3667614984 0.79239809 0.053369588  18 static     41
#> 171 -0.456033278 0.4238340650 0.89055622 0.021907877  19  enter     41
#> 190  1.500000000 0.3600092146 0.67551345 0.089175946  20   exit     41
#> 21   0.626120474 0.7552454052 0.83057943 0.026780652   3 static     42
#> 55  -0.208228961 0.8068563868 0.89560784 0.043131437   7  enter     42
#> 65  -0.359233712 0.2884054463 0.89276196 0.062064298   8  enter     42
#> 166  0.340781093 0.3667614984 0.79239809 0.053369588  18 static     42
#> 172 -0.148266227 0.4238340650 0.89055622 0.021907877  19  enter     42
#> 22   0.626120474 0.7552454052 0.83057943 0.026780652   3    raw     43
#> 56   0.347959581 0.8068563868 0.89560784 0.043131437   7  enter     43
#> 66  -0.090897976 0.2884054463 0.89276196 0.062064298   8  enter     43
#> 167  0.340781093 0.3667614984 0.79239809 0.053369588  18    raw     43
#> 173  0.522226277 0.4238340650 0.89055622 0.021907877  19  enter     43
#> 23   0.755584108 0.7552454052 0.83057943 0.026780652   3   exit     44
#> 57   0.603259240 0.8068563868 0.89560784 0.043131437   7  enter     44
#> 67   0.032272525 0.2884054463 0.89276196 0.062064298   8  enter     44
#> 168  0.920390547 0.3667614984 0.79239809 0.053369588  18   exit     44
#> 174  0.829993328 0.4238340650 0.89055622 0.021907877  19  enter     44
#> 24   1.370536367 0.7552454052 0.83057943 0.026780652   3   exit     45
#> 58   0.639730620 0.8068563868 0.89560784 0.043131437   7 static     45
#> 68   0.049868311 0.2884054463 0.89276196 0.062064298   8 static     45
#> 169  1.500000000 0.3667614984 0.79239809 0.053369588  18   exit     45
#> 175  0.873960049 0.4238340650 0.89055622 0.021907877  19 static     45
#> 25   1.500000000 0.7552454052 0.83057943 0.026780652   3   exit     46
#> 59   0.639730620 0.8068563868 0.89560784 0.043131437   7 static     46
#> 69   0.049868311 0.2884054463 0.89276196 0.062064298   8 static     46
#> 176  0.873960049 0.4238340650 0.89055622 0.021907877  19    raw     46
#> 60   0.639730620 0.8068563868 0.89560784 0.043131437   7    raw     47
#> 70   0.049868311 0.2884054463 0.89276196 0.062064298   8 static     47
#> 177  1.186980025 0.4238340650 0.89055622 0.021907877  19   exit     47
#> 61   1.069865310 0.8068563868 0.89560784 0.043131437   7   exit     48
#> 71   0.049868311 0.2884054463 0.89276196 0.062064298   8    raw     48
#> 178  1.500000000 0.4238340650 0.89055622 0.021907877  19   exit     48
#> 62   1.500000000 0.8068563868 0.89560784 0.043131437   7   exit     49
#> 72   0.774934156 0.2884054463 0.89276196 0.062064298   8   exit     49
#> 73   1.500000000 0.2884054463 0.89276196 0.062064298   8   exit     50