Transition in and out of events

This tweening function is a more powerful version of tween_appear(), with support for newer features such as enter/exits and tween phase identification. The tweener treats each row in the data as unique events in time, and creates frames with the correct events present at any given time.

tween_events(
  .data,
  ease,
  nframes,
  start,
  end = NULL,
  range = NULL,
  enter = NULL,
  exit = NULL,
  enter_length = 0,
  exit_length = 0
)

Arguments

.data

A data.frame with components at different stages

ease

The easing function to use. Either a single string or one for each column in the data set.

nframes

The number of frames to calculate for the tween

start, end

The start (and potential end) of the event encoded in the row, as unquoted expressions. Will be evaluated in the context of .data so can refer to columns in it. If end = NULL the event will be without extend and only visible in a single frame, unless enter and/or exit is given.

range

The range of time points to include in the tween. If NULL it will use the range of time

enter, exit

functions that calculate a start state for new observations that appear in to or an end state for observations that are not present in to. If NULL the new/old observations will not be part of the tween. The function gets a data.frame with either the start state of the exiting observations, or the end state of the entering observations and must return a modified version of that data.frame. See the Match, Enter, and Exit section for more information.

enter_length, exit_length

The lenght of the opening and closing transitions if enter and/or exit is given. Measured in the same units as time

Value

A data.frame with the same columns as .data along with .id giving the component id, .phase giving the state of each component in each frame, and .frame giving the frame membership of each row.

See also

Examples

d <- data.frame(
  x = runif(20),
  y = runif(20),
  time = runif(20),
  duration = runif(20, max = 0.1)
)
from_left <- function(x) {
  x$x <- -0.5
  x
}
to_right <- function(x) {
  x$x <- 1.5
  x
}

tween_events(d, 'cubic-in-out', 50, start = time, end = time + duration,
             enter = from_left, exit = to_right, enter_length = 0.1,
             exit_length = 0.05)
#>                x          y       time    duration .id .phase .frame
#> 38  -0.500000000 0.21269498 0.04756962 0.040310297   5  enter      1
#> 11  -0.500000000 0.86849281 0.06861055 0.007175141   2  enter      2
#> 39  -0.420696688 0.21269498 0.04756962 0.040310297   5  enter      2
#> 12  -0.415110977 0.86849281 0.06861055 0.007175141   2  enter      3
#> 40   0.134426498 0.21269498 0.04756962 0.040310297   5  enter      3
#> 13   0.179112184 0.86849281 0.06861055 0.007175141   2  enter      4
#> 41   0.689549684 0.21269498 0.04756962 0.040310297   5  enter      4
#> 14   0.773335346 0.86849281 0.06861055 0.007175141   2  enter      5
#> 42   0.768852996 0.21269498 0.04756962 0.040310297   5 static      5
#> 169 -0.500000000 0.05898969 0.14639181 0.002609718  18  enter      5
#> 15   0.858224369 0.86849281 0.06861055 0.007175141   2 static      6
#> 43   0.768852996 0.21269498 0.04756962 0.040310297   5 static      6
#> 77  -0.500000000 0.38000985 0.16034686 0.058068283   9  enter      6
#> 170 -0.460720886 0.05898969 0.14639181 0.002609718  18  enter      6
#> 16   0.858224369 0.86849281 0.06861055 0.007175141   2    raw      7
#> 44   0.768852996 0.21269498 0.04756962 0.040310297   5    raw      7
#> 78  -0.483947686 0.38000985 0.16034686 0.058068283   9  enter      7
#> 171 -0.185767087 0.05898969 0.14639181 0.002609718  18  enter      7
#> 17   1.179112184 0.86849281 0.06861055 0.007175141   2   exit      8
#> 45   0.877171071 0.21269498 0.04756962 0.040310297   5   exit      8
#> 79  -0.371581487 0.38000985 0.16034686 0.058068283   9  enter      8
#> 119 -0.500000000 0.99175951 0.20367474 0.002437521  13  enter      8
#> 172  0.413239403 0.05898969 0.14639181 0.002609718  18  enter      8
#> 18   1.500000000 0.86849281 0.06861055 0.007175141   2   exit      9
#> 46   1.391681925 0.21269498 0.04756962 0.040310297   5   exit      9
#> 48  -0.500000000 0.58885715 0.22009549 0.049507835   6  enter      9
#> 80  -0.126783696 0.38000985 0.16034686 0.058068283   9  enter      9
#> 120 -0.408903075 0.99175951 0.20367474 0.002437521  13  enter      9
#> 173  0.688193201 0.05898969 0.14639181 0.002609718  18  enter      9
#> 47   1.500000000 0.21269498 0.04756962 0.040310297   5   exit     10
#> 49  -0.451559401 0.58885715 0.22009549 0.049507835   6  enter     10
#> 81  -0.014417497 0.38000985 0.16034686 0.058068283   9  enter     10
#> 121  0.228775400 0.99175951 0.20367474 0.002437521  13  enter     10
#> 174  0.727472316 0.05898969 0.14639181 0.002609718  18    raw     10
#> 50  -0.112475205 0.58885715 0.22009549 0.049507835   6  enter     11
#> 82   0.001634817 0.38000985 0.16034686 0.058068283   9 static     11
#> 122  0.866453875 0.99175951 0.20367474 0.002437521  13  enter     11
#> 175  1.113736158 0.05898969 0.14639181 0.002609718  18   exit     11
#> 19  -0.500000000 0.32716288 0.29021491 0.021835057   3  enter     12
#> 51   0.226608990 0.58885715 0.22009549 0.049507835   6  enter     12
#> 83   0.001634817 0.38000985 0.16034686 0.058068283   9 static     12
#> 123  0.957550800 0.99175951 0.20367474 0.002437521  13 static     12
#> 176  1.500000000 0.05898969 0.14639181 0.002609718  18   exit     12
#> 20  -0.467789805 0.32716288 0.29021491 0.021835057   3  enter     13
#> 52   0.275049590 0.58885715 0.22009549 0.049507835   6 static     13
#> 84   0.001634817 0.38000985 0.16034686 0.058068283   9    raw     13
#> 124  0.957550800 0.99175951 0.20367474 0.002437521  13    raw     13
#> 21  -0.242318443 0.32716288 0.29021491 0.021835057   3  enter     14
#> 28  -0.500000000 0.05213590 0.33185730 0.062583699   4  enter     14
#> 53   0.275049590 0.58885715 0.22009549 0.049507835   6 static     14
#> 85   0.750817408 0.38000985 0.16034686 0.058068283   9   exit     14
#> 125  1.228775400 0.99175951 0.20367474 0.002437521  13   exit     14
#> 158 -0.500000000 0.14173536 0.34290293 0.067361505  17  enter     14
#> 22  -0.016847082 0.32716288 0.29021491 0.021835057   3  enter     15
#> 29  -0.435937384 0.05213590 0.33185730 0.062583699   4  enter     15
#> 54   0.275049590 0.58885715 0.22009549 0.049507835   6    raw     15
#> 86   1.500000000 0.38000985 0.16034686 0.058068283   9   exit     15
#> 126  1.500000000 0.99175951 0.20367474 0.002437521  13   exit     15
#> 159 -0.470040710 0.14173536 0.34290293 0.067361505  17  enter     15
#> 23   0.015363113 0.32716288 0.29021491 0.021835057   3 static     16
#> 30   0.012500926 0.05213590 0.33185730 0.062583699   4  enter     16
#> 55   0.456523725 0.58885715 0.22009549 0.049507835   6   exit     16
#> 160 -0.260325679 0.14173536 0.34290293 0.067361505  17  enter     16
#> 24   0.015363113 0.32716288 0.29021491 0.021835057   3    raw     17
#> 31   0.460939235 0.05213590 0.33185730 0.062583699   4  enter     17
#> 56   1.318525865 0.58885715 0.22009549 0.049507835   6   exit     17
#> 161  0.196553495 0.14173536 0.34290293 0.067361505  17  enter     17
#> 25   0.235309319 0.32716288 0.29021491 0.021835057   3   exit     18
#> 32   0.525001851 0.05213590 0.33185730 0.062583699   4 static     18
#> 57   1.500000000 0.58885715 0.22009549 0.049507835   6   exit     18
#> 162  0.406268526 0.14173536 0.34290293 0.067361505  17  enter     18
#> 26   1.280053795 0.32716288 0.29021491 0.021835057   3   exit     19
#> 33   0.525001851 0.05213590 0.33185730 0.062583699   4 static     19
#> 147 -0.500000000 0.23434229 0.44656472 0.067479455  16  enter     19
#> 163  0.436227816 0.14173536 0.34290293 0.067361505  17 static     19
#> 177 -0.500000000 0.37900468 0.45838146 0.013171037  19  enter     19
#> 27   1.500000000 0.32716288 0.29021491 0.021835057   3   exit     20
#> 34   0.525001851 0.05213590 0.33185730 0.062583699   4 static     20
#> 148 -0.407462385 0.23434229 0.44656472 0.067479455  16  enter     20
#> 164  0.436227816 0.14173536 0.34290293 0.067361505  17 static     20
#> 178 -0.471235702 0.37900468 0.45838146 0.013171037  19  enter     20
#> 35   0.525001851 0.05213590 0.33185730 0.062583699   4    raw     21
#> 149  0.240300920 0.23434229 0.44656472 0.067479455  16  enter     21
#> 165  0.436227816 0.14173536 0.34290293 0.067361505  17 static     21
#> 179 -0.269885616 0.37900468 0.45838146 0.013171037  19  enter     21
#> 186 -0.500000000 0.29750579 0.49012306 0.070524061  20  enter     21
#> 36   1.012500926 0.05213590 0.33185730 0.062583699   4   exit     22
#> 150  0.888064225 0.23434229 0.44656472 0.067479455  16  enter     22
#> 166  0.436227816 0.14173536 0.34290293 0.067361505  17    raw     22
#> 180  0.168769929 0.37900468 0.45838146 0.013171037  19  enter     22
#> 187 -0.433960424 0.29750579 0.49012306 0.070524061  20  enter     22
#> 37   1.500000000 0.05213590 0.33185730 0.062583699   4   exit     23
#> 151  0.980601840 0.23434229 0.44656472 0.067479455  16 static     23
#> 167  0.968113908 0.14173536 0.34290293 0.067361505  17   exit     23
#> 181  0.370120015 0.37900468 0.45838146 0.013171037  19  enter     23
#> 188  0.028316605 0.29750579 0.49012306 0.070524061  20  enter     23
#> 152  0.980601840 0.23434229 0.44656472 0.067479455  16 static     24
#> 168  1.500000000 0.14173536 0.34290293 0.067361505  17   exit     24
#> 182  0.398884313 0.37900468 0.45838146 0.013171037  19    raw     24
#> 189  0.490593634 0.29750579 0.49012306 0.070524061  20  enter     24
#> 1   -0.500000000 0.31529975 0.58437079 0.049774764   1  enter     25
#> 58  -0.500000000 0.56646479 0.57395511 0.048203051   7  enter     25
#> 153  0.980601840 0.23434229 0.44656472 0.067479455  16 static     25
#> 183  0.562012563 0.37900468 0.45838146 0.013171037  19   exit     25
#> 190  0.556633209 0.29750579 0.49012306 0.070524061  20 static     25
#> 2   -0.470026901 0.31529975 0.58437079 0.049774764   1  enter     26
#> 59  -0.450954314 0.56646479 0.57395511 0.048203051   7  enter     26
#> 87  -0.500000000 0.73179445 0.61521229 0.029676887  10  enter     26
#> 154  0.980601840 0.23434229 0.44656472 0.067479455  16    raw     26
#> 184  1.336871750 0.37900468 0.45838146 0.013171037  19   exit     26
#> 191  0.556633209 0.29750579 0.49012306 0.070524061  20 static     26
#> 3   -0.260215206 0.31529975 0.58437079 0.049774764   1  enter     27
#> 60  -0.107634508 0.56646479 0.57395511 0.048203051   7  enter     27
#> 88  -0.477798556 0.73179445 0.61521229 0.029676887  10  enter     27
#> 155  1.057549716 0.23434229 0.44656472 0.067479455  16   exit     27
#> 185  1.500000000 0.37900468 0.45838146 0.013171037  19   exit     27
#> 192  0.556633209 0.29750579 0.49012306 0.070524061  20 static     27
#> 4    0.196874557 0.31529975 0.58437079 0.049774764   1  enter     28
#> 61   0.235685297 0.56646479 0.57395511 0.048203051   7  enter     28
#> 89  -0.322388445 0.73179445 0.61521229 0.029676887  10  enter     28
#> 156  1.423052125 0.23434229 0.44656472 0.067479455  16   exit     28
#> 193  0.556633209 0.29750579 0.49012306 0.070524061  20    raw     28
#> 5    0.406686251 0.31529975 0.58437079 0.049774764   1  enter     29
#> 62   0.284730984 0.56646479 0.57395511 0.048203051   7 static     29
#> 67  -0.500000000 0.46864643 0.66211890 0.049663310   8  enter     29
#> 90   0.016183583 0.73179445 0.61521229 0.029676887  10  enter     29
#> 157  1.500000000 0.23434229 0.44656472 0.067479455  16   exit     29
#> 194  0.696391252 0.29750579 0.49012306 0.070524061  20   exit     29
#> 6    0.436659351 0.31529975 0.58437079 0.049774764   1 static     30
#> 63   0.284730984 0.56646479 0.57395511 0.048203051   7 static     30
#> 68  -0.418216842 0.46864643 0.66211890 0.049663310   8  enter     30
#> 91   0.171593694 0.73179445 0.61521229 0.029676887  10  enter     30
#> 195  1.360241957 0.29750579 0.49012306 0.070524061  20   exit     30
#> 7    0.436659351 0.31529975 0.58437079 0.049774764   1 static     31
#> 64   0.284730984 0.56646479 0.57395511 0.048203051   7    raw     31
#> 69   0.154265266 0.46864643 0.66211890 0.049663310   8  enter     31
#> 92   0.193795139 0.73179445 0.61521229 0.029676887  10 static     31
#> 97  -0.500000000 0.13926519 0.72642365 0.098835685  11  enter     31
#> 196  1.500000000 0.29750579 0.49012306 0.070524061  20   exit     31
#> 8    0.436659351 0.31529975 0.58437079 0.049774764   1    raw     32
#> 65   0.892365492 0.56646479 0.57395511 0.048203051   7   exit     32
#> 70   0.726747374 0.46864643 0.66211890 0.049663310   8  enter     32
#> 93   0.193795139 0.73179445 0.61521229 0.029676887  10    raw     32
#> 98  -0.456113963 0.13926519 0.72642365 0.098835685  11  enter     32
#> 9    0.968329675 0.31529975 0.58437079 0.049774764   1   exit     33
#> 66   1.500000000 0.56646479 0.57395511 0.048203051   7   exit     33
#> 71   0.808530532 0.46864643 0.66211890 0.049663310   8 static     33
#> 94   0.387306970 0.73179445 0.61521229 0.029676887  10   exit     33
#> 99  -0.148911701 0.13926519 0.72642365 0.098835685  11  enter     33
#> 10   1.500000000 0.31529975 0.58437079 0.049774764   1   exit     34
#> 72   0.808530532 0.46864643 0.66211890 0.049663310   8 static     34
#> 95   1.306488169 0.73179445 0.61521229 0.029676887  10   exit     34
#> 100  0.520350368 0.13926519 0.72642365 0.098835685  11  enter     34
#> 73   0.808530532 0.46864643 0.66211890 0.049663310   8    raw     35
#> 96   1.500000000 0.73179445 0.61521229 0.029676887  10   exit     35
#> 101  0.827552629 0.13926519 0.72642365 0.098835685  11  enter     35
#> 127 -0.500000000 0.59345445 0.81551969 0.079031139  14  enter     35
#> 74   0.910970453 0.46864643 0.66211890 0.049663310   8   exit     36
#> 102  0.871438666 0.13926519 0.72642365 0.098835685  11 static     36
#> 128 -0.471751602 0.59345445 0.81551969 0.079031139  14  enter     36
#> 75   1.397560079 0.46864643 0.66211890 0.049663310   8   exit     37
#> 103  0.871438666 0.13926519 0.72642365 0.098835685  11 static     37
#> 129 -0.274012813 0.59345445 0.81551969 0.079031139  14  enter     37
#> 76   1.500000000 0.46864643 0.66211890 0.049663310   8   exit     38
#> 104  0.871438666 0.13926519 0.72642365 0.098835685  11 static     38
#> 130  0.156775263 0.59345445 0.81551969 0.079031139  14  enter     38
#> 105  0.871438666 0.13926519 0.72642365 0.098835685  11 static     39
#> 131  0.354514052 0.59345445 0.81551969 0.079031139  14  enter     39
#> 106  0.871438666 0.13926519 0.72642365 0.098835685  11    raw     40
#> 132  0.382762450 0.59345445 0.81551969 0.079031139  14 static     40
#> 107  0.964558864 0.13926519 0.72642365 0.098835685  11   exit     41
#> 133  0.382762450 0.59345445 0.81551969 0.079031139  14 static     41
#> 139 -0.500000000 0.54304227 0.94840329 0.001302604  15  enter     41
#> 108  1.406879802 0.13926519 0.72642365 0.098835685  11   exit     42
#> 110 -0.500000000 0.08721259 0.97308099 0.017434827  12  enter     42
#> 134  0.382762450 0.59345445 0.81551969 0.079031139  14 static     42
#> 140 -0.464650798 0.54304227 0.94840329 0.001302604  15  enter     42
#> 109  1.500000000 0.13926519 0.72642365 0.098835685  11   exit     43
#> 111 -0.479809317 0.08721259 0.97308099 0.017434827  12  enter     43
#> 135  0.382762450 0.59345445 0.81551969 0.079031139  14    raw     43
#> 141 -0.217206387 0.54304227 0.94840329 0.001302604  15  enter     43
#> 112 -0.338474533 0.08721259 0.97308099 0.017434827  12  enter     44
#> 136  0.548279124 0.59345445 0.81551969 0.079031139  14   exit     44
#> 142  0.321868937 0.54304227 0.94840329 0.001302604  15  enter     44
#> 113 -0.030566613 0.08721259 0.97308099 0.017434827  12  enter     45
#> 137  1.334483326 0.59345445 0.81551969 0.079031139  14   exit     45
#> 143  0.569313348 0.54304227 0.94840329 0.001302604  15  enter     45
#> 114  0.110768170 0.08721259 0.97308099 0.017434827  12  enter     46
#> 138  1.500000000 0.59345445 0.81551969 0.079031139  14   exit     46
#> 144  0.604662550 0.54304227 0.94840329 0.001302604  15    raw     46
#> 115  0.130958854 0.08721259 0.97308099 0.017434827  12 static     47
#> 145  1.052331275 0.54304227 0.94840329 0.001302604  15   exit     47
#> 116  0.130958854 0.08721259 0.97308099 0.017434827  12    raw     48
#> 146  1.500000000 0.54304227 0.94840329 0.001302604  15   exit     48
#> 117  0.815479427 0.08721259 0.97308099 0.017434827  12   exit     49
#> 118  1.500000000 0.08721259 0.97308099 0.017434827  12   exit     50