Transition in and out of events

This tweening function is a more powerful version of tween_appear(), with support for newer features such as enter/exits and tween phase identification. The tweener treats each row in the data as unique events in time, and creates frames with the correct events present at any given time.

tween_events(
  .data,
  ease,
  nframes,
  start,
  end = NULL,
  range = NULL,
  enter = NULL,
  exit = NULL,
  enter_length = 0,
  exit_length = 0
)

Arguments

.data

A data.frame with components at different stages

ease

The easing function to use. Either a single string or one for each column in the data set.

nframes

The number of frames to calculate for the tween

start, end

The start (and potential end) of the event encoded in the row, as unquoted expressions. Will be evaluated in the context of .data so can refer to columns in it. If end = NULL the event will be without extend and only visible in a single frame, unless enter and/or exit is given.

range

The range of time points to include in the tween. If NULL it will use the range of time

enter, exit

functions that calculate a start state for new observations that appear in to or an end state for observations that are not present in to. If NULL the new/old observations will not be part of the tween. The function gets a data.frame with either the start state of the exiting observations, or the end state of the entering observations and must return a modified version of that data.frame. See the Match, Enter, and Exit section for more information.

enter_length, exit_length

The lenght of the opening and closing transitions if enter and/or exit is given. Measured in the same units as time

Value

A data.frame with the same columns as .data along with .id giving the component id, .phase giving the state of each component in each frame, and .frame giving the frame membership of each row.

See also

Examples

d <- data.frame(
  x = runif(20),
  y = runif(20),
  time = runif(20),
  duration = runif(20, max = 0.1)
)
from_left <- function(x) {
  x$x <- -0.5
  x
}
to_right <- function(x) {
  x$x <- 1.5
  x
}

tween_events(d, 'cubic-in-out', 50, start = time, end = time + duration,
             enter = from_left, exit = to_right, enter_length = 0.1,
             exit_length = 0.05)
#>                x           y      time    duration .id .phase .frame
#> 105 -0.500000000 0.845180908 0.1605402 0.054145559  11  enter      1
#> 189 -0.500000000 0.760990321 0.1559188 0.037326735  19  enter      1
#> 106 -0.480006607 0.845180908 0.1605402 0.054145559  11  enter      2
#> 190 -0.475786423 0.760990321 0.1559188 0.037326735  19  enter      2
#> 199 -0.500000000 0.961489340 0.1866720 0.053637700  20  enter      2
#> 107 -0.340052853 0.845180908 0.1605402 0.054145559  11  enter      3
#> 191 -0.306291388 0.760990321 0.1559188 0.037326735  19  enter      3
#> 200 -0.481629066 0.961489340 0.1866720 0.053637700  20  enter      3
#> 75  -0.500000000 0.832993890 0.2274847 0.061351959   8  enter      4
#> 108 -0.035153605 0.845180908 0.1605402 0.054145559  11  enter      4
#> 179 -0.500000000 0.168574719 0.2213891 0.041427333  18  enter      4
#> 192  0.062965655 0.760990321 0.1559188 0.037326735  19  enter      4
#> 201 -0.353032527 0.961489340 0.1866720 0.053637700  20  enter      4
#> 76  -0.460035551 0.832993890 0.2274847 0.061351959   8  enter      5
#> 109  0.104800148 0.845180908 0.1605402 0.054145559  11  enter      5
#> 180 -0.456103471 0.168574719 0.2213891 0.041427333  18  enter      5
#> 193  0.232460690 0.760990321 0.1559188 0.037326735  19  enter      5
#> 202 -0.072875783 0.961489340 0.1866720 0.053637700  20  enter      5
#> 77  -0.180284412 0.832993890 0.2274847 0.061351959   8  enter      6
#> 110  0.124793541 0.845180908 0.1605402 0.054145559  11 static      6
#> 181 -0.148827771 0.168574719 0.2213891 0.041427333  18  enter      6
#> 194  0.256674267 0.760990321 0.1559188 0.037326735  19 static      6
#> 203  0.055720756 0.961489340 0.1866720 0.053637700  20  enter      6
#> 44  -0.500000000 0.927642091 0.2805449 0.047440976   5  enter      7
#> 78   0.429173428 0.832993890 0.2274847 0.061351959   8  enter      7
#> 111  0.124793541 0.845180908 0.1605402 0.054145559  11 static      7
#> 182  0.520594290 0.168574719 0.2213891 0.041427333  18  enter      7
#> 195  0.256674267 0.760990321 0.1559188 0.037326735  19 static      7
#> 204  0.074091690 0.961489340 0.1866720 0.053637700  20 static      7
#> 45  -0.462687030 0.927642091 0.2805449 0.047440976   5  enter      8
#> 79   0.708924567 0.832993890 0.2274847 0.061351959   8  enter      8
#> 112  0.124793541 0.845180908 0.1605402 0.054145559  11 static      8
#> 183  0.827869990 0.168574719 0.2213891 0.041427333  18  enter      8
#> 196  0.256674267 0.760990321 0.1559188 0.037326735  19    raw      8
#> 205  0.074091690 0.961489340 0.1866720 0.053637700  20 static      8
#> 46  -0.201496243 0.927642091 0.2805449 0.047440976   5  enter      9
#> 80   0.748889016 0.832993890 0.2274847 0.061351959   8 static      9
#> 113  0.124793541 0.845180908 0.1605402 0.054145559  11    raw      9
#> 184  0.871766519 0.168574719 0.2213891 0.041427333  18 static      9
#> 197  0.878337133 0.760990321 0.1559188 0.037326735  19   exit      9
#> 206  0.074091690 0.961489340 0.1866720 0.053637700  20 static      9
#> 47   0.367526544 0.927642091 0.2805449 0.047440976   5  enter     10
#> 81   0.748889016 0.832993890 0.2274847 0.061351959   8 static     10
#> 114  0.812396771 0.845180908 0.1605402 0.054145559  11   exit     10
#> 185  0.871766519 0.168574719 0.2213891 0.041427333  18 static     10
#> 198  1.500000000 0.760990321 0.1559188 0.037326735  19   exit     10
#> 207  0.074091690 0.961489340 0.1866720 0.053637700  20    raw     10
#> 31  -0.500000000 0.349759821 0.3626065 0.099389574   4  enter     11
#> 48   0.628717331 0.927642091 0.2805449 0.047440976   5  enter     11
#> 82   0.748889016 0.832993890 0.2274847 0.061351959   8 static     11
#> 115  1.500000000 0.845180908 0.1605402 0.054145559  11   exit     11
#> 186  0.871766519 0.168574719 0.2213891 0.041427333  18    raw     11
#> 208  0.787045845 0.961489340 0.1866720 0.053637700  20   exit     11
#> 32  -0.460887562 0.349759821 0.3626065 0.099389574   4  enter     12
#> 49   0.666030301 0.927642091 0.2805449 0.047440976   5 static     12
#> 83   0.748889016 0.832993890 0.2274847 0.061351959   8    raw     12
#> 187  1.185883260 0.168574719 0.2213891 0.041427333  18   exit     12
#> 209  1.500000000 0.961489340 0.1866720 0.053637700  20   exit     12
#> 33  -0.187100498 0.349759821 0.3626065 0.099389574   4  enter     13
#> 50   0.666030301 0.927642091 0.2805449 0.047440976   5 static     13
#> 84   1.124444508 0.832993890 0.2274847 0.061351959   8   exit     13
#> 188  1.500000000 0.168574719 0.2213891 0.041427333  18   exit     13
#> 34   0.409364176 0.349759821 0.3626065 0.099389574   4  enter     14
#> 51   0.666030301 0.927642091 0.2805449 0.047440976   5    raw     14
#> 85   1.500000000 0.832993890 0.2274847 0.061351959   8   exit     14
#> 96  -0.500000000 0.962392115 0.4364163 0.013256631  10  enter     14
#> 35   0.683151240 0.349759821 0.3626065 0.099389574   4  enter     15
#> 52   1.083015151 0.927642091 0.2805449 0.047440976   5   exit     15
#> 97  -0.459611638 0.962392115 0.4364163 0.013256631  10  enter     15
#> 36   0.722263678 0.349759821 0.3626065 0.099389574   4 static     16
#> 53   1.500000000 0.927642091 0.2805449 0.047440976   5   exit     16
#> 98  -0.176893104 0.962392115 0.4364163 0.013256631  10  enter     16
#> 135 -0.500000000 0.970215719 0.4708386 0.077525796  14  enter     16
#> 37   0.722263678 0.349759821 0.3626065 0.099389574   4 static     17
#> 99   0.439029417 0.962392115 0.4364163 0.013256631  10  enter     17
#> 136 -0.459451596 0.970215719 0.4708386 0.077525796  14  enter     17
#> 38   0.722263678 0.349759821 0.3626065 0.099389574   4 static     18
#> 100  0.721747951 0.962392115 0.4364163 0.013256631  10  enter     18
#> 137 -0.175612765 0.970215719 0.4708386 0.077525796  14  enter     18
#> 39   0.722263678 0.349759821 0.3626065 0.099389574   4 static     19
#> 101  0.762136313 0.962392115 0.4364163 0.013256631  10 static     19
#> 116 -0.500000000 0.993130847 0.5275659 0.033376212  12  enter     19
#> 138  0.442750400 0.970215719 0.4708386 0.077525796  14  enter     19
#> 40   0.722263678 0.349759821 0.3626065 0.099389574   4    raw     20
#> 102  0.762136313 0.962392115 0.4364163 0.013256631  10    raw     20
#> 117 -0.414908053 0.993130847 0.5275659 0.033376212  12  enter     20
#> 139  0.726589231 0.970215719 0.4708386 0.077525796  14  enter     20
#> 41   0.837483874 0.349759821 0.3626065 0.099389574   4   exit     21
#> 103  1.131068156 0.962392115 0.4364163 0.013256631  10   exit     21
#> 118  0.180735574 0.993130847 0.5275659 0.033376212  12  enter     21
#> 140  0.767137635 0.970215719 0.4708386 0.077525796  14 static     21
#> 13  -0.500000000 0.279508535 0.5994430 0.026269116   2  enter     22
#> 42   1.384779804 0.349759821 0.3626065 0.099389574   4   exit     22
#> 104  1.500000000 0.962392115 0.4364163 0.013256631  10   exit     22
#> 119  0.776379202 0.993130847 0.5275659 0.033376212  12  enter     22
#> 141  0.767137635 0.970215719 0.4708386 0.077525796  14 static     22
#> 147 -0.500000000 0.484465477 0.5899820 0.075115477  15  enter     22
#> 14  -0.458632279 0.279508535 0.5994430 0.026269116   2  enter     23
#> 43   1.500000000 0.349759821 0.3626065 0.099389574   4   exit     23
#> 54  -0.500000000 0.782430034 0.6190287 0.028157971   6  enter     23
#> 120  0.861471149 0.993130847 0.5275659 0.033376212  12 static     23
#> 142  0.767137635 0.970215719 0.4708386 0.077525796  14 static     23
#> 148 -0.435917636 0.484465477 0.5899820 0.075115477  15  enter     23
#> 15  -0.169058230 0.279508535 0.5994430 0.026269116   2  enter     24
#> 55  -0.478277557 0.782430034 0.6190287 0.028157971   6  enter     24
#> 121  0.861471149 0.993130847 0.5275659 0.033376212  12 static     24
#> 143  0.767137635 0.970215719 0.4708386 0.077525796  14    raw     24
#> 149  0.012658911 0.484465477 0.5899820 0.075115477  15  enter     24
#> 16   0.461799519 0.279508535 0.5994430 0.026269116   2  enter     25
#> 56  -0.326220459 0.782430034 0.6190287 0.028157971   6  enter     25
#> 122  0.861471149 0.993130847 0.5275659 0.033376212  12    raw     25
#> 144  0.875709837 0.970215719 0.4708386 0.077525796  14   exit     25
#> 150  0.461235458 0.484465477 0.5899820 0.075115477  15  enter     25
#> 17   0.751373568 0.279508535 0.5994430 0.026269116   2  enter     26
#> 57   0.005046792 0.782430034 0.6190287 0.028157971   6  enter     26
#> 123  1.180735574 0.993130847 0.5275659 0.033376212  12   exit     26
#> 145  1.391427798 0.970215719 0.4708386 0.077525796  14   exit     26
#> 151  0.525317822 0.484465477 0.5899820 0.075115477  15 static     26
#> 18   0.792741289 0.279508535 0.5994430 0.026269116   2 static     27
#> 58   0.157103891 0.782430034 0.6190287 0.028157971   6  enter     27
#> 124  1.500000000 0.993130847 0.5275659 0.033376212  12   exit     27
#> 125 -0.500000000 0.323079476 0.7126230 0.031832359  13  enter     27
#> 146  1.500000000 0.970215719 0.4708386 0.077525796  14   exit     27
#> 152  0.525317822 0.484465477 0.5899820 0.075115477  15 static     27
#> 19   0.792741289 0.279508535 0.5994430 0.026269116   2    raw     28
#> 59   0.178826334 0.782430034 0.6190287 0.028157971   6 static     28
#> 126 -0.483268781 0.323079476 0.7126230 0.031832359  13  enter     28
#> 153  0.525317822 0.484465477 0.5899820 0.075115477  15 static     28
#> 171 -0.500000000 0.005595589 0.7132856 0.005832804  17  enter     28
#> 20   1.146370645 0.279508535 0.5994430 0.026269116   2   exit     29
#> 60   0.178826334 0.782430034 0.6190287 0.028157971   6    raw     29
#> 127 -0.366150250 0.323079476 0.7126230 0.031832359  13  enter     29
#> 154  0.525317822 0.484465477 0.5899820 0.075115477  15 static     29
#> 172 -0.445667304 0.005595589 0.7132856 0.005832804  17  enter     29
#> 21   1.500000000 0.279508535 0.5994430 0.026269116   2   exit     30
#> 61   0.374555766 0.782430034 0.6190287 0.028157971   6   exit     30
#> 128 -0.110999165 0.323079476 0.7126230 0.031832359  13  enter     30
#> 155  0.525317822 0.484465477 0.5899820 0.075115477  15    raw     30
#> 173 -0.065338432 0.005595589 0.7132856 0.005832804  17  enter     30
#> 62   1.304270568 0.782430034 0.6190287 0.028157971   6   exit     31
#> 129  0.006119365 0.323079476 0.7126230 0.031832359  13  enter     31
#> 156  1.012658911 0.484465477 0.5899820 0.075115477  15   exit     31
#> 158 -0.500000000 0.495376588 0.7937081 0.097441273  16  enter     31
#> 174  0.314990440 0.005595589 0.7132856 0.005832804  17  enter     31
#> 63   1.500000000 0.782430034 0.6190287 0.028157971   6   exit     32
#> 130  0.022850584 0.323079476 0.7126230 0.031832359  13 static     32
#> 157  1.500000000 0.484465477 0.5899820 0.075115477  15   exit     32
#> 159 -0.458922715 0.495376588 0.7937081 0.097441273  16  enter     32
#> 175  0.369323136 0.005595589 0.7132856 0.005832804  17 static     32
#> 131  0.022850584 0.323079476 0.7126230 0.031832359  13 static     33
#> 160 -0.171381724 0.495376588 0.7937081 0.097441273  16  enter     33
#> 176  0.369323136 0.005595589 0.7132856 0.005832804  17    raw     33
#> 64  -0.500000000 0.004487753 0.8458243 0.066282865   7  enter     34
#> 132  0.022850584 0.323079476 0.7126230 0.031832359  13    raw     34
#> 161  0.455046865 0.495376588 0.7937081 0.097441273  16  enter     34
#> 177  0.934661568 0.005595589 0.7132856 0.005832804  17   exit     34
#> 65  -0.470938096 0.004487753 0.8458243 0.066282865   7  enter     35
#> 133  0.761425292 0.323079476 0.7126230 0.031832359  13   exit     35
#> 162  0.742587856 0.495376588 0.7937081 0.097441273  16  enter     35
#> 178  1.500000000 0.005595589 0.7132856 0.005832804  17   exit     35
#> 66  -0.267504770 0.004487753 0.8458243 0.066282865   7  enter     36
#> 134  1.500000000 0.323079476 0.7126230 0.031832359  13   exit     36
#> 163  0.783665141 0.495376588 0.7937081 0.097441273  16 static     36
#> 22  -0.500000000 0.732620544 0.9126814 0.025139544   3  enter     37
#> 67   0.175689264 0.004487753 0.8458243 0.066282865   7  enter     37
#> 86  -0.500000000 0.097194439 0.9172177 0.041003593   9  enter     37
#> 164  0.783665141 0.495376588 0.7937081 0.097441273  16 static     37
#> 23  -0.457387782 0.732620544 0.9126814 0.025139544   3  enter     38
#> 68   0.379122590 0.004487753 0.8458243 0.066282865   7  enter     38
#> 87  -0.456607856 0.097194439 0.9172177 0.041003593   9  enter     38
#> 165  0.783665141 0.495376588 0.7937081 0.097441273  16 static     38
#> 1   -0.500000000 0.133680098 0.9519911 0.084033233   1  enter     39
#> 24  -0.159102258 0.732620544 0.9126814 0.025139544   3  enter     39
#> 69   0.408184494 0.004487753 0.8458243 0.066282865   7 static     39
#> 88  -0.152862846 0.097194439 0.9172177 0.041003593   9  enter     39
#> 166  0.783665141 0.495376588 0.7937081 0.097441273  16 static     39
#> 2   -0.456974212 0.133680098 0.9519911 0.084033233   1  enter     40
#> 25   0.490734064 0.732620544 0.9126814 0.025139544   3  enter     40
#> 70   0.408184494 0.004487753 0.8458243 0.066282865   7 static     40
#> 89   0.508867352 0.097194439 0.9172177 0.041003593   9  enter     40
#> 167  0.783665141 0.495376588 0.7937081 0.097441273  16 static     40
#> 3   -0.155793694 0.133680098 0.9519911 0.084033233   1  enter     41
#> 26   0.789019588 0.732620544 0.9126814 0.025139544   3  enter     41
#> 71   0.408184494 0.004487753 0.8458243 0.066282865   7 static     41
#> 90   0.812612362 0.097194439 0.9172177 0.041003593   9  enter     41
#> 168  0.783665141 0.495376588 0.7937081 0.097441273  16    raw     41
#> 4    0.500349575 0.133680098 0.9519911 0.084033233   1  enter     42
#> 27   0.831631806 0.732620544 0.9126814 0.025139544   3 static     42
#> 72   0.408184494 0.004487753 0.8458243 0.066282865   7    raw     42
#> 91   0.856004506 0.097194439 0.9172177 0.041003593   9 static     42
#> 169  1.141832570 0.495376588 0.7937081 0.097441273  16   exit     42
#> 5    0.801530093 0.133680098 0.9519911 0.084033233   1  enter     43
#> 28   0.831631806 0.732620544 0.9126814 0.025139544   3    raw     43
#> 73   0.954092247 0.004487753 0.8458243 0.066282865   7   exit     43
#> 92   0.856004506 0.097194439 0.9172177 0.041003593   9 static     43
#> 170  1.500000000 0.495376588 0.7937081 0.097441273  16   exit     43
#> 6    0.844555881 0.133680098 0.9519911 0.084033233   1 static     44
#> 29   1.165815903 0.732620544 0.9126814 0.025139544   3   exit     44
#> 74   1.500000000 0.004487753 0.8458243 0.066282865   7   exit     44
#> 93   0.856004506 0.097194439 0.9172177 0.041003593   9    raw     44
#> 7    0.844555881 0.133680098 0.9519911 0.084033233   1 static     45
#> 30   1.500000000 0.732620544 0.9126814 0.025139544   3   exit     45
#> 94   1.178002253 0.097194439 0.9172177 0.041003593   9   exit     45
#> 8    0.844555881 0.133680098 0.9519911 0.084033233   1 static     46
#> 95   1.500000000 0.097194439 0.9172177 0.041003593   9   exit     46
#> 9    0.844555881 0.133680098 0.9519911 0.084033233   1 static     47
#> 10   0.844555881 0.133680098 0.9519911 0.084033233   1    raw     48
#> 11   1.172277940 0.133680098 0.9519911 0.084033233   1   exit     49
#> 12   1.500000000 0.133680098 0.9519911 0.084033233   1   exit     50