Transition in and out of events

This tweening function is a more powerful version of tween_appear(), with support for newer features such as enter/exits and tween phase identification. The tweener treats each row in the data as unique events in time, and creates frames with the correct events present at any given time.

tween_events(
  .data,
  ease,
  nframes,
  start,
  end = NULL,
  range = NULL,
  enter = NULL,
  exit = NULL,
  enter_length = 0,
  exit_length = 0
)

Arguments

.data

A data.frame with components at different stages

ease

The easing function to use. Either a single string or one for each column in the data set.

nframes

The number of frames to calculate for the tween

start, end

The start (and potential end) of the event encoded in the row, as unquoted expressions. Will be evaluated in the context of .data so can refer to columns in it. If end = NULL the event will be without extend and only visible in a single frame, unless enter and/or exit is given.

range

The range of time points to include in the tween. If NULL it will use the range of time

enter, exit

functions that calculate a start state for new observations that appear in to or an end state for observations that are not present in to. If NULL the new/old observations will not be part of the tween. The function gets a data.frame with either the start state of the exiting observations, or the end state of the entering observations and must return a modified version of that data.frame. See the Match, Enter, and Exit section for more information.

enter_length, exit_length

The lenght of the opening and closing transitions if enter and/or exit is given. Measured in the same units as time

Value

A data.frame with the same columns as .data along with .id giving the component id, .phase giving the state of each component in each frame, and .frame giving the frame membership of each row.

See also

Examples

d <- data.frame(
  x = runif(20),
  y = runif(20),
  time = runif(20),
  duration = runif(20, max = 0.1)
)
from_left <- function(x) {
  x$x <- -0.5
  x
}
to_right <- function(x) {
  x$x <- 1.5
  x
}

tween_events(d, 'cubic-in-out', 50, start = time, end = time + duration,
             enter = from_left, exit = to_right, enter_length = 0.1,
             exit_length = 0.05)
#>               x          y       time   duration .id .phase .frame
#> 24  -0.50000000 0.12311748 0.03155371 0.05359034   3  enter      1
#> 25  -0.46744868 0.12311748 0.03155371 0.05359034   3  enter      2
#> 26  -0.23958942 0.12311748 0.03155371 0.05359034   3  enter      3
#> 27  -0.01173016 0.12311748 0.03155371 0.05359034   3  enter      4
#> 53  -0.50000000 0.04512714 0.09503306 0.06473133   6  enter      4
#> 28   0.02082116 0.12311748 0.03155371 0.05359034   3 static      5
#> 54  -0.41158275 0.04512714 0.09503306 0.06473133   6  enter      5
#> 29   0.02082116 0.12311748 0.03155371 0.05359034   3 static      6
#> 55   0.20733802 0.04512714 0.09503306 0.06473133   6  enter      6
#> 30   0.02082116 0.12311748 0.03155371 0.05359034   3 static      7
#> 56   0.82625879 0.04512714 0.09503306 0.06473133   6  enter      7
#> 31   0.02082116 0.12311748 0.03155371 0.05359034   3    raw      8
#> 57   0.91467604 0.04512714 0.09503306 0.06473133   6 static      8
#> 152 -0.50000000 0.70714464 0.19657555 0.03377961  16  enter      8
#> 32   0.76041058 0.12311748 0.03155371 0.05359034   3   exit      9
#> 58   0.91467604 0.04512714 0.09503306 0.06473133   6 static      9
#> 153 -0.45917358 0.70714464 0.19657555 0.03377961  16  enter      9
#> 33   1.50000000 0.12311748 0.03155371 0.05359034   3   exit     10
#> 59   0.91467604 0.04512714 0.09503306 0.06473133   6 static     10
#> 154 -0.17338861 0.70714464 0.19657555 0.03377961  16  enter     10
#> 60   0.91467604 0.04512714 0.09503306 0.06473133   6    raw     11
#> 63  -0.50000000 0.39646343 0.26380957 0.01322144   7  enter     11
#> 155  0.44921436 0.70714464 0.19657555 0.03377961  16  enter     11
#> 61   1.20733802 0.04512714 0.09503306 0.06473133   6   exit     12
#> 64  -0.46376502 0.39646343 0.26380957 0.01322144   7  enter     12
#> 156  0.73499933 0.70714464 0.19657555 0.03377961  16  enter     12
#> 62   1.50000000 0.04512714 0.09503306 0.06473133   6   exit     13
#> 65  -0.21012013 0.39646343 0.26380957 0.01322144   7  enter     13
#> 157  0.77582575 0.70714464 0.19657555 0.03377961  16 static     13
#> 66   0.34246337 0.39646343 0.26380957 0.01322144   7  enter     14
#> 158  0.77582575 0.70714464 0.19657555 0.03377961  16    raw     14
#> 67   0.59610826 0.39646343 0.26380957 0.01322144   7  enter     15
#> 159  0.88311083 0.70714464 0.19657555 0.03377961  16   exit     15
#> 68   0.63234324 0.39646343 0.26380957 0.01322144   7    raw     16
#> 160  1.39271493 0.70714464 0.19657555 0.03377961  16   exit     16
#> 1   -0.50000000 0.41696457 0.40038338 0.08455362   1  enter     17
#> 69   0.76088498 0.39646343 0.26380957 0.01322144   7   exit     17
#> 161  1.50000000 0.70714464 0.19657555 0.03377961  16   exit     17
#> 174 -0.50000000 0.32125271 0.39179325 0.02957061  18  enter     17
#> 2   -0.47909982 0.41696457 0.40038338 0.08455362   1  enter     18
#> 70   1.37145826 0.39646343 0.26380957 0.01322144   7   exit     18
#> 175 -0.47001323 0.32125271 0.39179325 0.02957061  18  enter     18
#> 3   -0.33279855 0.41696457 0.40038338 0.08455362   1  enter     19
#> 13  -0.50000000 0.62592877 0.44471795 0.07302651   2  enter     19
#> 71   1.50000000 0.39646343 0.26380957 0.01322144   7   exit     19
#> 176 -0.26010583 0.32125271 0.39179325 0.02957061  18  enter     19
#> 4   -0.01407080 0.41696457 0.40038338 0.08455362   1  enter     20
#> 14  -0.47039148 0.62592877 0.44471795 0.07302651   2  enter     20
#> 177  0.19719242 0.32125271 0.39179325 0.02957061  18  enter     20
#> 5    0.13223047 0.41696457 0.40038338 0.08455362   1  enter     21
#> 15  -0.26313185 0.62592877 0.44471795 0.07302651   2  enter     21
#> 178  0.40709982 0.32125271 0.39179325 0.02957061  18  enter     21
#> 6    0.15313065 0.41696457 0.40038338 0.08455362   1 static     22
#> 16   0.18839806 0.62592877 0.44471795 0.07302651   2  enter     22
#> 119 -0.50000000 0.12211946 0.50980157 0.09758357  13  enter     22
#> 179  0.43708659 0.32125271 0.39179325 0.02957061  18 static     22
#> 195 -0.50000000 0.50941747 0.49659111 0.05843160  20  enter     22
#> 7    0.15313065 0.41696457 0.40038338 0.08455362   1 static     23
#> 17   0.39565768 0.62592877 0.44471795 0.07302651   2  enter     23
#> 120 -0.46027419 0.12211946 0.50980157 0.09758357  13  enter     23
#> 180  0.43708659 0.32125271 0.39179325 0.02957061  18    raw     23
#> 196 -0.42621131 0.50941747 0.49659111 0.05843160  20  enter     23
#> 8    0.15313065 0.41696457 0.40038338 0.08455362   1 static     24
#> 18   0.42526620 0.62592877 0.44471795 0.07302651   2 static     24
#> 121 -0.18219356 0.12211946 0.50980157 0.09758357  13  enter     24
#> 181  0.96854329 0.32125271 0.39179325 0.02957061  18   exit     24
#> 197  0.09030956 0.50941747 0.49659111 0.05843160  20  enter     24
#> 9    0.15313065 0.41696457 0.40038338 0.08455362   1 static     25
#> 19   0.42526620 0.62592877 0.44471795 0.07302651   2 static     25
#> 122  0.42362498 0.12211946 0.50980157 0.09758357  13  enter     25
#> 182  1.50000000 0.32125271 0.39179325 0.02957061  18   exit     25
#> 198  0.60683042 0.50941747 0.49659111 0.05843160  20  enter     25
#> 10   0.15313065 0.41696457 0.40038338 0.08455362   1    raw     26
#> 20   0.42526620 0.62592877 0.44471795 0.07302651   2 static     26
#> 123  0.70170562 0.12211946 0.50980157 0.09758357  13  enter     26
#> 199  0.68061911 0.50941747 0.49659111 0.05843160  20 static     26
#> 11   0.82656532 0.41696457 0.40038338 0.08455362   1   exit     27
#> 21   0.42526620 0.62592877 0.44471795 0.07302651   2    raw     27
#> 124  0.74143142 0.12211946 0.50980157 0.09758357  13 static     27
#> 200  0.68061911 0.50941747 0.49659111 0.05843160  20 static     27
#> 12   1.50000000 0.41696457 0.40038338 0.08455362   1   exit     28
#> 22   0.96263310 0.62592877 0.44471795 0.07302651   2   exit     28
#> 125  0.74143142 0.12211946 0.50980157 0.09758357  13 static     28
#> 201  0.68061911 0.50941747 0.49659111 0.05843160  20 static     28
#> 23   1.50000000 0.62592877 0.44471795 0.07302651   2   exit     29
#> 99  -0.50000000 0.11905015 0.65027131 0.08775780  11  enter     29
#> 126  0.74143142 0.12211946 0.50980157 0.09758357  13 static     29
#> 162 -0.50000000 0.87107782 0.65893379 0.09433127  17  enter     29
#> 202  0.68061911 0.50941747 0.49659111 0.05843160  20    raw     29
#> 100 -0.43617911 0.11905015 0.65027131 0.08775780  11  enter     30
#> 127  0.74143142 0.12211946 0.50980157 0.09758357  13 static     30
#> 163 -0.47301777 0.87107782 0.65893379 0.09433127  17  enter     30
#> 203  1.09030956 0.50941747 0.49659111 0.05843160  20   exit     30
#> 101  0.01056713 0.11905015 0.65027131 0.08775780  11  enter     31
#> 128  0.74143142 0.12211946 0.50980157 0.09758357  13    raw     31
#> 164 -0.28414215 0.87107782 0.65893379 0.09433127  17  enter     31
#> 204  1.50000000 0.50941747 0.49659111 0.05843160  20   exit     31
#> 42  -0.50000000 0.60693041 0.73322772 0.07534158   5  enter     32
#> 80  -0.50000000 0.36896670 0.73311643 0.02627676   9  enter     32
#> 102  0.45731336 0.11905015 0.65027131 0.08775780  11  enter     32
#> 129  1.12071571 0.12211946 0.50980157 0.09758357  13   exit     32
#> 165  0.12733686 0.87107782 0.65893379 0.09433127  17  enter     32
#> 43  -0.46583581 0.60693041 0.73322772 0.07534158   5  enter     33
#> 81  -0.45230004 0.36896670 0.73311643 0.02627676   9  enter     33
#> 103  0.52113426 0.11905015 0.65027131 0.08775780  11 static     33
#> 110 -0.50000000 0.78499299 0.74451926 0.02297705  12  enter     33
#> 130  1.50000000 0.12211946 0.50980157 0.09758357  13   exit     33
#> 131 -0.50000000 0.60409728 0.75016146 0.07494755  14  enter     33
#> 166  0.31621248 0.87107782 0.65893379 0.09433127  17  enter     33
#> 44  -0.22668647 0.60693041 0.73322772 0.07534158   5  enter     34
#> 82  -0.11840035 0.36896670 0.73311643 0.02627676   9  enter     34
#> 104  0.52113426 0.11905015 0.65027131 0.08775780  11 static     34
#> 111 -0.45214241 0.78499299 0.74451926 0.02297705  12  enter     34
#> 132 -0.45288197 0.60409728 0.75016146 0.07494755  14  enter     34
#> 167  0.34319471 0.87107782 0.65893379 0.09433127  17 static     34
#> 45   0.29431746 0.60693041 0.73322772 0.07534158   5  enter     35
#> 83   0.60902399 0.36896670 0.73311643 0.02627676   9  enter     35
#> 105  0.52113426 0.11905015 0.65027131 0.08775780  11 static     35
#> 112 -0.11713929 0.78499299 0.74451926 0.02297705  12  enter     35
#> 133 -0.12305573 0.60409728 0.75016146 0.07494755  14  enter     35
#> 168  0.34319471 0.87107782 0.65893379 0.09433127  17 static     35
#> 46   0.53346680 0.60693041 0.73322772 0.07534158   5  enter     36
#> 72  -0.50000000 0.19175646 0.80805495 0.01563595   8  enter     36
#> 84   0.94292369 0.36896670 0.73311643 0.02627676   9  enter     36
#> 106  0.52113426 0.11905015 0.65027131 0.08775780  11 static     36
#> 113  0.21786383 0.78499299 0.74451926 0.02297705  12  enter     36
#> 134  0.59549427 0.60409728 0.75016146 0.07494755  14  enter     36
#> 169  0.34319471 0.87107782 0.65893379 0.09433127  17 static     36
#> 183 -0.50000000 0.09886206 0.81721655 0.09709381  19  enter     36
#> 47   0.56763099 0.60693041 0.73322772 0.07534158   5 static     37
#> 73  -0.42308241 0.19175646 0.80805495 0.01563595   8  enter     37
#> 85   0.99062365 0.36896670 0.73311643 0.02627676   9 static     37
#> 89  -0.50000000 0.09311332 0.83822458 0.04323011  10  enter     37
#> 107  0.52113426 0.11905015 0.65027131 0.08775780  11    raw     37
#> 114  0.26572142 0.78499299 0.74451926 0.02297705  12 static     37
#> 135  0.92532051 0.60409728 0.75016146 0.07494755  14  enter     37
#> 170  0.34319471 0.87107782 0.65893379 0.09433127  17 static     37
#> 184 -0.46524357 0.09886206 0.81721655 0.09709381  19  enter     37
#> 48   0.56763099 0.60693041 0.73322772 0.07534158   5 static     38
#> 74   0.11534070 0.19175646 0.80805495 0.01563595   8  enter     38
#> 86   0.99062365 0.36896670 0.73311643 0.02627676   9    raw     38
#> 90  -0.47722692 0.09311332 0.83822458 0.04323011  10  enter     38
#> 108  1.01056713 0.11905015 0.65027131 0.08775780  11   exit     38
#> 115  0.26572142 0.78499299 0.74451926 0.02297705  12    raw     38
#> 136  0.97243854 0.60409728 0.75016146 0.07494755  14 static     38
#> 171  0.34319471 0.87107782 0.65893379 0.09433127  17    raw     38
#> 185 -0.22194857 0.09886206 0.81721655 0.09709381  19  enter     38
#> 49   0.56763099 0.60693041 0.73322772 0.07534158   5 static     39
#> 75   0.65376380 0.19175646 0.80805495 0.01563595   8  enter     39
#> 87   1.24531182 0.36896670 0.73311643 0.02627676   9   exit     39
#> 91  -0.31781536 0.09311332 0.83822458 0.04323011  10  enter     39
#> 109  1.50000000 0.11905015 0.65027131 0.08775780  11   exit     39
#> 116  0.44857751 0.78499299 0.74451926 0.02297705  12   exit     39
#> 137  0.97243854 0.60409728 0.75016146 0.07494755  14 static     39
#> 172  0.92159735 0.87107782 0.65893379 0.09433127  17   exit     39
#> 186  0.30808698 0.09886206 0.81721655 0.09709381  19  enter     39
#> 50   0.56763099 0.60693041 0.73322772 0.07534158   5    raw     40
#> 76   0.73068139 0.19175646 0.80805495 0.01563595   8 static     40
#> 88   1.50000000 0.36896670 0.73311643 0.02627676   9   exit     40
#> 92   0.02947410 0.09311332 0.83822458 0.04323011  10  enter     40
#> 117  1.31714391 0.78499299 0.74451926 0.02297705  12   exit     40
#> 138  0.97243854 0.60409728 0.75016146 0.07494755  14 static     40
#> 173  1.50000000 0.87107782 0.65893379 0.09433127  17   exit     40
#> 187  0.55138198 0.09886206 0.81721655 0.09709381  19  enter     40
#> 51   1.03381550 0.60693041 0.73322772 0.07534158   5   exit     41
#> 77   0.73068139 0.19175646 0.80805495 0.01563595   8    raw     41
#> 93   0.18888565 0.09311332 0.83822458 0.04323011  10  enter     41
#> 118  1.50000000 0.78499299 0.74451926 0.02297705  12   exit     41
#> 139  0.97243854 0.60409728 0.75016146 0.07494755  14    raw     41
#> 142 -0.50000000 0.81666312 0.93558743 0.04225078  15  enter     41
#> 188  0.58613841 0.09886206 0.81721655 0.09709381  19 static     41
#> 34  -0.50000000 0.47298435 0.93998265 0.01528374   4  enter     42
#> 52   1.50000000 0.60693041 0.73322772 0.07534158   5   exit     42
#> 78   1.11534070 0.19175646 0.80805495 0.01563595   8   exit     42
#> 94   0.21165873 0.09311332 0.83822458 0.04323011  10 static     42
#> 140  1.23621927 0.60409728 0.75016146 0.07494755  14   exit     42
#> 143 -0.47155526 0.81666312 0.93558743 0.04225078  15  enter     42
#> 189  0.58613841 0.09886206 0.81721655 0.09709381  19 static     42
#> 35  -0.45780097 0.47298435 0.93998265 0.01528374   4  enter     43
#> 79   1.50000000 0.19175646 0.80805495 0.01563595   8   exit     43
#> 95   0.21165873 0.09311332 0.83822458 0.04323011  10    raw     43
#> 141  1.50000000 0.60409728 0.75016146 0.07494755  14   exit     43
#> 144 -0.27244211 0.81666312 0.93558743 0.04225078  15  enter     43
#> 190  0.58613841 0.09886206 0.81721655 0.09709381  19 static     43
#> 36  -0.16240774 0.47298435 0.93998265 0.01528374   4  enter     44
#> 96   0.40252411 0.09311332 0.83822458 0.04323011  10   exit     44
#> 145  0.16134011 0.81666312 0.93558743 0.04225078  15  enter     44
#> 191  0.58613841 0.09886206 0.81721655 0.09709381  19 static     44
#> 37   0.13298549 0.47298435 0.93998265 0.01528374   4  enter     45
#> 97   1.30913463 0.09311332 0.83822458 0.04323011  10   exit     45
#> 146  0.36045326 0.81666312 0.93558743 0.04225078  15  enter     45
#> 192  0.58613841 0.09886206 0.81721655 0.09709381  19    raw     45
#> 38   0.17518452 0.47298435 0.93998265 0.01528374   4 static     46
#> 98   1.50000000 0.09311332 0.83822458 0.04323011  10   exit     46
#> 147  0.38889799 0.81666312 0.93558743 0.04225078  15 static     46
#> 193  1.04306920 0.09886206 0.81721655 0.09709381  19   exit     46
#> 39   0.17518452 0.47298435 0.93998265 0.01528374   4    raw     47
#> 148  0.38889799 0.81666312 0.93558743 0.04225078  15 static     47
#> 194  1.50000000 0.09886206 0.81721655 0.09709381  19   exit     47
#> 40   0.83759226 0.47298435 0.93998265 0.01528374   4   exit     48
#> 149  0.38889799 0.81666312 0.93558743 0.04225078  15    raw     48
#> 41   1.50000000 0.47298435 0.93998265 0.01528374   4   exit     49
#> 150  0.94444900 0.81666312 0.93558743 0.04225078  15   exit     49
#> 151  1.50000000 0.81666312 0.93558743 0.04225078  15   exit     50