Transition in and out of events

This tweening function is a more powerful version of tween_appear(), with support for newer features such as enter/exits and tween phase identification. The tweener treats each row in the data as unique events in time, and creates frames with the correct events present at any given time.

tween_events(
  .data,
  ease,
  nframes,
  start,
  end = NULL,
  range = NULL,
  enter = NULL,
  exit = NULL,
  enter_length = 0,
  exit_length = 0
)

Arguments

.data
A data.frame with components at different stages

ease
The easing function to use. Either a single string or one for each column in the data set.

nframes
The number of frames to calculate for the tween

start, end
The start (and potential end) of the event encoded in the row, as unquoted expressions. Will be evaluated in the context of .data so can refer to columns in it. If end = NULL the event will be without extend and only visible in a single frame, unless enter and/or exit is given.

range
The range of time points to include in the tween. If NULL it will use the range of time

enter
functions that calculate a start state for new observations that appear in to or an end state for observations that are not present in to. If NULL the new/old observations will not be part of the tween. The function gets a data.frame with either the start state of the exiting observations, or the end state of the entering observations and must return a modified version of that data.frame. See the Match, Enter, and Exit section for more information.

exit
functions that calculate a start state for new observations that appear in to or an end state for observations that are not present in to. If NULL the new/old observations will not be part of the tween. The function gets a data.frame with either the start state of the exiting observations, or the end state of the entering observations and must return a modified version of that data.frame. See the Match, Enter, and Exit section for more information.

enter_length
The lenght of the opening and closing transitions if enter and/or exit is given. Measured in the same units as time

exit_length
The lenght of the opening and closing transitions if enter and/or exit is given. Measured in the same units as time

Value

A data.frame with the same columns as .data along with .id giving the component id, .phase giving the state of each component in each frame, and .frame giving the frame membership of each row.

See also

Examples

d <- data.frame(
  x = runif(20),
  y = runif(20),
  time = runif(20),
  duration = runif(20, max = 0.1)
)
from_left <- function(x) {
  x$x <- -0.5
  x
}
to_right <- function(x) {
  x$x <- 1.5
  x
}

tween_events(d, 'cubic-in-out', 50, start = time, end = time + duration,
             enter = from_left, exit = to_right, enter_length = 0.1,
             exit_length = 0.05)

#>               x          y       time   duration .id .phase .frame
#> 176 -0.50000000 0.63454427 0.02387676 0.09815419  19  enter      1
#> 177 -0.46118069 0.63454427 0.02387676 0.09815419  19  enter      2
#> 62  -0.50000000 0.51680410 0.06994552 0.08973044   7  enter      3
#> 178 -0.18944556 0.63454427 0.02387676 0.09815419  19  enter      3
#> 63  -0.44722319 0.51680410 0.06994552 0.08973044   7  enter      4
#> 179  0.08228958 0.63454427 0.02387676 0.09815419  19  enter      4
#> 53  -0.50000000 0.45185844 0.10904879 0.04357182   6  enter      5
#> 64  -0.07778548 0.51680410 0.06994552 0.08973044   7  enter      5
#> 180  0.12110888 0.63454427 0.02387676 0.09815419  19 static      5
#> 54  -0.41263109 0.45185844 0.10904879 0.04357182   6  enter      6
#> 65   0.29165222 0.51680410 0.06994552 0.08973044   7  enter      6
#> 107 -0.50000000 0.43052065 0.14871035 0.05109649  12  enter      6
#> 181  0.12110888 0.63454427 0.02387676 0.09815419  19 static      6
#> 55   0.19895125 0.45185844 0.10904879 0.04357182   6  enter      7
#> 66   0.34442903 0.51680410 0.06994552 0.08973044   7 static      7
#> 108 -0.46168453 0.43052065 0.14871035 0.05109649  12  enter      7
#> 182  0.12110888 0.63454427 0.02387676 0.09815419  19 static      7
#> 56   0.81053360 0.45185844 0.10904879 0.04357182   6  enter      8
#> 67   0.34442903 0.51680410 0.06994552 0.08973044   7 static      8
#> 109 -0.19347625 0.43052065 0.14871035 0.05109649  12  enter      8
#> 126 -0.50000000 0.45112033 0.20123842 0.02417761  14  enter      8
#> 183  0.12110888 0.63454427 0.02387676 0.09815419  19 static      8
#> 31  -0.50000000 0.66589079 0.22511539 0.09202758   4  enter      9
#> 57   0.89790251 0.45185844 0.10904879 0.04357182   6 static      9
#> 68   0.34442903 0.51680410 0.06994552 0.08973044   7 static      9
#> 110  0.07473203 0.43052065 0.14871035 0.05109649  12  enter      9
#> 127 -0.46326135 0.45112033 0.20123842 0.02417761  14  enter      9
#> 184  0.12110888 0.63454427 0.02387676 0.09815419  19    raw      9
#> 32  -0.47639367 0.66589079 0.22511539 0.09202758   4  enter     10
#> 58   0.89790251 0.45185844 0.10904879 0.04357182   6 static     10
#> 69   0.34442903 0.51680410 0.06994552 0.08973044   7 static     10
#> 111  0.11304750 0.43052065 0.14871035 0.05109649  12 static     10
#> 128 -0.20609081 0.45112033 0.20123842 0.02417761  14  enter     10
#> 185  0.81055444 0.63454427 0.02387676 0.09815419  19   exit     10
#> 33  -0.31114938 0.66589079 0.22511539 0.09202758   4  enter     11
#> 59   0.89790251 0.45185844 0.10904879 0.04357182   6    raw     11
#> 70   0.34442903 0.51680410 0.06994552 0.08973044   7    raw     11
#> 90  -0.50000000 0.73122915 0.25920274 0.05386331  10  enter     11
#> 112  0.11304750 0.43052065 0.14871035 0.05109649  12 static     11
#> 129  0.35417359 0.45112033 0.20123842 0.02417761  14  enter     11
#> 186  1.50000000 0.63454427 0.02387676 0.09815419  19   exit     11
#> 34   0.04884712 0.66589079 0.22511539 0.09202758   4  enter     12
#> 60   1.19895125 0.45185844 0.10904879 0.04357182   6   exit     12
#> 71   0.92221452 0.51680410 0.06994552 0.08973044   7   exit     12
#> 91  -0.41547298 0.73122915 0.25920274 0.05386331  10  enter     12
#> 113  0.11304750 0.43052065 0.14871035 0.05109649  12 static     12
#> 130  0.61134413 0.45112033 0.20123842 0.02417761  14  enter     12
#> 35   0.21409141 0.66589079 0.22511539 0.09202758   4  enter     13
#> 61   1.50000000 0.45185844 0.10904879 0.04357182   6   exit     13
#> 72   1.50000000 0.51680410 0.06994552 0.08973044   7   exit     13
#> 92   0.17621619 0.73122915 0.25920274 0.05386331  10  enter     13
#> 114  0.11304750 0.43052065 0.14871035 0.05109649  12    raw     13
#> 131  0.64808278 0.45112033 0.20123842 0.02417761  14 static     13
#> 36   0.23769774 0.66589079 0.22511539 0.09202758   4 static     14
#> 93   0.76790535 0.73122915 0.25920274 0.05386331  10  enter     14
#> 115  0.80652375 0.43052065 0.14871035 0.05109649  12   exit     14
#> 132  0.64808278 0.45112033 0.20123842 0.02417761  14    raw     14
#> 37   0.23769774 0.66589079 0.22511539 0.09202758   4 static     15
#> 94   0.85243238 0.73122915 0.25920274 0.05386331  10 static     15
#> 116  1.50000000 0.43052065 0.14871035 0.05109649  12   exit     15
#> 133  1.07404139 0.45112033 0.20123842 0.02417761  14   exit     15
#> 38   0.23769774 0.66589079 0.22511539 0.09202758   4 static     16
#> 95   0.85243238 0.73122915 0.25920274 0.05386331  10 static     16
#> 134  1.50000000 0.45112033 0.20123842 0.02417761  14   exit     16
#> 39   0.23769774 0.66589079 0.22511539 0.09202758   4    raw     17
#> 96   0.85243238 0.73122915 0.25920274 0.05386331  10    raw     17
#> 40   0.42470548 0.66589079 0.22511539 0.09202758   4   exit     18
#> 97   1.17621619 0.73122915 0.25920274 0.05386331  10   exit     18
#> 41   1.31299226 0.66589079 0.22511539 0.09202758   4   exit     19
#> 98   1.50000000 0.73122915 0.25920274 0.05386331  10   exit     19
#> 42   1.50000000 0.66589079 0.22511539 0.09202758   4   exit     20
#> 22  -0.50000000 0.03442923 0.52870927 0.04327882   3  enter     22
#> 23  -0.41366794 0.03442923 0.52870927 0.04327882   3  enter     23
#> 24   0.19065646 0.03442923 0.52870927 0.04327882   3  enter     24
#> 25   0.79498087 0.03442923 0.52870927 0.04327882   3  enter     25
#> 26   0.88131292 0.03442923 0.52870927 0.04327882   3 static     26
#> 12  -0.50000000 0.61208279 0.65111864 0.06000373   2  enter     27
#> 27   0.88131292 0.03442923 0.52870927 0.04327882   3 static     27
#> 117 -0.50000000 0.14540735 0.64068048 0.05437927  13  enter     27
#> 13  -0.41633563 0.61208279 0.65111864 0.06000373   2  enter     28
#> 28   0.88131292 0.03442923 0.52870927 0.04327882   3    raw     28
#> 83  -0.50000000 0.88093309 0.65832223 0.01281158   9  enter     28
#> 118 -0.40810861 0.14540735 0.64068048 0.05437927  13  enter     28
#> 14   0.16931499 0.61208279 0.65111864 0.06000373   2  enter     29
#> 29   1.19065646 0.03442923 0.52870927 0.04327882   3   exit     29
#> 84  -0.42305602 0.88093309 0.65832223 0.01281158   9  enter     29
#> 119  0.23513114 0.14540735 0.64068048 0.05437927  13  enter     29
#> 15   0.75496562 0.61208279 0.65111864 0.06000373   2  enter     30
#> 30   1.50000000 0.03442923 0.52870927 0.04327882   3   exit     30
#> 85   0.11555182 0.88093309 0.65832223 0.01281158   9  enter     30
#> 120  0.87837089 0.14540735 0.64068048 0.05437927  13  enter     30
#> 16   0.83862999 0.61208279 0.65111864 0.06000373   2 static     31
#> 86   0.65415967 0.88093309 0.65832223 0.01281158   9  enter     31
#> 121  0.97026228 0.14540735 0.64068048 0.05437927  13 static     31
#> 135 -0.50000000 0.57878149 0.74362132 0.08031441  15  enter     31
#> 17   0.83862999 0.61208279 0.65111864 0.06000373   2 static     32
#> 87   0.73110365 0.88093309 0.65832223 0.01281158   9    raw     32
#> 122  0.97026228 0.14540735 0.64068048 0.05437927  13 static     32
#> 136 -0.46524543 0.57878149 0.74362132 0.08031441  15  enter     32
#> 18   0.83862999 0.61208279 0.65111864 0.06000373   2 static     33
#> 73  -0.50000000 0.67661080 0.77764551 0.07690135   8  enter     33
#> 88   1.11555182 0.88093309 0.65832223 0.01281158   9   exit     33
#> 123  0.97026228 0.14540735 0.64068048 0.05437927  13    raw     33
#> 137 -0.22196348 0.57878149 0.74362132 0.08031441  15  enter     33
#> 187 -0.50000000 0.13662858 0.79537771 0.03541867  20  enter     33
#> 19   0.83862999 0.61208279 0.65111864 0.06000373   2    raw     34
#> 74  -0.42119636 0.67661080 0.77764551 0.07690135   8  enter     34
#> 89   1.50000000 0.88093309 0.65832223 0.01281158   9   exit     34
#> 124  1.23513114 0.14540735 0.64068048 0.05437927  13   exit     34
#> 138  0.02131848 0.57878149 0.74362132 0.08031441  15  enter     34
#> 167 -0.50000000 0.54922071 0.80341257 0.05147446  18  enter     34
#> 188 -0.42941147 0.13662858 0.79537771 0.03541867  20  enter     34
#> 20   1.16931499 0.61208279 0.65111864 0.06000373   2   exit     35
#> 75   0.13042911 0.67661080 0.77764551 0.07690135   8  enter     35
#> 125  1.50000000 0.14540735 0.64068048 0.05437927  13   exit     35
#> 139  0.05607304 0.57878149 0.74362132 0.08031441  15 static     35
#> 146 -0.50000000 0.07041712 0.84763743 0.09479137  16  enter     35
#> 168 -0.43060627 0.54922071 0.80341257 0.05147446  18  enter     35
#> 189  0.06470825 0.13662858 0.79537771 0.03541867  20  enter     35
#> 21   1.50000000 0.61208279 0.65111864 0.06000373   2   exit     36
#> 76   0.68205459 0.67661080 0.77764551 0.07690135   8  enter     36
#> 140  0.05607304 0.57878149 0.74362132 0.08031441  15 static     36
#> 147 -0.46885961 0.07041712 0.84763743 0.09479137  16  enter     36
#> 158 -0.50000000 0.74232618 0.85363587 0.04073613  17  enter     36
#> 169  0.05514981 0.54922071 0.80341257 0.05147446  18  enter     36
#> 190  0.55882798 0.13662858 0.79537771 0.03541867  20  enter     36
#> 77   0.76085823 0.67661080 0.77764551 0.07690135   8 static     37
#> 141  0.05607304 0.57878149 0.74362132 0.08031441  15 static     37
#> 148 -0.25087690 0.07041712 0.84763743 0.09479137  16  enter     37
#> 159 -0.45033594 0.74232618 0.85363587 0.04073613  17  enter     37
#> 170  0.54090590 0.54922071 0.80341257 0.05147446  18  enter     37
#> 191  0.62941651 0.13662858 0.79537771 0.03541867  20 static     37
#> 78   0.76085823 0.67661080 0.77764551 0.07690135   8 static     38
#> 99  -0.50000000 0.39861982 0.91833784 0.01290162  11  enter     38
#> 142  0.05607304 0.57878149 0.74362132 0.08031441  15 static     38
#> 149  0.22401400 0.07041712 0.84763743 0.09479137  16  enter     38
#> 160 -0.10268749 0.74232618 0.85363587 0.04073613  17  enter     38
#> 171  0.61029962 0.54922071 0.80341257 0.05147446  18 static     38
#> 192  0.62941651 0.13662858 0.79537771 0.03541867  20 static     38
#> 1   -0.50000000 0.71195530 0.93023350 0.08865505   1  enter     39
#> 79   0.76085823 0.67661080 0.77764551 0.07690135   8 static     39
#> 100 -0.45277382 0.39861982 0.91833784 0.01290162  11  enter     39
#> 143  0.05607304 0.57878149 0.74362132 0.08031441  15    raw     39
#> 150  0.44199671 0.07041712 0.84763743 0.09479137  16  enter     39
#> 161  0.24496095 0.74232618 0.85363587 0.04073613  17  enter     39
#> 172  0.61029962 0.54922071 0.80341257 0.05147446  18 static     39
#> 193  0.62941651 0.13662858 0.79537771 0.03541867  20    raw     39
#> 2   -0.41504463 0.71195530 0.93023350 0.08865505   1  enter     40
#> 80   0.76085823 0.67661080 0.77764551 0.07690135   8    raw     40
#> 101 -0.12219058 0.39861982 0.91833784 0.01290162  11  enter     40
#> 144  0.77803652 0.57878149 0.74362132 0.08031441  15   exit     40
#> 151  0.47313710 0.07041712 0.84763743 0.09479137  16 static     40
#> 162  0.29462501 0.74232618 0.85363587 0.04073613  17 static     40
#> 173  0.61029962 0.54922071 0.80341257 0.05147446  18    raw     40
#> 194  1.06470825 0.13662858 0.79537771 0.03541867  20   exit     40
#> 3    0.17964295 0.71195530 0.93023350 0.08865505   1  enter     41
#> 43  -0.50000000 0.65296728 0.98326687 0.06160144   5  enter     41
#> 81   1.13042911 0.67661080 0.77764551 0.07690135   8   exit     41
#> 102  0.59800862 0.39861982 0.91833784 0.01290162  11  enter     41
#> 145  1.50000000 0.57878149 0.74362132 0.08031441  15   exit     41
#> 152  0.47313710 0.07041712 0.84763743 0.09479137  16 static     41
#> 163  0.29462501 0.74232618 0.85363587 0.04073613  17 static     41
#> 174  1.05514981 0.54922071 0.80341257 0.05147446  18   exit     41
#> 195  1.50000000 0.13662858 0.79537771 0.03541867  20   exit     41
#> 4    0.77433053 0.71195530 0.93023350 0.08865505   1  enter     42
#> 44  -0.43258557 0.65296728 0.98326687 0.06160144   5  enter     42
#> 82   1.50000000 0.67661080 0.77764551 0.07690135   8   exit     42
#> 103  0.92859186 0.39861982 0.91833784 0.01290162  11  enter     42
#> 153  0.47313710 0.07041712 0.84763743 0.09479137  16 static     42
#> 164  0.29462501 0.74232618 0.85363587 0.04073613  17    raw     42
#> 175  1.50000000 0.54922071 0.80341257 0.05147446  18   exit     42
#> 5    0.85928589 0.71195530 0.93023350 0.08865505   1 static     43
#> 45   0.03931542 0.65296728 0.98326687 0.06160144   5  enter     43
#> 104  0.97581804 0.39861982 0.91833784 0.01290162  11    raw     43
#> 154  0.47313710 0.07041712 0.84763743 0.09479137  16 static     43
#> 165  0.89731251 0.74232618 0.85363587 0.04073613  17   exit     43
#> 6    0.85928589 0.71195530 0.93023350 0.08865505   1 static     44
#> 46   0.51121642 0.65296728 0.98326687 0.06160144   5  enter     44
#> 105  1.23790902 0.39861982 0.91833784 0.01290162  11   exit     44
#> 155  0.47313710 0.07041712 0.84763743 0.09479137  16    raw     44
#> 166  1.50000000 0.74232618 0.85363587 0.04073613  17   exit     44
#> 7    0.85928589 0.71195530 0.93023350 0.08865505   1 static     45
#> 47   0.57863085 0.65296728 0.98326687 0.06160144   5 static     45
#> 106  1.50000000 0.39861982 0.91833784 0.01290162  11   exit     45
#> 156  0.98656855 0.07041712 0.84763743 0.09479137  16   exit     45
#> 8    0.85928589 0.71195530 0.93023350 0.08865505   1 static     46
#> 48   0.57863085 0.65296728 0.98326687 0.06160144   5 static     46
#> 157  1.50000000 0.07041712 0.84763743 0.09479137  16   exit     46
#> 9    0.85928589 0.71195530 0.93023350 0.08865505   1    raw     47
#> 49   0.57863085 0.65296728 0.98326687 0.06160144   5 static     47
#> 10   1.17964295 0.71195530 0.93023350 0.08865505   1   exit     48
#> 50   0.57863085 0.65296728 0.98326687 0.06160144   5    raw     48
#> 11   1.50000000 0.71195530 0.93023350 0.08865505   1   exit     49
#> 51   1.03931542 0.65296728 0.98326687 0.06160144   5   exit     49
#> 52   1.50000000 0.65296728 0.98326687 0.06160144   5   exit     50