Transition in and out of events

This tweening function is a more powerful version of tween_appear(), with support for newer features such as enter/exits and tween phase identification. The tweener treats each row in the data as unique events in time, and creates frames with the correct events present at any given time.

tween_events(
  .data,
  ease,
  nframes,
  start,
  end = NULL,
  range = NULL,
  enter = NULL,
  exit = NULL,
  enter_length = 0,
  exit_length = 0
)

Arguments

.data

A data.frame with components at different stages

ease

The easing function to use. Either a single string or one for each column in the data set.

nframes

The number of frames to calculate for the tween

start, end

The start (and potential end) of the event encoded in the row, as unquoted expressions. Will be evaluated in the context of .data so can refer to columns in it. If end = NULL the event will be without extend and only visible in a single frame, unless enter and/or exit is given.

range

The range of time points to include in the tween. If NULL it will use the range of time

enter, exit

functions that calculate a start state for new observations that appear in to or an end state for observations that are not present in to. If NULL the new/old observations will not be part of the tween. The function gets a data.frame with either the start state of the exiting observations, or the end state of the entering observations and must return a modified version of that data.frame. See the Match, Enter, and Exit section for more information.

enter_length, exit_length

The lenght of the opening and closing transitions if enter and/or exit is given. Measured in the same units as time

Value

A data.frame with the same columns as .data along with .id giving the component id, .phase giving the state of each component in each frame, and .frame giving the frame membership of each row.

See also

Examples

d <- data.frame(
  x = runif(20),
  y = runif(20),
  time = runif(20),
  duration = runif(20, max = 0.1)
)
from_left <- function(x) {
  x$x <- -0.5
  x
}
to_right <- function(x) {
  x$x <- 1.5
  x
}

tween_events(d, 'cubic-in-out', 50, start = time, end = time + duration,
             enter = from_left, exit = to_right, enter_length = 0.1,
             exit_length = 0.05)
#>                x          y       time    duration .id .phase .frame
#> 66  -0.500000000 0.51128858 0.01163884 0.047463264   8  enter      1
#> 67  -0.422055098 0.51128858 0.01163884 0.047463264   8  enter      2
#> 68   0.123559218 0.51128858 0.01163884 0.047463264   8  enter      3
#> 176 -0.500000000 0.93824842 0.05652764 0.029135661  20  enter      3
#> 69   0.669173534 0.51128858 0.01163884 0.047463264   8  enter      4
#> 177 -0.456710346 0.93824842 0.05652764 0.029135661  20  enter      4
#> 70   0.747118436 0.51128858 0.01163884 0.047463264   8 static      5
#> 178 -0.153682771 0.93824842 0.05652764 0.029135661  20  enter      5
#> 71   0.747118436 0.51128858 0.01163884 0.047463264   8 static      6
#> 179  0.149344803 0.93824842 0.05652764 0.029135661  20  enter      6
#> 72   0.747118436 0.51128858 0.01163884 0.047463264   8    raw      7
#> 180  0.192634457 0.93824842 0.05652764 0.029135661  20 static      7
#> 19  -0.500000000 0.10484113 0.17611345 0.083708017   3  enter      8
#> 73   1.123559218 0.51128858 0.01163884 0.047463264   8   exit      8
#> 181  0.192634457 0.93824842 0.05652764 0.029135661  20    raw      8
#> 20  -0.447150295 0.10484113 0.17611345 0.083708017   3  enter      9
#> 49  -0.500000000 0.11356415 0.21753560 0.008825088   6  enter      9
#> 74   1.500000000 0.51128858 0.01163884 0.047463264   8   exit      9
#> 182  0.846317229 0.93824842 0.05652764 0.029135661  20   exit      9
#> 21  -0.077202361 0.10484113 0.17611345 0.083708017   3  enter     10
#> 50  -0.461413265 0.11356415 0.21753560 0.008825088   6  enter     10
#> 183  1.500000000 0.93824842 0.05652764 0.029135661  20   exit     10
#> 22   0.292745574 0.10484113 0.17611345 0.083708017   3  enter     11
#> 51  -0.191306119 0.11356415 0.21753560 0.008825088   6  enter     11
#> 23   0.345595279 0.10484113 0.17611345 0.083708017   3 static     12
#> 52   0.078801027 0.11356415 0.21753560 0.008825088   6  enter     12
#> 24   0.345595279 0.10484113 0.17611345 0.083708017   3 static     13
#> 53   0.117387762 0.11356415 0.21753560 0.008825088   6 static     13
#> 25   0.345595279 0.10484113 0.17611345 0.083708017   3 static     14
#> 54   0.117387762 0.11356415 0.21753560 0.008825088   6    raw     14
#> 57  -0.500000000 0.16104651 0.33015309 0.041038760   7  enter     14
#> 26   0.345595279 0.10484113 0.17611345 0.083708017   3    raw     15
#> 55   0.808693881 0.11356415 0.21753560 0.008825088   6   exit     15
#> 58  -0.468458781 0.16104651 0.33015309 0.041038760   7  enter     15
#> 27   0.922797639 0.10484113 0.17611345 0.083708017   3   exit     16
#> 56   1.500000000 0.11356415 0.21753560 0.008825088   6   exit     16
#> 59  -0.247670251 0.16104651 0.33015309 0.041038760   7  enter     16
#> 147 -0.500000000 0.61891769 0.39148159 0.081205372  17  enter     16
#> 166 -0.500000000 0.85651102 0.38575146 0.068572579  19  enter     16
#> 28   1.500000000 0.10484113 0.17611345 0.083708017   3   exit     17
#> 60  -0.026881721 0.16104651 0.33015309 0.041038760   7  enter     17
#> 84  -0.500000000 0.06802661 0.40071388 0.078687881  10  enter     17
#> 148 -0.456280223 0.61891769 0.39148159 0.081205372  17  enter     17
#> 167 -0.418578514 0.85651102 0.38575146 0.068572579  19  enter     17
#> 61   0.004659497 0.16104651 0.33015309 0.041038760   7 static     18
#> 85  -0.416800653 0.06802661 0.40071388 0.078687881  10  enter     18
#> 149 -0.150241786 0.61891769 0.39148159 0.081205372  17  enter     18
#> 168  0.151371889 0.85651102 0.38575146 0.068572579  19  enter     18
#> 62   0.004659497 0.16104651 0.33015309 0.041038760   7 static     19
#> 86   0.165594775 0.06802661 0.40071388 0.078687881  10  enter     19
#> 138 -0.500000000 0.47442698 0.45236010 0.045554379  16  enter     19
#> 150  0.155796651 0.61891769 0.39148159 0.081205372  17  enter     19
#> 169  0.721322291 0.85651102 0.38575146 0.068572579  19  enter     19
#> 63   0.004659497 0.16104651 0.33015309 0.041038760   7    raw     20
#> 87   0.747990204 0.06802661 0.40071388 0.078687881  10  enter     20
#> 112 -0.500000000 0.77534198 0.48686312 0.001953115  13  enter     20
#> 139 -0.433727603 0.47442698 0.45236010 0.045554379  16  enter     20
#> 151  0.199516427 0.61891769 0.39148159 0.081205372  17 static     20
#> 170  0.802743777 0.85651102 0.38575146 0.068572579  19 static     20
#> 1   -0.500000000 0.01592739 0.50891464 0.028767555   1  enter     21
#> 64   0.752329749 0.16104651 0.33015309 0.041038760   7   exit     21
#> 88   0.831189550 0.06802661 0.40071388 0.078687881  10 static     21
#> 113 -0.467346272 0.77534198 0.48686312 0.001953115  13  enter     21
#> 119 -0.500000000 0.35649725 0.51160164 0.048572367  14  enter     21
#> 140  0.030179178 0.47442698 0.45236010 0.045554379  16  enter     21
#> 152  0.199516427 0.61891769 0.39148159 0.081205372  17 static     21
#> 171  0.802743777 0.85651102 0.38575146 0.068572579  19 static     21
#> 2   -0.435507693 0.01592739 0.50891464 0.028767555   1  enter     22
#> 65   1.500000000 0.16104651 0.33015309 0.041038760   7   exit     22
#> 89   0.831189550 0.06802661 0.40071388 0.078687881  10 static     22
#> 114 -0.238770174 0.77534198 0.48686312 0.001953115  13  enter     22
#> 120 -0.452323907 0.35649725 0.51160164 0.048572367  14  enter     22
#> 141  0.494085960 0.47442698 0.45236010 0.045554379  16  enter     22
#> 153  0.199516427 0.61891769 0.39148159 0.081205372  17 static     22
#> 172  0.802743777 0.85651102 0.38575146 0.068572579  19 static     22
#> 3    0.015938460 0.01592739 0.50891464 0.028767555   1  enter     23
#> 90   0.831189550 0.06802661 0.40071388 0.078687881  10 static     23
#> 115 -0.010194077 0.77534198 0.48686312 0.001953115  13  enter     23
#> 121 -0.118591255 0.35649725 0.51160164 0.048572367  14  enter     23
#> 142  0.560358357 0.47442698 0.45236010 0.045554379  16 static     23
#> 154  0.199516427 0.61891769 0.39148159 0.081205372  17 static     23
#> 173  0.802743777 0.85651102 0.38575146 0.068572579  19    raw     23
#> 4    0.467384612 0.01592739 0.50891464 0.028767555   1  enter     24
#> 91   0.831189550 0.06802661 0.40071388 0.078687881  10    raw     24
#> 116  0.022459652 0.77534198 0.48686312 0.001953115  13    raw     24
#> 122  0.215141397 0.35649725 0.51160164 0.048572367  14  enter     24
#> 143  0.560358357 0.47442698 0.45236010 0.045554379  16 static     24
#> 155  0.199516427 0.61891769 0.39148159 0.081205372  17    raw     24
#> 174  1.151371889 0.85651102 0.38575146 0.068572579  19   exit     24
#> 5    0.531876920 0.01592739 0.50891464 0.028767555   1 static     25
#> 92   1.165594775 0.06802661 0.40071388 0.078687881  10   exit     25
#> 117  0.761229826 0.77534198 0.48686312 0.001953115  13   exit     25
#> 123  0.262817491 0.35649725 0.51160164 0.048572367  14 static     25
#> 144  0.560358357 0.47442698 0.45236010 0.045554379  16    raw     25
#> 156  0.849758214 0.61891769 0.39148159 0.081205372  17   exit     25
#> 175  1.500000000 0.85651102 0.38575146 0.068572579  19   exit     25
#> 6    0.531876920 0.01592739 0.50891464 0.028767555   1    raw     26
#> 75  -0.500000000 0.03618118 0.63779044 0.034961803   9  enter     26
#> 93   1.500000000 0.06802661 0.40071388 0.078687881  10   exit     26
#> 118  1.500000000 0.77534198 0.48686312 0.001953115  13   exit     26
#> 124  0.262817491 0.35649725 0.51160164 0.048572367  14 static     26
#> 145  1.030179178 0.47442698 0.45236010 0.045554379  16   exit     26
#> 157  1.500000000 0.61891769 0.39148159 0.081205372  17   exit     26
#> 7    1.015938460 0.01592739 0.50891464 0.028767555   1   exit     27
#> 76  -0.442756309 0.03618118 0.63779044 0.034961803   9  enter     27
#> 125  0.262817491 0.35649725 0.51160164 0.048572367  14    raw     27
#> 146  1.500000000 0.47442698 0.45236010 0.045554379  16   exit     27
#> 8    1.500000000 0.01592739 0.50891464 0.028767555   1   exit     28
#> 39  -0.500000000 0.83606381 0.67109100 0.092045692   5  enter     28
#> 77  -0.042050475 0.03618118 0.63779044 0.034961803   9  enter     28
#> 126  0.881408745 0.35649725 0.51160164 0.048572367  14   exit     28
#> 40  -0.431453087 0.83606381 0.67109100 0.092045692   5  enter     29
#> 78   0.358655360 0.03618118 0.63779044 0.034961803   9  enter     29
#> 127  1.500000000 0.35649725 0.51160164 0.048572367  14   exit     29
#> 158 -0.500000000 0.51228037 0.70513912 0.022211988  18  enter     29
#> 29  -0.500000000 0.20620797 0.73801541 0.054452760   4  enter     30
#> 41   0.048375302 0.83606381 0.67109100 0.092045692   5  enter     30
#> 79   0.415899051 0.03618118 0.63779044 0.034961803   9 static     30
#> 159 -0.410282231 0.51228037 0.70513912 0.022211988  18  enter     30
#> 30  -0.425796802 0.20620797 0.73801541 0.054452760   4  enter     31
#> 42   0.528203691 0.83606381 0.67109100 0.092045692   5  enter     31
#> 80   0.415899051 0.03618118 0.63779044 0.034961803   9 static     31
#> 160  0.217742153 0.51228037 0.70513912 0.022211988  18  enter     31
#> 9   -0.500000000 0.09910131 0.78383703 0.072020109   2  enter     32
#> 31   0.093625582 0.20620797 0.73801541 0.054452760   4  enter     32
#> 43   0.596750604 0.83606381 0.67109100 0.092045692   5 static     32
#> 81   0.415899051 0.03618118 0.63779044 0.034961803   9    raw     32
#> 161  0.845766536 0.51228037 0.70513912 0.022211988  18  enter     32
#> 10  -0.435151951 0.09910131 0.78383703 0.072020109   2  enter     33
#> 32   0.613047967 0.20620797 0.73801541 0.054452760   4  enter     33
#> 44   0.596750604 0.83606381 0.67109100 0.092045692   5 static     33
#> 82   0.957949525 0.03618118 0.63779044 0.034961803   9   exit     33
#> 162  0.935484305 0.51228037 0.70513912 0.022211988  18 static     33
#> 11   0.018784390 0.09910131 0.78383703 0.072020109   2  enter     34
#> 33   0.687251164 0.20620797 0.73801541 0.054452760   4 static     34
#> 45   0.596750604 0.83606381 0.67109100 0.092045692   5 static     34
#> 83   1.500000000 0.03618118 0.63779044 0.034961803   9   exit     34
#> 94  -0.500000000 0.45091928 0.83502619 0.007346922  11  enter     34
#> 102 -0.500000000 0.69774674 0.82275058 0.075898010  12  enter     34
#> 163  0.935484305 0.51228037 0.70513912 0.022211988  18    raw     34
#> 12   0.472720732 0.09910131 0.78383703 0.072020109   2  enter     35
#> 34   0.687251164 0.20620797 0.73801541 0.054452760   4 static     35
#> 46   0.596750604 0.83606381 0.67109100 0.092045692   5    raw     35
#> 95  -0.446211309 0.45091928 0.83502619 0.007346922  11  enter     35
#> 103 -0.408132852 0.69774674 0.82275058 0.075898010  12  enter     35
#> 164  1.217742153 0.51228037 0.70513912 0.022211988  18   exit     35
#> 13   0.537568781 0.09910131 0.78383703 0.072020109   2 static     36
#> 35   0.687251164 0.20620797 0.73801541 0.054452760   4 static     36
#> 47   1.048375302 0.83606381 0.67109100 0.092045692   5   exit     36
#> 96  -0.069690469 0.45091928 0.83502619 0.007346922  11  enter     36
#> 104  0.234937181 0.69774674 0.82275058 0.075898010  12  enter     36
#> 165  1.500000000 0.51228037 0.70513912 0.022211988  18   exit     36
#> 14   0.537568781 0.09910131 0.78383703 0.072020109   2 static     37
#> 36   0.687251164 0.20620797 0.73801541 0.054452760   4    raw     37
#> 48   1.500000000 0.83606381 0.67109100 0.092045692   5   exit     37
#> 97   0.306830371 0.45091928 0.83502619 0.007346922  11  enter     37
#> 105  0.878007214 0.69774674 0.82275058 0.075898010  12  enter     37
#> 15   0.537568781 0.09910131 0.78383703 0.072020109   2 static     38
#> 37   1.093625582 0.20620797 0.73801541 0.054452760   4   exit     38
#> 98   0.360619063 0.45091928 0.83502619 0.007346922  11 static     38
#> 106  0.969874362 0.69774674 0.82275058 0.075898010  12 static     38
#> 16   0.537568781 0.09910131 0.78383703 0.072020109   2    raw     39
#> 38   1.500000000 0.20620797 0.73801541 0.054452760   4   exit     39
#> 99   0.360619063 0.45091928 0.83502619 0.007346922  11    raw     39
#> 107  0.969874362 0.69774674 0.82275058 0.075898010  12 static     39
#> 17   1.018784390 0.09910131 0.78383703 0.072020109   2   exit     40
#> 100  0.930309531 0.45091928 0.83502619 0.007346922  11   exit     40
#> 108  0.969874362 0.69774674 0.82275058 0.075898010  12 static     40
#> 18   1.500000000 0.09910131 0.78383703 0.072020109   2   exit     41
#> 101  1.500000000 0.45091928 0.83502619 0.007346922  11   exit     41
#> 109  0.969874362 0.69774674 0.82275058 0.075898010  12    raw     41
#> 128 -0.500000000 0.30752641 0.99217647 0.085187192  15  enter     41
#> 110  1.234937181 0.69774674 0.82275058 0.075898010  12   exit     42
#> 129 -0.458007421 0.30752641 0.99217647 0.085187192  15  enter     42
#> 111  1.500000000 0.69774674 0.82275058 0.075898010  12   exit     43
#> 130 -0.164059369 0.30752641 0.99217647 0.085187192  15  enter     43
#> 131  0.129888684 0.30752641 0.99217647 0.085187192  15  enter     44
#> 132  0.171881263 0.30752641 0.99217647 0.085187192  15 static     45
#> 133  0.171881263 0.30752641 0.99217647 0.085187192  15 static     46
#> 134  0.171881263 0.30752641 0.99217647 0.085187192  15 static     47
#> 135  0.171881263 0.30752641 0.99217647 0.085187192  15    raw     48
#> 136  0.835940631 0.30752641 0.99217647 0.085187192  15   exit     49
#> 137  1.500000000 0.30752641 0.99217647 0.085187192  15   exit     50