Transition in and out of events

This tweening function is a more powerful version of tween_appear(), with support for newer features such as enter/exits and tween phase identification. The tweener treats each row in the data as unique events in time, and creates frames with the correct events present at any given time.

tween_events(
  .data,
  ease,
  nframes,
  start,
  end = NULL,
  range = NULL,
  enter = NULL,
  exit = NULL,
  enter_length = 0,
  exit_length = 0
)

Arguments

.data

A data.frame with components at different stages

ease

The easing function to use. Either a single string or one for each column in the data set.

nframes

The number of frames to calculate for the tween

start, end

The start (and potential end) of the event encoded in the row, as unquoted expressions. Will be evaluated in the context of .data so can refer to columns in it. If end = NULL the event will be without extend and only visible in a single frame, unless enter and/or exit is given.

range

The range of time points to include in the tween. If NULL it will use the range of time

enter, exit

functions that calculate a start state for new observations that appear in to or an end state for observations that are not present in to. If NULL the new/old observations will not be part of the tween. The function gets a data.frame with either the start state of the exiting observations, or the end state of the entering observations and must return a modified version of that data.frame. See the Match, Enter, and Exit section for more information.

enter_length, exit_length

The lenght of the opening and closing transitions if enter and/or exit is given. Measured in the same units as time

Value

A data.frame with the same columns as .data along with .id giving the component id, .phase giving the state of each component in each frame, and .frame giving the frame membership of each row.

See also

Examples

d <- data.frame(
  x = runif(20),
  y = runif(20),
  time = runif(20),
  duration = runif(20, max = 0.1)
)
from_left <- function(x) {
  x$x <- -0.5
  x
}
to_right <- function(x) {
  x$x <- 1.5
  x
}

tween_events(d, 'cubic-in-out', 50, start = time, end = time + duration,
             enter = from_left, exit = to_right, enter_length = 0.1,
             exit_length = 0.05)
#>                x         y       time    duration .id .phase .frame
#> 73  -0.500000000 0.9038736 0.01415646 0.062066943   9  enter      1
#> 111 -0.500000000 0.5843085 0.02597290 0.004862295  13  enter      1
#> 74  -0.422007974 0.9038736 0.01415646 0.062066943   9  enter      2
#> 112 -0.476281544 0.5843085 0.02597290 0.004862295  13  enter      2
#> 75   0.123936207 0.9038736 0.01415646 0.062066943   9  enter      3
#> 83  -0.500000000 0.9040876 0.06502352 0.052697441  10  enter      3
#> 113 -0.310252353 0.5843085 0.02597290 0.004862295  13  enter      3
#> 76   0.669880389 0.9038736 0.01415646 0.062066943   9  enter      4
#> 84  -0.440221038 0.9040876 0.06502352 0.052697441  10  enter      4
#> 92  -0.500000000 0.5072273 0.09475695 0.061304461  11  enter      4
#> 114  0.051454100 0.5843085 0.02597290 0.004862295  13  enter      4
#> 77   0.747872415 0.9038736 0.01415646 0.062066943   9 static      5
#> 85  -0.021768300 0.9040876 0.06502352 0.052697441  10  enter      5
#> 93  -0.414650065 0.5072273 0.09475695 0.061304461  11  enter      5
#> 115  0.217483292 0.5843085 0.02597290 0.004862295  13  enter      5
#> 78   0.747872415 0.9038736 0.01415646 0.062066943   9 static      6
#> 86   0.396684437 0.9040876 0.06502352 0.052697441  10  enter      6
#> 94   0.182799477 0.5072273 0.09475695 0.061304461  11  enter      6
#> 116  0.241201748 0.5843085 0.02597290 0.004862295  13    raw      6
#> 63  -0.500000000 0.8096616 0.15295271 0.081698396   8  enter      7
#> 79   0.747872415 0.9038736 0.01415646 0.062066943   9 static      7
#> 87   0.456463400 0.9040876 0.06502352 0.052697441  10 static      7
#> 95   0.780249019 0.5072273 0.09475695 0.061304461  11  enter      7
#> 117  0.870600874 0.5843085 0.02597290 0.004862295  13   exit      7
#> 20  -0.500000000 0.3618477 0.17488532 0.039950198   3  enter      8
#> 64  -0.425222480 0.8096616 0.15295271 0.081698396   8  enter      8
#> 80   0.747872415 0.9038736 0.01415646 0.062066943   9    raw      8
#> 88   0.456463400 0.9040876 0.06502352 0.052697441  10 static      8
#> 96   0.865598953 0.5072273 0.09475695 0.061304461  11 static      8
#> 118  1.500000000 0.5843085 0.02597290 0.004862295  13   exit      8
#> 21  -0.426709524 0.3618477 0.17488532 0.039950198   3  enter      9
#> 65   0.098220162 0.8096616 0.15295271 0.081698396   8  enter      9
#> 81   1.123936207 0.9038736 0.01415646 0.062066943   9   exit      9
#> 89   0.456463400 0.9040876 0.06502352 0.052697441  10    raw      9
#> 97   0.865598953 0.5072273 0.09475695 0.061304461  11 static      9
#> 22   0.086323812 0.3618477 0.17488532 0.039950198   3  enter     10
#> 66   0.621662804 0.8096616 0.15295271 0.081698396   8  enter     10
#> 82   1.500000000 0.9038736 0.01415646 0.062066943   9   exit     10
#> 90   0.978231700 0.9040876 0.06502352 0.052697441  10   exit     10
#> 98   0.865598953 0.5072273 0.09475695 0.061304461  11 static     10
#> 23   0.599357147 0.3618477 0.17488532 0.039950198   3  enter     11
#> 67   0.696440324 0.8096616 0.15295271 0.081698396   8 static     11
#> 91   1.500000000 0.9040876 0.06502352 0.052697441  10   exit     11
#> 99   0.865598953 0.5072273 0.09475695 0.061304461  11    raw     11
#> 24   0.672647623 0.3618477 0.17488532 0.039950198   3 static     12
#> 68   0.696440324 0.8096616 0.15295271 0.081698396   8 static     12
#> 100  1.182799477 0.5072273 0.09475695 0.061304461  11   exit     12
#> 25   0.672647623 0.3618477 0.17488532 0.039950198   3    raw     13
#> 69   0.696440324 0.8096616 0.15295271 0.081698396   8 static     13
#> 101  1.500000000 0.5072273 0.09475695 0.061304461  11   exit     13
#> 26   1.086323812 0.3618477 0.17488532 0.039950198   3   exit     14
#> 70   0.696440324 0.8096616 0.15295271 0.081698396   8    raw     14
#> 155 -0.500000000 0.5131994 0.32447862 0.016311288  18  enter     14
#> 27   1.500000000 0.3618477 0.17488532 0.039950198   3   exit     15
#> 71   1.098220162 0.8096616 0.15295271 0.081698396   8   exit     15
#> 156 -0.406717032 0.5131994 0.32447862 0.016311288  18  enter     15
#> 12  -0.500000000 0.2690190 0.37029830 0.021646584   2  enter     16
#> 72   1.500000000 0.8096616 0.15295271 0.081698396   8   exit     16
#> 102 -0.500000000 0.3824690 0.37328208 0.046754489  12  enter     16
#> 157  0.246263745 0.5131994 0.32447862 0.016311288  18  enter     16
#> 13  -0.442593518 0.2690190 0.37029830 0.021646584   2  enter     17
#> 103 -0.446153182 0.3824690 0.37328208 0.046754489  12  enter     17
#> 158  0.899244523 0.5131994 0.32447862 0.016311288  18  enter     17
#> 14  -0.040748142 0.2690190 0.37029830 0.021646584   2  enter     18
#> 104 -0.069225458 0.3824690 0.37328208 0.046754489  12  enter     18
#> 159  0.992527491 0.5131994 0.32447862 0.016311288  18 static     18
#> 15   0.361097234 0.2690190 0.37029830 0.021646584   2  enter     19
#> 28  -0.500000000 0.2882138 0.44153688 0.019410767   4  enter     19
#> 105  0.307702266 0.3824690 0.37328208 0.046754489  12  enter     19
#> 160  0.992527491 0.5131994 0.32447862 0.016311288  18    raw     19
#> 16   0.418503717 0.2690190 0.37029830 0.021646584   2 static     20
#> 29  -0.441658044 0.2882138 0.44153688 0.019410767   4  enter     20
#> 106  0.361549083 0.3824690 0.37328208 0.046754489  12 static     20
#> 161  1.246263745 0.5131994 0.32447862 0.016311288  18   exit     20
#> 17   0.418503717 0.2690190 0.37029830 0.021646584   2    raw     21
#> 30  -0.033264352 0.2882138 0.44153688 0.019410767   4  enter     21
#> 107  0.361549083 0.3824690 0.37328208 0.046754489  12 static     21
#> 162  1.500000000 0.5131994 0.32447862 0.016311288  18   exit     21
#> 18   0.959251858 0.2690190 0.37029830 0.021646584   2   exit     22
#> 31   0.375129340 0.2882138 0.44153688 0.019410767   4  enter     22
#> 108  0.361549083 0.3824690 0.37328208 0.046754489  12    raw     22
#> 129 -0.500000000 0.2735813 0.51639866 0.037739585  15  enter     22
#> 137 -0.500000000 0.7194192 0.52675321 0.065465239  16  enter     22
#> 19   1.500000000 0.2690190 0.37029830 0.021646584   2   exit     23
#> 32   0.433471296 0.2882138 0.44153688 0.019410767   4    raw     23
#> 109  0.930774542 0.3824690 0.37328208 0.046754489  12   exit     23
#> 130 -0.460044862 0.2735813 0.51639866 0.037739585  15  enter     23
#> 138 -0.425418082 0.7194192 0.52675321 0.065465239  16  enter     23
#> 33   0.591475548 0.2882138 0.44153688 0.019410767   4   exit     24
#> 110  1.500000000 0.3824690 0.37328208 0.046754489  12   exit     24
#> 119 -0.500000000 0.5786310 0.57896891 0.065306551  14  enter     24
#> 131 -0.180358893 0.2735813 0.51639866 0.037739585  15  enter     24
#> 139  0.096655345 0.7194192 0.52675321 0.065465239  16  enter     24
#> 1   -0.500000000 0.8252333 0.60510799 0.080920276   1  enter     25
#> 34   1.341995748 0.2882138 0.44153688 0.019410767   4   exit     25
#> 120 -0.466949554 0.5786310 0.57896891 0.065306551  14  enter     25
#> 132  0.099327075 0.2735813 0.51639866 0.037739585  15  enter     25
#> 140  0.618728771 0.7194192 0.52675321 0.065465239  16  enter     25
#> 2   -0.426338745 0.8252333 0.60510799 0.080920276   1  enter     26
#> 35   1.500000000 0.2882138 0.44153688 0.019410767   4   exit     26
#> 121 -0.235596430 0.5786310 0.57896891 0.065306551  14  enter     26
#> 133  0.139282214 0.2735813 0.51639866 0.037739585  15 static     26
#> 141  0.693310689 0.7194192 0.52675321 0.065465239  16 static     26
#> 3    0.089290044 0.8252333 0.60510799 0.080920276   1  enter     27
#> 122 -0.004243306 0.5786310 0.57896891 0.065306551  14  enter     27
#> 134  0.139282214 0.2735813 0.51639866 0.037739585  15    raw     27
#> 142  0.693310689 0.7194192 0.52675321 0.065465239  16 static     27
#> 147 -0.500000000 0.6846671 0.64276190 0.022905262  17  enter     27
#> 163 -0.500000000 0.4033370 0.64978453 0.011862204  19  enter     27
#> 4    0.604918832 0.8252333 0.60510799 0.080920276   1  enter     28
#> 123  0.028807140 0.5786310 0.57896891 0.065306551  14 static     28
#> 135  0.819641107 0.2735813 0.51639866 0.037739585  15   exit     28
#> 143  0.693310689 0.7194192 0.52675321 0.065465239  16 static     28
#> 148 -0.432941566 0.6846671 0.64276190 0.022905262  17  enter     28
#> 164 -0.410480953 0.4033370 0.64978453 0.011862204  19  enter     28
#> 5    0.678580087 0.8252333 0.60510799 0.080920276   1 static     29
#> 124  0.028807140 0.5786310 0.57896891 0.065306551  14 static     29
#> 136  1.500000000 0.2735813 0.51639866 0.037739585  15   exit     29
#> 144  0.693310689 0.7194192 0.52675321 0.065465239  16    raw     29
#> 149  0.036467473 0.6846671 0.64276190 0.022905262  17  enter     29
#> 165  0.216152373 0.4033370 0.64978453 0.011862204  19  enter     29
#> 6    0.678580087 0.8252333 0.60510799 0.080920276   1 static     30
#> 125  0.028807140 0.5786310 0.57896891 0.065306551  14 static     30
#> 145  1.096655345 0.7194192 0.52675321 0.065465239  16   exit     30
#> 150  0.505876511 0.6846671 0.64276190 0.022905262  17  enter     30
#> 166  0.842785700 0.4033370 0.64978453 0.011862204  19  enter     30
#> 7    0.678580087 0.8252333 0.60510799 0.080920276   1 static     31
#> 126  0.028807140 0.5786310 0.57896891 0.065306551  14    raw     31
#> 146  1.500000000 0.7194192 0.52675321 0.065465239  16   exit     31
#> 151  0.572934945 0.6846671 0.64276190 0.022905262  17 static     31
#> 167  0.932304746 0.4033370 0.64978453 0.011862204  19 static     31
#> 8    0.678580087 0.8252333 0.60510799 0.080920276   1 static     32
#> 127  0.764403570 0.5786310 0.57896891 0.065306551  14   exit     32
#> 152  0.572934945 0.6846671 0.64276190 0.022905262  17    raw     32
#> 168  0.932304746 0.4033370 0.64978453 0.011862204  19    raw     32
#> 9    0.678580087 0.8252333 0.60510799 0.080920276   1    raw     33
#> 128  1.500000000 0.5786310 0.57896891 0.065306551  14   exit     33
#> 153  1.036467473 0.6846671 0.64276190 0.022905262  17   exit     33
#> 169  1.216152373 0.4033370 0.64978453 0.011862204  19   exit     33
#> 10   1.089290044 0.8252333 0.60510799 0.080920276   1   exit     34
#> 54  -0.500000000 0.1160586 0.82860539 0.035006098   7  enter     34
#> 154  1.500000000 0.6846671 0.64276190 0.022905262  17   exit     34
#> 170  1.500000000 0.4033370 0.64978453 0.011862204  19   exit     34
#> 11   1.500000000 0.8252333 0.60510799 0.080920276   1   exit     35
#> 55  -0.469272585 0.1160586 0.82860539 0.035006098   7  enter     35
#> 171 -0.500000000 0.3317948 0.83426745 0.043995851  20  enter     35
#> 56  -0.254180683 0.1160586 0.82860539 0.035006098   7  enter     36
#> 172 -0.413497262 0.3317948 0.83426745 0.043995851  20  enter     36
#> 57   0.214412389 0.1160586 0.82860539 0.035006098   7  enter     37
#> 173  0.192021907 0.3317948 0.83426745 0.043995851  20  enter     37
#> 58   0.429504292 0.1160586 0.82860539 0.035006098   7  enter     38
#> 174  0.797541076 0.3317948 0.83426745 0.043995851  20  enter     38
#> 59   0.460231706 0.1160586 0.82860539 0.035006098   7 static     39
#> 175  0.884043815 0.3317948 0.83426745 0.043995851  20 static     39
#> 60   0.460231706 0.1160586 0.82860539 0.035006098   7    raw     40
#> 176  0.884043815 0.3317948 0.83426745 0.043995851  20 static     40
#> 36  -0.500000000 0.5704255 0.98920100 0.066561265   5  enter     41
#> 46  -0.500000000 0.1464387 0.99480482 0.001393738   6  enter     41
#> 61   0.980115853 0.1160586 0.82860539 0.035006098   7   exit     41
#> 177  0.884043815 0.3317948 0.83426745 0.043995851  20    raw     41
#> 37  -0.459773640 0.5704255 0.98920100 0.066561265   5  enter     42
#> 47  -0.413226649 0.1464387 0.99480482 0.001393738   6  enter     42
#> 62   1.500000000 0.1160586 0.82860539 0.035006098   7   exit     42
#> 178  1.192021907 0.3317948 0.83426745 0.043995851  20   exit     42
#> 38  -0.178189120 0.5704255 0.98920100 0.066561265   5  enter     43
#> 48   0.194186811 0.1464387 0.99480482 0.001393738   6  enter     43
#> 179  1.500000000 0.3317948 0.83426745 0.043995851  20   exit     43
#> 39   0.103395400 0.5704255 0.98920100 0.066561265   5  enter     44
#> 49   0.801600271 0.1464387 0.99480482 0.001393738   6  enter     44
#> 40   0.143621759 0.5704255 0.98920100 0.066561265   5 static     45
#> 50   0.888373622 0.1464387 0.99480482 0.001393738   6    raw     45
#> 41   0.143621759 0.5704255 0.98920100 0.066561265   5 static     46
#> 51   0.978984938 0.1464387 0.99480482 0.001393738   6   exit     46
#> 42   0.143621759 0.5704255 0.98920100 0.066561265   5 static     47
#> 52   1.409388685 0.1464387 0.99480482 0.001393738   6   exit     47
#> 43   0.143621759 0.5704255 0.98920100 0.066561265   5    raw     48
#> 53   1.500000000 0.1464387 0.99480482 0.001393738   6   exit     48
#> 44   0.821810880 0.5704255 0.98920100 0.066561265   5   exit     49
#> 45   1.500000000 0.5704255 0.98920100 0.066561265   5   exit     50