Transition in and out of events

This tweening function is a more powerful version of tween_appear(), with support for newer features such as enter/exits and tween phase identification. The tweener treats each row in the data as unique events in time, and creates frames with the correct events present at any given time.

tween_events(
  .data,
  ease,
  nframes,
  start,
  end = NULL,
  range = NULL,
  enter = NULL,
  exit = NULL,
  enter_length = 0,
  exit_length = 0
)

Arguments

.data

A data.frame with components at different stages

ease

The easing function to use. Either a single string or one for each column in the data set.

nframes

The number of frames to calculate for the tween

start, end

The start (and potential end) of the event encoded in the row, as unquoted expressions. Will be evaluated in the context of .data so can refer to columns in it. If end = NULL the event will be without extend and only visible in a single frame, unless enter and/or exit is given.

range

The range of time points to include in the tween. If NULL it will use the range of time

enter, exit

functions that calculate a start state for new observations that appear in to or an end state for observations that are not present in to. If NULL the new/old observations will not be part of the tween. The function gets a data.frame with either the start state of the exiting observations, or the end state of the entering observations and must return a modified version of that data.frame. See the Match, Enter, and Exit section for more information.

enter_length, exit_length

The lenght of the opening and closing transitions if enter and/or exit is given. Measured in the same units as time

Value

A data.frame with the same columns as .data along with .id giving the component id, .phase giving the state of each component in each frame, and .frame giving the frame membership of each row.

See also

Examples

d <- data.frame(
  x = runif(20),
  y = runif(20),
  time = runif(20),
  duration = runif(20, max = 0.1)
)
from_left <- function(x) {
  x$x <- -0.5
  x
}
to_right <- function(x) {
  x$x <- 1.5
  x
}

tween_events(d, 'cubic-in-out', 50, start = time, end = time + duration,
             enter = from_left, exit = to_right, enter_length = 0.1,
             exit_length = 0.05)
#>                x         y      time    duration .id .phase .frame
#> 57  -0.500000000 0.0522231 0.1670316 0.058207267   6  enter      1
#> 58  -0.454431274 0.0522231 0.1670316 0.058207267   6  enter      2
#> 151 -0.500000000 0.3888482 0.1820347 0.095984629  14  enter      2
#> 59  -0.135450189 0.0522231 0.1670316 0.058207267   6  enter      3
#> 152 -0.479324895 0.3888482 0.1820347 0.095984629  14  enter      3
#> 60   0.559472888 0.0522231 0.1670316 0.058207267   6  enter      4
#> 153 -0.334599161 0.3888482 0.1820347 0.095984629  14  enter      4
#> 61   0.878453973 0.0522231 0.1670316 0.058207267   6  enter      5
#> 154 -0.019303810 0.3888482 0.1820347 0.095984629  14  enter      5
#> 62   0.924022699 0.0522231 0.1670316 0.058207267   6 static      6
#> 155  0.125421924 0.3888482 0.1820347 0.095984629  14  enter      6
#> 63   0.924022699 0.0522231 0.1670316 0.058207267   6 static      7
#> 141 -0.500000000 0.3323523 0.2777635 0.007030151  13  enter      7
#> 156  0.146097029 0.3888482 0.1820347 0.095984629  14 static      7
#> 64   0.924022699 0.0522231 0.1670316 0.058207267   6 static      8
#> 142 -0.459497730 0.3323523 0.2777635 0.007030151  13  enter      8
#> 157  0.146097029 0.3888482 0.1820347 0.095984629  14 static      8
#> 65   0.924022699 0.0522231 0.1670316 0.058207267   6 static      9
#> 143 -0.175981840 0.3323523 0.2777635 0.007030151  13  enter      9
#> 158  0.146097029 0.3888482 0.1820347 0.095984629  14 static      9
#> 224 -0.500000000 0.5782739 0.3132521 0.082680291  20  enter      9
#> 66   0.924022699 0.0522231 0.1670316 0.058207267   6    raw     10
#> 144  0.441677777 0.3323523 0.2777635 0.007030151  13  enter     10
#> 159  0.146097029 0.3888482 0.1820347 0.095984629  14 static     10
#> 225 -0.466120022 0.5782739 0.3132521 0.082680291  20  enter     10
#> 67   1.212011350 0.0522231 0.1670316 0.058207267   6   exit     11
#> 69  -0.500000000 0.5001318 0.3468455 0.019439141   7  enter     11
#> 145  0.725193667 0.3323523 0.2777635 0.007030151  13  enter     11
#> 160  0.146097029 0.3888482 0.1820347 0.095984629  14 static     11
#> 189 -0.500000000 0.1978992 0.3454362 0.045283925  17  enter     11
#> 226 -0.228960173 0.5782739 0.3132521 0.082680291  20  enter     11
#> 23  -0.500000000 0.7131297 0.3658759 0.091476528   3  enter     12
#> 68   1.500000000 0.0522231 0.1670316 0.058207267   6   exit     12
#> 70  -0.456950272 0.5001318 0.3468455 0.019439141   7  enter     12
#> 79  -0.500000000 0.3549670 0.3734955 0.046710959   8  enter     12
#> 146  0.765695937 0.3323523 0.2777635 0.007030151  13 static     12
#> 161  0.146097029 0.3888482 0.1820347 0.095984629  14 static     12
#> 190 -0.477987332 0.1978992 0.3454362 0.045283925  17  enter     12
#> 227  0.287709497 0.5782739 0.3132521 0.082680291  20  enter     12
#> 24  -0.463336192 0.7131297 0.3658759 0.091476528   3  enter     13
#> 71  -0.155602175 0.5001318 0.3468455 0.019439141   7  enter     13
#> 80  -0.479156054 0.3549670 0.3734955 0.046710959   8  enter     13
#> 147  0.765695937 0.3323523 0.2777635 0.007030151  13    raw     13
#> 162  0.146097029 0.3888482 0.1820347 0.095984629  14    raw     13
#> 191 -0.323898658 0.1978992 0.3454362 0.045283925  17  enter     13
#> 228  0.524869345 0.5782739 0.3132521 0.082680291  20  enter     13
#> 25  -0.206689533 0.7131297 0.3658759 0.091476528   3  enter     14
#> 72   0.500906180 0.5001318 0.3468455 0.019439141   7  enter     14
#> 81  -0.333248435 0.3549670 0.3734955 0.046710959   8  enter     14
#> 148  0.874481724 0.3323523 0.2777635 0.007030151  13   exit     14
#> 163  0.823048515 0.3888482 0.1820347 0.095984629  14   exit     14
#> 192  0.011794526 0.1978992 0.3454362 0.045283925  17  enter     14
#> 229  0.558749324 0.5782739 0.3132521 0.082680291  20 static     14
#> 26   0.352433545 0.7131297 0.3658759 0.091476528   3  enter     15
#> 73   0.802254278 0.5001318 0.3468455 0.019439141   7  enter     15
#> 82   0.062786533 0.3549670 0.3734955 0.046710959   8  enter     15
#> 149  1.391214213 0.3323523 0.2777635 0.007030151  13   exit     15
#> 164  1.500000000 0.3888482 0.1820347 0.095984629  14   exit     15
#> 193  0.165883200 0.1978992 0.3454362 0.045283925  17  enter     15
#> 200 -0.500000000 0.9275100 0.4162140 0.067066822  18  enter     15
#> 230  0.558749324 0.5782739 0.3132521 0.082680291  20 static     15
#> 27   0.609080203 0.7131297 0.3658759 0.091476528   3  enter     16
#> 74   0.845304006 0.5001318 0.3468455 0.019439141   7 static     16
#> 83   0.458821501 0.3549670 0.3734955 0.046710959   8  enter     16
#> 150  1.500000000 0.3323523 0.2777635 0.007030151  13   exit     16
#> 194  0.187895868 0.1978992 0.3454362 0.045283925  17 static     16
#> 201 -0.462334009 0.9275100 0.4162140 0.067066822  18  enter     16
#> 231  0.558749324 0.5782739 0.3132521 0.082680291  20 static     16
#> 28   0.645744012 0.7131297 0.3658759 0.091476528   3 static     17
#> 75   0.845304006 0.5001318 0.3468455 0.019439141   7    raw     17
#> 84   0.604729121 0.3549670 0.3734955 0.046710959   8  enter     17
#> 91  -0.500000000 0.8689928 0.4613760 0.069255435   9  enter     17
#> 195  0.187895868 0.1978992 0.3454362 0.045283925  17 static     17
#> 202 -0.198672069 0.9275100 0.4162140 0.067066822  18  enter     17
#> 232  0.558749324 0.5782739 0.3132521 0.082680291  20 static     17
#> 29   0.645744012 0.7131297 0.3658759 0.091476528   3 static     18
#> 76   0.942296005 0.5001318 0.3468455 0.019439141   7   exit     18
#> 85   0.625573066 0.3549670 0.3734955 0.046710959   8 static     18
#> 92  -0.485540368 0.8689928 0.4613760 0.069255435   9  enter     18
#> 196  0.187895868 0.1978992 0.3454362 0.045283925  17 static     18
#> 203  0.375734298 0.9275100 0.4162140 0.067066822  18  enter     18
#> 233  0.558749324 0.5782739 0.3132521 0.082680291  20 static     18
#> 30   0.645744012 0.7131297 0.3658759 0.091476528   3 static     19
#> 77   1.403008001 0.5001318 0.3468455 0.019439141   7   exit     19
#> 86   0.625573066 0.3549670 0.3734955 0.046710959   8 static     19
#> 93  -0.384322941 0.8689928 0.4613760 0.069255435   9  enter     19
#> 165 -0.500000000 0.6840792 0.5019538 0.051301926  15  enter     19
#> 197  0.187895868 0.1978992 0.3454362 0.045283925  17    raw     19
#> 204  0.639396237 0.9275100 0.4162140 0.067066822  18  enter     19
#> 234  0.558749324 0.5782739 0.3132521 0.082680291  20    raw     19
#> 31   0.645744012 0.7131297 0.3658759 0.091476528   3 static     20
#> 78   1.500000000 0.5001318 0.3468455 0.019439141   7   exit     20
#> 87   0.625573066 0.3549670 0.3734955 0.046710959   8    raw     20
#> 94  -0.109589927 0.8689928 0.4613760 0.069255435   9  enter     20
#> 166 -0.482341268 0.6840792 0.5019538 0.051301926  15  enter     20
#> 198  0.843947934 0.1978992 0.3454362 0.045283925  17   exit     20
#> 205  0.677062229 0.9275100 0.4162140 0.067066822  18 static     20
#> 235  0.698193868 0.5782739 0.3132521 0.082680291  20   exit     20
#> 32   0.645744012 0.7131297 0.3658759 0.091476528   3 static     21
#> 88   0.755117797 0.3549670 0.3734955 0.046710959   8   exit     21
#> 95   0.165143087 0.8689928 0.4613760 0.069255435   9  enter     21
#> 167 -0.358730146 0.6840792 0.5019538 0.051301926  15  enter     21
#> 199  1.500000000 0.1978992 0.3454362 0.045283925  17   exit     21
#> 206  0.677062229 0.9275100 0.4162140 0.067066822  18 static     21
#> 236  1.360555455 0.5782739 0.3132521 0.082680291  20   exit     21
#> 33   0.645744012 0.7131297 0.3658759 0.091476528   3    raw     22
#> 89   1.370455269 0.3549670 0.3734955 0.046710959   8   exit     22
#> 96   0.266360513 0.8689928 0.4613760 0.069255435   9  enter     22
#> 168 -0.023214242 0.6840792 0.5019538 0.051301926  15  enter     22
#> 207  0.677062229 0.9275100 0.4162140 0.067066822  18 static     22
#> 237  1.500000000 0.5782739 0.3132521 0.082680291  20   exit     22
#> 34   0.772300454 0.7131297 0.3658759 0.091476528   3   exit     23
#> 46  -0.500000000 0.3716502 0.5750621 0.033851793   5  enter     23
#> 90   1.500000000 0.3549670 0.3734955 0.046710959   8   exit     23
#> 97   0.280820145 0.8689928 0.4613760 0.069255435   9 static     23
#> 169  0.312301661 0.6840792 0.5019538 0.051301926  15  enter     23
#> 208  0.677062229 0.9275100 0.4162140 0.067066822  18 static     23
#> 35   1.373443557 0.7131297 0.3658759 0.091476528   3   exit     24
#> 47  -0.481745357 0.3716502 0.5750621 0.033851793   5  enter     24
#> 98   0.280820145 0.8689928 0.4613760 0.069255435   9 static     24
#> 170  0.435912784 0.6840792 0.5019538 0.051301926  15  enter     24
#> 209  0.677062229 0.9275100 0.4162140 0.067066822  18    raw     24
#> 1   -0.500000000 0.5705489 0.6126578 0.034559183   1  enter     25
#> 36   1.500000000 0.7131297 0.3658759 0.091476528   3   exit     25
#> 48  -0.353962854 0.3716502 0.5750621 0.033851793   5  enter     25
#> 99   0.280820145 0.8689928 0.4613760 0.069255435   9 static     25
#> 171  0.453571515 0.6840792 0.5019538 0.051301926  15 static     25
#> 210  1.088531114 0.9275100 0.4162140 0.067066822  18   exit     25
#> 2   -0.479133321 0.5705489 0.6126578 0.034559183   1  enter     26
#> 49  -0.007124632 0.3716502 0.5750621 0.033851793   5  enter     26
#> 100  0.280820145 0.8689928 0.4613760 0.069255435   9    raw     26
#> 172  0.453571515 0.6840792 0.5019538 0.051301926  15 static     26
#> 211  1.500000000 0.9275100 0.4162140 0.067066822  18   exit     26
#> 3   -0.333066564 0.5705489 0.6126578 0.034559183   1  enter     27
#> 37  -0.500000000 0.4675865 0.6382104 0.010875044   4  enter     27
#> 50   0.339713590 0.3716502 0.5750621 0.033851793   5  enter     27
#> 101  0.461439383 0.8689928 0.4613760 0.069255435   9   exit     27
#> 173  0.453571515 0.6840792 0.5019538 0.051301926  15 static     27
#> 4    0.063400346 0.5705489 0.6126578 0.034559183   1  enter     28
#> 38  -0.463610168 0.4675865 0.6382104 0.010875044   4  enter     28
#> 51   0.467496093 0.3716502 0.5750621 0.033851793   5  enter     28
#> 102  1.319380762 0.8689928 0.4613760 0.069255435   9   exit     28
#> 174  0.453571515 0.6840792 0.5019538 0.051301926  15    raw     28
#> 5    0.459867256 0.5705489 0.6126578 0.034559183   1  enter     29
#> 39  -0.208881341 0.4675865 0.6382104 0.010875044   4  enter     29
#> 52   0.485750736 0.3716502 0.5750621 0.033851793   5 static     29
#> 103  1.500000000 0.8689928 0.4613760 0.069255435   9   exit     29
#> 175  0.976785758 0.6840792 0.5019538 0.051301926  15   exit     29
#> 6    0.605934012 0.5705489 0.6126578 0.034559183   1  enter     30
#> 40   0.346063604 0.4675865 0.6382104 0.010875044   4  enter     30
#> 53   0.485750736 0.3716502 0.5750621 0.033851793   5 static     30
#> 176  1.500000000 0.6840792 0.5019538 0.051301926  15   exit     30
#> 7    0.626800692 0.5705489 0.6126578 0.034559183   1 static     31
#> 41   0.600792431 0.4675865 0.6382104 0.010875044   4  enter     31
#> 54   0.485750736 0.3716502 0.5750621 0.033851793   5    raw     31
#> 8    0.626800692 0.5705489 0.6126578 0.034559183   1 static     32
#> 12  -0.500000000 0.3866962 0.7289341 0.039013092   2  enter     32
#> 42   0.637182263 0.4675865 0.6382104 0.010875044   4 static     32
#> 55   0.992875368 0.3716502 0.5750621 0.033851793   5   exit     32
#> 9    0.626800692 0.5705489 0.6126578 0.034559183   1    raw     33
#> 13  -0.467800222 0.3866962 0.7289341 0.039013092   2  enter     33
#> 43   0.637182263 0.4675865 0.6382104 0.010875044   4    raw     33
#> 56   1.500000000 0.3716502 0.5750621 0.033851793   5   exit     33
#> 126 -0.500000000 0.6922982 0.7584476 0.094643861  12  enter     33
#> 212 -0.500000000 0.1046808 0.7509264 0.062706545  19  enter     33
#> 10   1.063400346 0.5705489 0.6126578 0.034559183   1   exit     34
#> 14  -0.242401776 0.3866962 0.7289341 0.039013092   2  enter     34
#> 44   1.068591131 0.4675865 0.6382104 0.010875044   4   exit     34
#> 127 -0.474178826 0.6922982 0.7584476 0.094643861  12  enter     34
#> 213 -0.467530514 0.1046808 0.7509264 0.062706545  19  enter     34
#> 11   1.500000000 0.5705489 0.6126578 0.034559183   1   exit     35
#> 15   0.248644839 0.3866962 0.7289341 0.039013092   2  enter     35
#> 45   1.500000000 0.4675865 0.6382104 0.010875044   4   exit     35
#> 128 -0.293430610 0.6922982 0.7584476 0.094643861  12  enter     35
#> 177 -0.500000000 0.9527881 0.7888264 0.046727419  16  enter     35
#> 214 -0.240244110 0.1046808 0.7509264 0.062706545  19  enter     35
#> 16   0.474043285 0.3866962 0.7289341 0.039013092   2  enter     36
#> 129  0.197171692 0.6922982 0.7584476 0.094643861  12  enter     36
#> 178 -0.470205622 0.9527881 0.7888264 0.046727419  16  enter     36
#> 215  0.254915556 0.1046808 0.7509264 0.062706545  19  enter     36
#> 17   0.506243063 0.3866962 0.7289341 0.039013092   2 static     37
#> 104 -0.500000000 0.8595477 0.8347282 0.014703074  10  enter     37
#> 130  0.687773994 0.6922982 0.7584476 0.094643861  12  enter     37
#> 179 -0.261644973 0.9527881 0.7888264 0.046727419  16  enter     37
#> 216  0.482201960 0.1046808 0.7509264 0.062706545  19  enter     37
#> 18   0.506243063 0.3866962 0.7289341 0.039013092   2 static     38
#> 105 -0.485517850 0.8595477 0.8347282 0.014703074  10  enter     38
#> 131  0.868522211 0.6922982 0.7584476 0.094643861  12  enter     38
#> 180  0.192719297 0.9527881 0.7888264 0.046727419  16  enter     38
#> 217  0.514671446 0.1046808 0.7509264 0.062706545  19 static     38
#> 19   0.506243063 0.3866962 0.7289341 0.039013092   2    raw     39
#> 106 -0.384142801 0.8595477 0.8347282 0.014703074  10  enter     39
#> 114 -0.500000000 0.6701551 0.8624805 0.052627436  11  enter     39
#> 132  0.894343385 0.6922982 0.7584476 0.094643861  12 static     39
#> 181  0.401279945 0.9527881 0.7888264 0.046727419  16  enter     39
#> 218  0.514671446 0.1046808 0.7509264 0.062706545  19 static     39
#> 20   0.653466313 0.3866962 0.7289341 0.039013092   2   exit     40
#> 107 -0.108981952 0.8595477 0.8347282 0.014703074  10  enter     40
#> 115 -0.457835988 0.6701551 0.8624805 0.052627436  11  enter     40
#> 133  0.894343385 0.6922982 0.7584476 0.094643861  12 static     40
#> 182  0.431074324 0.9527881 0.7888264 0.046727419  16 static     40
#> 219  0.514671446 0.1046808 0.7509264 0.062706545  19 static     40
#> 21   1.352776750 0.3866962 0.7289341 0.039013092   2   exit     41
#> 108  0.166178897 0.8595477 0.8347282 0.014703074  10  enter     41
#> 116 -0.162687906 0.6701551 0.8624805 0.052627436  11  enter     41
#> 134  0.894343385 0.6922982 0.7584476 0.094643861  12 static     41
#> 183  0.431074324 0.9527881 0.7888264 0.046727419  16 static     41
#> 220  0.514671446 0.1046808 0.7509264 0.062706545  19 static     41
#> 22   1.500000000 0.3866962 0.7289341 0.039013092   2   exit     42
#> 109  0.267553946 0.8595477 0.8347282 0.014703074  10  enter     42
#> 117  0.480313274 0.6701551 0.8624805 0.052627436  11  enter     42
#> 135  0.894343385 0.6922982 0.7584476 0.094643861  12 static     42
#> 184  0.431074324 0.9527881 0.7888264 0.046727419  16 static     42
#> 221  0.514671446 0.1046808 0.7509264 0.062706545  19    raw     42
#> 110  0.282036096 0.8595477 0.8347282 0.014703074  10 static     43
#> 118  0.775461356 0.6701551 0.8624805 0.052627436  11  enter     43
#> 136  0.894343385 0.6922982 0.7584476 0.094643861  12 static     43
#> 185  0.431074324 0.9527881 0.7888264 0.046727419  16    raw     43
#> 222  1.007335723 0.1046808 0.7509264 0.062706545  19   exit     43
#> 111  0.282036096 0.8595477 0.8347282 0.014703074  10    raw     44
#> 119  0.817625368 0.6701551 0.8624805 0.052627436  11 static     44
#> 137  0.894343385 0.6922982 0.7584476 0.094643861  12    raw     44
#> 186  0.589433683 0.9527881 0.7888264 0.046727419  16   exit     44
#> 223  1.500000000 0.1046808 0.7509264 0.062706545  19   exit     44
#> 112  0.891018048 0.8595477 0.8347282 0.014703074  10   exit     45
#> 120  0.817625368 0.6701551 0.8624805 0.052627436  11 static     45
#> 138  0.984070291 0.6922982 0.7584476 0.094643861  12   exit     45
#> 187  1.341640641 0.9527881 0.7888264 0.046727419  16   exit     45
#> 113  1.500000000 0.8595477 0.8347282 0.014703074  10   exit     46
#> 121  0.817625368 0.6701551 0.8624805 0.052627436  11 static     46
#> 139  1.410273094 0.6922982 0.7584476 0.094643861  12   exit     46
#> 188  1.500000000 0.9527881 0.7888264 0.046727419  16   exit     46
#> 122  0.817625368 0.6701551 0.8624805 0.052627436  11    raw     47
#> 140  1.500000000 0.6922982 0.7584476 0.094643861  12   exit     47
#> 123  0.918717906 0.6701551 0.8624805 0.052627436  11   exit     48
#> 124  1.398907462 0.6701551 0.8624805 0.052627436  11   exit     49
#> 125  1.500000000 0.6701551 0.8624805 0.052627436  11   exit     50