Transition in and out of events

This tweening function is a more powerful version of tween_appear(), with support for newer features such as enter/exits and tween phase identification. The tweener treats each row in the data as unique events in time, and creates frames with the correct events present at any given time.

tween_events(
  .data,
  ease,
  nframes,
  start,
  end = NULL,
  range = NULL,
  enter = NULL,
  exit = NULL,
  enter_length = 0,
  exit_length = 0
)

Arguments

.data

A data.frame with components at different stages

ease

The easing function to use. Either a single string or one for each column in the data set.

nframes

The number of frames to calculate for the tween

start, end

The start (and potential end) of the event encoded in the row, as unquoted expressions. Will be evaluated in the context of .data so can refer to columns in it. If end = NULL the event will be without extend and only visible in a single frame, unless enter and/or exit is given.

range

The range of time points to include in the tween. If NULL it will use the range of time

enter

functions that calculate a start state for new observations that appear in to or an end state for observations that are not present in to. If NULL the new/old observations will not be part of the tween. The function gets a data.frame with either the start state of the exiting observations, or the end state of the entering observations and must return a modified version of that data.frame. See the Match, Enter, and Exit section for more information.

exit

functions that calculate a start state for new observations that appear in to or an end state for observations that are not present in to. If NULL the new/old observations will not be part of the tween. The function gets a data.frame with either the start state of the exiting observations, or the end state of the entering observations and must return a modified version of that data.frame. See the Match, Enter, and Exit section for more information.

enter_length

The lenght of the opening and closing transitions if enter and/or exit is given. Measured in the same units as time

exit_length

The lenght of the opening and closing transitions if enter and/or exit is given. Measured in the same units as time

Value

A data.frame with the same columns as .data along with .id giving the component id, .phase giving the state of each component in each frame, and .frame giving the frame membership of each row.

See also

Examples

d <- data.frame(
  x = runif(20),
  y = runif(20),
  time = runif(20),
  duration = runif(20, max = 0.1)
)
from_left <- function(x) {
  x$x <- -0.5
  x
}
to_right <- function(x) {
  x$x <- 1.5
  x
}

tween_events(d, 'cubic-in-out', 50, start = time, end = time + duration,
             enter = from_left, exit = to_right, enter_length = 0.1,
             exit_length = 0.05)
#>                x          y       time    duration .id .phase .frame
#> 96  -0.500000000 0.54524506 0.09245657 0.007491694  10  enter      1
#> 97  -0.454133352 0.54524506 0.09245657 0.007491694  10  enter      2
#> 169 -0.500000000 0.39613608 0.11701265 0.014395456  17  enter      2
#> 98  -0.133066819 0.54524506 0.09245657 0.007491694  10  enter      3
#> 170 -0.483848844 0.39613608 0.11701265 0.014395456  17  enter      3
#> 99   0.566399557 0.54524506 0.09245657 0.007491694  10  enter      4
#> 146 -0.500000000 0.82182576 0.15797363 0.053700030  15  enter      4
#> 171 -0.370790754 0.39613608 0.11701265 0.014395456  17  enter      4
#> 100  0.887466090 0.54524506 0.09245657 0.007491694  10  enter      5
#> 147 -0.459114613 0.82182576 0.15797363 0.053700030  15  enter      5
#> 172 -0.124485629 0.39613608 0.11701265 0.014395456  17  enter      5
#> 101  0.933332738 0.54524506 0.09245657 0.007491694  10    raw      6
#> 148 -0.172916901 0.82182576 0.15797363 0.053700030  15  enter      6
#> 173 -0.011427538 0.39613608 0.11701265 0.014395456  17  enter      6
#> 102  1.216666369 0.54524506 0.09245657 0.007491694  10   exit      7
#> 149  0.450585256 0.82182576 0.15797363 0.053700030  15  enter      7
#> 174  0.004723618 0.39613608 0.11701265 0.014395456  17 static      7
#> 85  -0.500000000 0.08186153 0.24870477 0.057198536   9  enter      8
#> 103  1.500000000 0.54524506 0.09245657 0.007491694  10   exit      8
#> 150  0.736782967 0.82182576 0.15797363 0.053700030  15  enter      8
#> 175  0.004723618 0.39613608 0.11701265 0.014395456  17    raw      8
#> 86  -0.457043743 0.08186153 0.24870477 0.057198536   9  enter      9
#> 151  0.777668354 0.82182576 0.15797363 0.053700030  15 static      9
#> 176  0.752361809 0.39613608 0.11701265 0.014395456  17   exit      9
#> 87  -0.156349945 0.08186153 0.24870477 0.057198536   9  enter     10
#> 152  0.777668354 0.82182576 0.15797363 0.053700030  15 static     10
#> 177  1.500000000 0.39613608 0.11701265 0.014395456  17   exit     10
#> 88   0.498732972 0.08186153 0.24870477 0.057198536   9  enter     11
#> 153  0.777668354 0.82182576 0.15797363 0.053700030  15    raw     11
#> 89   0.799426770 0.08186153 0.24870477 0.057198536   9  enter     12
#> 154  0.884680450 0.82182576 0.15797363 0.053700030  15   exit     12
#> 64  -0.500000000 0.12500265 0.33837448 0.051938470   7  enter     13
#> 90   0.842383027 0.08186153 0.24870477 0.057198536   9 static     13
#> 155  1.392987904 0.82182576 0.15797363 0.053700030  15   exit     13
#> 65  -0.435535450 0.12500265 0.33837448 0.051938470   7  enter     14
#> 91   0.842383027 0.08186153 0.24870477 0.057198536   9 static     14
#> 156  1.500000000 0.82182576 0.15797363 0.053700030  15   exit     14
#> 66   0.015716398 0.12500265 0.33837448 0.051938470   7  enter     15
#> 92   0.842383027 0.08186153 0.24870477 0.057198536   9 static     15
#> 67   0.466968247 0.12500265 0.33837448 0.051938470   7  enter     16
#> 93   0.842383027 0.08186153 0.24870477 0.057198536   9    raw     16
#> 68   0.531432797 0.12500265 0.33837448 0.051938470   7 static     17
#> 74  -0.500000000 0.91625194 0.43373561 0.063994016   8  enter     17
#> 94   1.171191513 0.08186153 0.24870477 0.057198536   9   exit     17
#> 178 -0.500000000 0.84463410 0.42371958 0.071639226  18  enter     17
#> 69   0.531432797 0.12500265 0.33837448 0.051938470   7 static     18
#> 75  -0.465012856 0.91625194 0.43373561 0.063994016   8  enter     18
#> 95   1.500000000 0.08186153 0.24870477 0.057198536   9   exit     18
#> 157 -0.500000000 0.65744056 0.46189914 0.072038273  16  enter     18
#> 179 -0.406726710 0.84463410 0.42371958 0.071639226  18  enter     18
#> 22  -0.500000000 0.31126828 0.47436285 0.042816520   3  enter     19
#> 70   0.531432797 0.12500265 0.33837448 0.051938470   7 static     19
#> 76  -0.220102844 0.91625194 0.43373561 0.063994016   8  enter     19
#> 158 -0.472851048 0.65744056 0.46189914 0.072038273  16  enter     19
#> 180  0.246186317 0.84463410 0.42371958 0.071639226  18  enter     19
#> 1   -0.500000000 0.24981905 0.50313321 0.066587885   1  enter     20
#> 23  -0.465172313 0.31126828 0.47436285 0.042816520   3  enter     20
#> 71   0.531432797 0.12500265 0.33837448 0.051938470   7    raw     20
#> 77   0.313451109 0.91625194 0.43373561 0.063994016   8  enter     20
#> 159 -0.282808385 0.65744056 0.46189914 0.072038273  16  enter     20
#> 181  0.899099345 0.84463410 0.42371958 0.071639226  18  enter     20
#> 2   -0.481802470 0.24981905 0.50313321 0.066587885   1  enter     21
#> 24  -0.221378505 0.31126828 0.47436285 0.042816520   3  enter     21
#> 72   1.015716398 0.12500265 0.33837448 0.051938470   7   exit     21
#> 78   0.558361120 0.91625194 0.43373561 0.063994016   8  enter     21
#> 160  0.131213131 0.65744056 0.46189914 0.072038273  16  enter     21
#> 182  0.992372634 0.84463410 0.42371958 0.071639226  18 static     21
#> 3   -0.354419757 0.24981905 0.50313321 0.066587885   1  enter     22
#> 25   0.309743720 0.31126828 0.47436285 0.042816520   3  enter     22
#> 73   1.500000000 0.12500265 0.33837448 0.051938470   7   exit     22
#> 79   0.593348264 0.91625194 0.43373561 0.063994016   8 static     22
#> 161  0.321255795 0.65744056 0.46189914 0.072038273  16  enter     22
#> 183  0.992372634 0.84463410 0.42371958 0.071639226  18 static     22
#> 4   -0.076907418 0.24981905 0.50313321 0.066587885   1  enter     23
#> 26   0.553537528 0.31126828 0.47436285 0.042816520   3  enter     23
#> 80   0.593348264 0.91625194 0.43373561 0.063994016   8 static     23
#> 162  0.348404747 0.65744056 0.46189914 0.072038273  16 static     23
#> 184  0.992372634 0.84463410 0.42371958 0.071639226  18 static     23
#> 5    0.050475295 0.24981905 0.50313321 0.066587885   1  enter     24
#> 27   0.588365215 0.31126828 0.47436285 0.042816520   3 static     24
#> 81   0.593348264 0.91625194 0.43373561 0.063994016   8 static     24
#> 163  0.348404747 0.65744056 0.46189914 0.072038273  16 static     24
#> 185  0.992372634 0.84463410 0.42371958 0.071639226  18 static     24
#> 6    0.068672825 0.24981905 0.50313321 0.066587885   1 static     25
#> 28   0.588365215 0.31126828 0.47436285 0.042816520   3 static     25
#> 82   0.593348264 0.91625194 0.43373561 0.063994016   8    raw     25
#> 164  0.348404747 0.65744056 0.46189914 0.072038273  16 static     25
#> 186  0.992372634 0.84463410 0.42371958 0.071639226  18    raw     25
#> 7    0.068672825 0.24981905 0.50313321 0.066587885   1 static     26
#> 29   0.588365215 0.31126828 0.47436285 0.042816520   3    raw     26
#> 83   1.046674132 0.91625194 0.43373561 0.063994016   8   exit     26
#> 165  0.348404747 0.65744056 0.46189914 0.072038273  16    raw     26
#> 187  1.246186317 0.84463410 0.42371958 0.071639226  18   exit     26
#> 8    0.068672825 0.24981905 0.50313321 0.066587885   1 static     27
#> 30   1.044182607 0.31126828 0.47436285 0.042816520   3   exit     27
#> 84   1.500000000 0.91625194 0.43373561 0.063994016   8   exit     27
#> 166  0.519011451 0.65744056 0.46189914 0.072038273  16   exit     27
#> 188  1.500000000 0.84463410 0.42371958 0.071639226  18   exit     27
#> 198 -0.500000000 0.75936890 0.64222981 0.055785868  20  enter     27
#> 9    0.068672825 0.24981905 0.50313321 0.066587885   1    raw     28
#> 31   1.500000000 0.31126828 0.47436285 0.042816520   3   exit     28
#> 167  1.329393296 0.65744056 0.46189914 0.072038273  16   exit     28
#> 199 -0.430056499 0.75936890 0.64222981 0.055785868  20  enter     28
#> 10   0.784336413 0.24981905 0.50313321 0.066587885   1   exit     29
#> 168  1.500000000 0.65744056 0.46189914 0.072038273  16   exit     29
#> 200  0.059548005 0.75936890 0.64222981 0.055785868  20  enter     29
#> 11   1.500000000 0.24981905 0.50313321 0.066587885   1   exit     30
#> 201  0.549152509 0.75936890 0.64222981 0.055785868  20  enter     30
#> 136 -0.500000000 0.09406659 0.73277134 0.049931901  14  enter     31
#> 189 -0.500000000 0.54652374 0.73339749 0.019173902  19  enter     31
#> 202  0.619096010 0.75936890 0.64222981 0.055785868  20 static     31
#> 32  -0.500000000 0.22517505 0.74397948 0.050543083   4  enter     32
#> 137 -0.480714869 0.09406659 0.73277134 0.049931901  14  enter     32
#> 190 -0.470817896 0.54652374 0.73339749 0.019173902  19  enter     32
#> 203  0.619096010 0.75936890 0.64222981 0.055785868  20 static     32
#> 33  -0.464702162 0.22517505 0.74397948 0.050543083   4  enter     33
#> 138 -0.345718949 0.09406659 0.73277134 0.049931901  14  enter     33
#> 191 -0.266543166 0.54652374 0.73339749 0.019173902  19  enter     33
#> 204  0.619096010 0.75936890 0.64222981 0.055785868  20 static     33
#> 34  -0.217617293 0.22517505 0.74397948 0.050543083   4  enter     34
#> 104 -0.500000000 0.91269112 0.80274723 0.022691750  11  enter     34
#> 139 -0.051620696 0.09406659 0.73277134 0.049931901  14  enter     34
#> 192  0.178483923 0.54652374 0.73339749 0.019173902  19  enter     34
#> 205  0.619096010 0.75936890 0.64222981 0.055785868  20    raw     34
#> 35   0.029467575 0.22517505 0.74397948 0.050543083   4  enter     35
#> 53  -0.500000000 0.53436505 0.81980757 0.059172467   6  enter     35
#> 105 -0.474049852 0.91269112 0.80274723 0.022691750  11  enter     35
#> 140  0.083375223 0.09406659 0.73277134 0.049931901  14  enter     35
#> 193  0.382758653 0.54652374 0.73339749 0.019173902  19  enter     35
#> 206  1.059548005 0.75936890 0.64222981 0.055785868  20   exit     35
#> 36   0.064765414 0.22517505 0.74397948 0.050543083   4 static     36
#> 54  -0.460076817 0.53436505 0.81980757 0.059172467   6  enter     36
#> 106 -0.292398818 0.91269112 0.80274723 0.022691750  11  enter     36
#> 141  0.102660355 0.09406659 0.73277134 0.049931901  14 static     36
#> 194  0.411940757 0.54652374 0.73339749 0.019173902  19 static     36
#> 207  1.500000000 0.75936890 0.64222981 0.055785868  20   exit     36
#> 12  -0.500000000 0.99024546 0.86340076 0.038464226   2  enter     37
#> 37   0.064765414 0.22517505 0.74397948 0.050543083   4 static     37
#> 55  -0.180614536 0.53436505 0.81980757 0.059172467   6  enter     37
#> 107  0.103340934 0.91269112 0.80274723 0.022691750  11  enter     37
#> 113 -0.500000000 0.21018664 0.86342903 0.089697751  12  enter     37
#> 142  0.102660355 0.09406659 0.73277134 0.049931901  14 static     37
#> 195  0.411940757 0.54652374 0.73339749 0.019173902  19    raw     37
#> 13  -0.455802393 0.99024546 0.86340076 0.038464226   2  enter     38
#> 38   0.064765414 0.22517505 0.74397948 0.050543083   4 static     38
#> 56   0.428214004 0.53436505 0.81980757 0.059172467   6  enter     38
#> 108  0.284991968 0.91269112 0.80274723 0.022691750  11  enter     38
#> 114 -0.459137278 0.21018664 0.86342903 0.089697751  12  enter     38
#> 143  0.102660355 0.09406659 0.73277134 0.049931901  14    raw     38
#> 196  0.955970379 0.54652374 0.73339749 0.019173902  19   exit     38
#> 14  -0.146419145 0.99024546 0.86340076 0.038464226   2  enter     39
#> 39   0.064765414 0.22517505 0.74397948 0.050543083   4    raw     39
#> 42  -0.500000000 0.42936992 0.90370510 0.058994069   5  enter     39
#> 57   0.707676284 0.53436505 0.81980757 0.059172467   6  enter     39
#> 109  0.310942115 0.91269112 0.80274723 0.022691750  11 static     39
#> 115 -0.173098221 0.21018664 0.86342903 0.089697751  12  enter     39
#> 144  0.801330177 0.09406659 0.73277134 0.049931901  14   exit     39
#> 197  1.500000000 0.54652374 0.73339749 0.019173902  19   exit     39
#> 15   0.527594360 0.99024546 0.86340076 0.038464226   2  enter     40
#> 40   0.782382707 0.22517505 0.74397948 0.050543083   4   exit     40
#> 43  -0.477310146 0.42936992 0.90370510 0.058994069   5  enter     40
#> 58   0.747599467 0.53436505 0.81980757 0.059172467   6 static     40
#> 110  0.310942115 0.91269112 0.80274723 0.022691750  11    raw     40
#> 116  0.450058295 0.21018664 0.86342903 0.089697751  12  enter     40
#> 125 -0.500000000 0.88565723 0.92406699 0.064733850  13  enter     40
#> 145  1.500000000 0.09406659 0.73277134 0.049931901  14   exit     40
#> 16   0.836977608 0.99024546 0.86340076 0.038464226   2  enter     41
#> 41   1.500000000 0.22517505 0.74397948 0.050543083   4   exit     41
#> 44  -0.318481165 0.42936992 0.90370510 0.058994069   5  enter     41
#> 59   0.747599467 0.53436505 0.81980757 0.059172467   6 static     41
#> 111  0.905471058 0.91269112 0.80274723 0.022691750  11   exit     41
#> 117  0.736097352 0.21018664 0.86342903 0.089697751  12  enter     41
#> 126 -0.469940315 0.88565723 0.92406699 0.064733850  13  enter     41
#> 17   0.881175215 0.99024546 0.86340076 0.038464226   2 static     42
#> 45   0.027539116 0.42936992 0.90370510 0.058994069   5  enter     42
#> 60   0.747599467 0.53436505 0.81980757 0.059172467   6 static     42
#> 112  1.500000000 0.91269112 0.80274723 0.022691750  11   exit     42
#> 118  0.776960074 0.21018664 0.86342903 0.089697751  12 static     42
#> 127 -0.259522520 0.88565723 0.92406699 0.064733850  13  enter     42
#> 18   0.881175215 0.99024546 0.86340076 0.038464226   2 static     43
#> 46   0.186368097 0.42936992 0.90370510 0.058994069   5  enter     43
#> 61   0.747599467 0.53436505 0.81980757 0.059172467   6    raw     43
#> 119  0.776960074 0.21018664 0.86342903 0.089697751  12 static     43
#> 128  0.198887677 0.88565723 0.92406699 0.064733850  13  enter     43
#> 19   0.881175215 0.99024546 0.86340076 0.038464226   2    raw     44
#> 47   0.209057951 0.42936992 0.90370510 0.058994069   5 static     44
#> 62   1.123799734 0.53436505 0.81980757 0.059172467   6   exit     44
#> 120  0.776960074 0.21018664 0.86342903 0.089697751  12 static     44
#> 129  0.409305472 0.88565723 0.92406699 0.064733850  13  enter     44
#> 20   1.190587607 0.99024546 0.86340076 0.038464226   2   exit     45
#> 48   0.209057951 0.42936992 0.90370510 0.058994069   5 static     45
#> 63   1.500000000 0.53436505 0.81980757 0.059172467   6   exit     45
#> 121  0.776960074 0.21018664 0.86342903 0.089697751  12 static     45
#> 130  0.439365157 0.88565723 0.92406699 0.064733850  13 static     45
#> 21   1.500000000 0.99024546 0.86340076 0.038464226   2   exit     46
#> 49   0.209057951 0.42936992 0.90370510 0.058994069   5    raw     46
#> 122  0.776960074 0.21018664 0.86342903 0.089697751  12    raw     46
#> 131  0.439365157 0.88565723 0.92406699 0.064733850  13 static     46
#> 50   0.400308625 0.42936992 0.90370510 0.058994069   5   exit     47
#> 123  1.138480037 0.21018664 0.86342903 0.089697751  12   exit     47
#> 132  0.439365157 0.88565723 0.92406699 0.064733850  13 static     47
#> 51   1.308749326 0.42936992 0.90370510 0.058994069   5   exit     48
#> 124  1.500000000 0.21018664 0.86342903 0.089697751  12   exit     48
#> 133  0.439365157 0.88565723 0.92406699 0.064733850  13    raw     48
#> 52   1.500000000 0.42936992 0.90370510 0.058994069   5   exit     49
#> 134  0.969682579 0.88565723 0.92406699 0.064733850  13   exit     49
#> 135  1.500000000 0.88565723 0.92406699 0.064733850  13   exit     50