Transition in and out of events

This tweening function is a more powerful version of tween_appear(), with support for newer features such as enter/exits and tween phase identification. The tweener treats each row in the data as unique events in time, and creates frames with the correct events present at any given time.

tween_events(
  .data,
  ease,
  nframes,
  start,
  end = NULL,
  range = NULL,
  enter = NULL,
  exit = NULL,
  enter_length = 0,
  exit_length = 0
)

Arguments

.data

A data.frame with components at different stages

ease

The easing function to use. Either a single string or one for each column in the data set.

nframes

The number of frames to calculate for the tween

start, end

The start (and potential end) of the event encoded in the row, as unquoted expressions. Will be evaluated in the context of .data so can refer to columns in it. If end = NULL the event will be without extend and only visible in a single frame, unless enter and/or exit is given.

range

The range of time points to include in the tween. If NULL it will use the range of time

enter, exit

functions that calculate a start state for new observations that appear in to or an end state for observations that are not present in to. If NULL the new/old observations will not be part of the tween. The function gets a data.frame with either the start state of the exiting observations, or the end state of the entering observations and must return a modified version of that data.frame. See the Match, Enter, and Exit section for more information.

enter_length, exit_length

The lenght of the opening and closing transitions if enter and/or exit is given. Measured in the same units as time

Value

A data.frame with the same columns as .data along with .id giving the component id, .phase giving the state of each component in each frame, and .frame giving the frame membership of each row.

See also

Examples

d <- data.frame(
  x = runif(20),
  y = runif(20),
  time = runif(20),
  duration = runif(20, max = 0.1)
)
from_left <- function(x) {
  x$x <- -0.5
  x
}
to_right <- function(x) {
  x$x <- 1.5
  x
}

tween_events(d, 'cubic-in-out', 50, start = time, end = time + duration,
             enter = from_left, exit = to_right, enter_length = 0.1,
             exit_length = 0.05)
#>                x         y       time    duration .id .phase .frame
#> 82  -0.500000000 0.9631831 0.04911027 0.025281717   9  enter      1
#> 83  -0.465871646 0.9631831 0.04911027 0.025281717   9  enter      2
#> 128 -0.500000000 0.9936744 0.07602862 0.092675535  14  enter      2
#> 84  -0.226973170 0.9631831 0.04911027 0.025281717   9  enter      3
#> 129 -0.454299309 0.9936744 0.07602862 0.092675535  14  enter      3
#> 85   0.293484225 0.9631831 0.04911027 0.025281717   9  enter      4
#> 130 -0.134394468 0.9936744 0.07602862 0.092675535  14  enter      4
#> 140 -0.500000000 0.4537346 0.12196996 0.082984576  15  enter      4
#> 86   0.532382702 0.9631831 0.04911027 0.025281717   9  enter      5
#> 131  0.562541076 0.9936744 0.07602862 0.092675535  14  enter      5
#> 141 -0.470154423 0.4537346 0.12196996 0.082984576  15  enter      5
#> 33  -0.500000000 0.0378566 0.16536604 0.070887312   4  enter      6
#> 87   0.566511055 0.9631831 0.04911027 0.025281717   9 static      6
#> 132  0.882445916 0.9936744 0.07602862 0.092675535  14  enter      6
#> 142 -0.261235385 0.4537346 0.12196996 0.082984576  15  enter      6
#> 34  -0.464799804 0.0378566 0.16536604 0.070887312   4  enter      7
#> 88   0.566511055 0.9631831 0.04911027 0.025281717   9    raw      7
#> 133  0.928146607 0.9936744 0.07602862 0.092675535  14 static      7
#> 143  0.193909661 0.4537346 0.12196996 0.082984576  15  enter      7
#> 35  -0.218398435 0.0378566 0.16536604 0.070887312   4  enter      8
#> 89   1.033255528 0.9631831 0.04911027 0.025281717   9   exit      8
#> 134  0.928146607 0.9936744 0.07602862 0.092675535  14 static      8
#> 144  0.402828699 0.4537346 0.12196996 0.082984576  15  enter      8
#> 36   0.318404548 0.0378566 0.16536604 0.070887312   4  enter      9
#> 55  -0.500000000 0.2801357 0.22076992 0.007416802   6  enter      9
#> 90   1.500000000 0.9631831 0.04911027 0.025281717   9   exit      9
#> 135  0.928146607 0.9936744 0.07602862 0.092675535  14 static      9
#> 145  0.432674276 0.4537346 0.12196996 0.082984576  15 static      9
#> 37   0.564805917 0.0378566 0.16536604 0.070887312   4  enter     10
#> 56  -0.460749384 0.2801357 0.22076992 0.007416802   6  enter     10
#> 136  0.928146607 0.9936744 0.07602862 0.092675535  14 static     10
#> 146  0.432674276 0.4537346 0.12196996 0.082984576  15 static     10
#> 38   0.600006113 0.0378566 0.16536604 0.070887312   4 static     11
#> 57  -0.185995071 0.2801357 0.22076992 0.007416802   6  enter     11
#> 110 -0.500000000 0.6809464 0.25816687 0.005540609  12  enter     11
#> 137  0.928146607 0.9936744 0.07602862 0.092675535  14    raw     11
#> 147  0.432674276 0.4537346 0.12196996 0.082984576  15 static     11
#> 184 -0.500000000 0.6035061 0.27596280 0.067683530  19  enter     11
#> 39   0.600006113 0.0378566 0.16536604 0.070887312   4 static     12
#> 58   0.088759242 0.2801357 0.22076992 0.007416802   6  enter     12
#> 111 -0.443341418 0.6809464 0.25816687 0.005540609  12  enter     12
#> 138  1.214073304 0.9936744 0.07602862 0.092675535  14   exit     12
#> 148  0.432674276 0.4537346 0.12196996 0.082984576  15 static     12
#> 185 -0.482876710 0.6035061 0.27596280 0.067683530  19  enter     12
#> 40   0.600006113 0.0378566 0.16536604 0.070887312   4 static     13
#> 59   0.128009858 0.2801357 0.22076992 0.007416802   6 static     13
#> 112 -0.046731342 0.6809464 0.25816687 0.005540609  12  enter     13
#> 139  1.500000000 0.9936744 0.07602862 0.092675535  14   exit     13
#> 149  0.432674276 0.4537346 0.12196996 0.082984576  15    raw     13
#> 186 -0.363013680 0.6035061 0.27596280 0.067683530  19  enter     13
#> 1   -0.500000000 0.5404593 0.32691562 0.070656365   1  enter     14
#> 41   0.600006113 0.0378566 0.16536604 0.070887312   4    raw     14
#> 60   0.128009858 0.2801357 0.22076992 0.007416802   6    raw     14
#> 113  0.349878733 0.6809464 0.25816687 0.005540609  12  enter     14
#> 150  0.966337138 0.4537346 0.12196996 0.082984576  15   exit     14
#> 187 -0.101883507 0.6035061 0.27596280 0.067683530  19  enter     14
#> 2   -0.447679546 0.5404593 0.32691562 0.070656365   1  enter     15
#> 42   1.050003056 0.0378566 0.16536604 0.070887312   4   exit     15
#> 61   0.814004929 0.2801357 0.22076992 0.007416802   6   exit     15
#> 114  0.406537316 0.6809464 0.25816687 0.005540609  12    raw     15
#> 151  1.500000000 0.4537346 0.12196996 0.082984576  15   exit     15
#> 188  0.017979523 0.6035061 0.27596280 0.067683530  19  enter     15
#> 3   -0.081436368 0.5404593 0.32691562 0.070656365   1  enter     16
#> 43   1.500000000 0.0378566 0.16536604 0.070887312   4   exit     16
#> 62   1.500000000 0.2801357 0.22076992 0.007416802   6   exit     16
#> 115  0.568531787 0.6809464 0.25816687 0.005540609  12   exit     16
#> 189  0.035102813 0.6035061 0.27596280 0.067683530  19 static     16
#> 195 -0.500000000 0.1586181 0.37344820 0.034032644  20  enter     16
#> 4    0.284806810 0.5404593 0.32691562 0.070656365   1  enter     17
#> 116  1.338005528 0.6809464 0.25816687 0.005540609  12   exit     17
#> 190  0.035102813 0.6035061 0.27596280 0.067683530  19 static     17
#> 196 -0.438344007 0.1586181 0.37344820 0.034032644  20  enter     17
#> 5    0.337127264 0.5404593 0.32691562 0.070656365   1 static     18
#> 117  1.500000000 0.6809464 0.25816687 0.005540609  12   exit     18
#> 172 -0.500000000 0.8265874 0.42791153 0.087212033  18  enter     18
#> 191  0.035102813 0.6035061 0.27596280 0.067683530  19 static     18
#> 197 -0.006752054 0.1586181 0.37344820 0.034032644  20  enter     18
#> 6    0.337127264 0.5404593 0.32691562 0.070656365   1 static     19
#> 102 -0.500000000 0.9109253 0.44551098 0.010852773  11  enter     19
#> 173 -0.482465275 0.8265874 0.42791153 0.087212033  18  enter     19
#> 192  0.035102813 0.6035061 0.27596280 0.067683530  19    raw     19
#> 198  0.424839898 0.1586181 0.37344820 0.034032644  20  enter     19
#> 7    0.337127264 0.5404593 0.32691562 0.070656365   1 static     20
#> 103 -0.463012385 0.9109253 0.44551098 0.010852773  11  enter     20
#> 174 -0.359722202 0.8265874 0.42791153 0.087212033  18  enter     20
#> 193  0.767551406 0.6035061 0.27596280 0.067683530  19   exit     20
#> 199  0.486495891 0.1586181 0.37344820 0.034032644  20 static     20
#> 8    0.337127264 0.5404593 0.32691562 0.070656365   1    raw     21
#> 12  -0.500000000 0.2425622 0.49372667 0.023293364   2  enter     21
#> 104 -0.204099080 0.9109253 0.44551098 0.010852773  11  enter     21
#> 118 -0.500000000 0.2269400 0.48588740 0.055918151  13  enter     21
#> 175 -0.092317651 0.8265874 0.42791153 0.087212033  18  enter     21
#> 194  1.500000000 0.6035061 0.27596280 0.067683530  19   exit     21
#> 200  0.486495891 0.1586181 0.37344820 0.034032644  20 static     21
#> 9    0.509404706 0.5404593 0.32691562 0.070656365   1   exit     22
#> 13  -0.455017649 0.2425622 0.49372667 0.023293364   2  enter     22
#> 105  0.359962048 0.9109253 0.44551098 0.010852773  11  enter     22
#> 119 -0.473494711 0.2269400 0.48588740 0.055918151  13  enter     22
#> 176  0.030425422 0.8265874 0.42791153 0.087212033  18  enter     22
#> 201  0.486495891 0.1586181 0.37344820 0.034032644  20    raw     22
#> 10   1.327722558 0.5404593 0.32691562 0.070656365   1   exit     23
#> 14  -0.140141191 0.2425622 0.49372667 0.023293364   2  enter     23
#> 91  -0.500000000 0.1625001 0.53227011 0.071173631  10  enter     23
#> 106  0.618875353 0.9109253 0.44551098 0.010852773  11  enter     23
#> 120 -0.287957685 0.2269400 0.48588740 0.055918151  13  enter     23
#> 177  0.047960147 0.8265874 0.42791153 0.087212033  18 static     23
#> 202  0.993247946 0.1586181 0.37344820 0.034032644  20   exit     23
#> 11   1.500000000 0.5404593 0.32691562 0.070656365   1   exit     24
#> 15   0.545839663 0.2425622 0.49372667 0.023293364   2  enter     24
#> 92  -0.459737347 0.1625001 0.53227011 0.071173631  10  enter     24
#> 107  0.655862968 0.9109253 0.44551098 0.010852773  11    raw     24
#> 121  0.116247978 0.2269400 0.48588740 0.055918151  13  enter     24
#> 178  0.047960147 0.8265874 0.42791153 0.087212033  18 static     24
#> 203  1.500000000 0.1586181 0.37344820 0.034032644  20   exit     24
#> 16   0.860716121 0.2425622 0.49372667 0.023293364   2  enter     25
#> 93  -0.177898774 0.1625001 0.53227011 0.071173631  10  enter     25
#> 108  1.077931484 0.9109253 0.44551098 0.010852773  11   exit     25
#> 122  0.301785004 0.2269400 0.48588740 0.055918151  13  enter     25
#> 179  0.047960147 0.8265874 0.42791153 0.087212033  18 static     25
#> 17   0.905698472 0.2425622 0.49372667 0.023293364   2 static     26
#> 21  -0.500000000 0.8444094 0.60565337 0.080357996   3  enter     26
#> 94   0.436106688 0.1625001 0.53227011 0.071173631  10  enter     26
#> 109  1.500000000 0.9109253 0.44551098 0.010852773  11   exit     26
#> 123  0.328290293 0.2269400 0.48588740 0.055918151  13 static     26
#> 152 -0.500000000 0.6142832 0.59138045 0.095027423  16  enter     26
#> 180  0.047960147 0.8265874 0.42791153 0.087212033  18 static     26
#> 18   0.905698472 0.2425622 0.49372667 0.023293364   2    raw     27
#> 22  -0.470455135 0.8444094 0.60565337 0.080357996   3  enter     27
#> 95   0.717945261 0.1625001 0.53227011 0.071173631  10  enter     27
#> 124  0.328290293 0.2269400 0.48588740 0.055918151  13 static     27
#> 153 -0.424486630 0.6142832 0.59138045 0.095027423  16  enter     27
#> 181  0.047960147 0.8265874 0.42791153 0.087212033  18    raw     27
#> 19   1.202849236 0.2425622 0.49372667 0.023293364   2   exit     28
#> 23  -0.263641083 0.8444094 0.60565337 0.080357996   3  enter     28
#> 96   0.758207914 0.1625001 0.53227011 0.071173631  10 static     28
#> 125  0.328290293 0.2269400 0.48588740 0.055918151  13    raw     28
#> 154  0.104106961 0.6142832 0.59138045 0.095027423  16  enter     28
#> 182  0.773980074 0.8265874 0.42791153 0.087212033  18   exit     28
#> 20   1.500000000 0.2425622 0.49372667 0.023293364   2   exit     29
#> 24   0.186918103 0.8444094 0.60565337 0.080357996   3  enter     29
#> 97   0.758207914 0.1625001 0.53227011 0.071173631  10 static     29
#> 126  0.914145147 0.2269400 0.48588740 0.055918151  13   exit     29
#> 155  0.632700552 0.6142832 0.59138045 0.095027423  16  enter     29
#> 183  1.500000000 0.8265874 0.42791153 0.087212033  18   exit     29
#> 25   0.393732155 0.8444094 0.60565337 0.080357996   3  enter     30
#> 63  -0.500000000 0.9801391 0.67790537 0.001064969   7  enter     30
#> 98   0.758207914 0.1625001 0.53227011 0.071173631  10 static     30
#> 127  1.500000000 0.2269400 0.48588740 0.055918151  13   exit     30
#> 156  0.708213922 0.6142832 0.59138045 0.095027423  16 static     30
#> 26   0.423277020 0.8444094 0.60565337 0.080357996   3 static     31
#> 64  -0.426003978 0.9801391 0.67790537 0.001064969   7  enter     31
#> 99   0.758207914 0.1625001 0.53227011 0.071173631  10    raw     31
#> 157  0.708213922 0.6142832 0.59138045 0.095027423  16 static     31
#> 27   0.423277020 0.8444094 0.60565337 0.080357996   3 static     32
#> 65   0.091968174 0.9801391 0.67790537 0.001064969   7  enter     32
#> 100  1.129103957 0.1625001 0.53227011 0.071173631  10   exit     32
#> 158  0.708213922 0.6142832 0.59138045 0.095027423  16 static     32
#> 28   0.423277020 0.8444094 0.60565337 0.080357996   3 static     33
#> 66   0.609940327 0.9801391 0.67790537 0.001064969   7  enter     33
#> 101  1.500000000 0.1625001 0.53227011 0.071173631  10   exit     33
#> 159  0.708213922 0.6142832 0.59138045 0.095027423  16 static     33
#> 29   0.423277020 0.8444094 0.60565337 0.080357996   3 static     34
#> 67   0.683936349 0.9801391 0.67790537 0.001064969   7    raw     34
#> 71  -0.500000000 0.2620210 0.76193203 0.078500727   8  enter     34
#> 160  0.708213922 0.6142832 0.59138045 0.095027423  16 static     34
#> 30   0.423277020 0.8444094 0.60565337 0.080357996   3    raw     35
#> 44  -0.500000000 0.9168547 0.78825826 0.063559779   5  enter     35
#> 68   0.804834667 0.9801391 0.67790537 0.001064969   7   exit     35
#> 72  -0.452348690 0.2620210 0.76193203 0.078500727   8  enter     35
#> 161  0.708213922 0.6142832 0.59138045 0.095027423  16    raw     35
#> 31   0.961638510 0.8444094 0.60565337 0.080357996   3   exit     36
#> 45  -0.448849020 0.9168547 0.78825826 0.063559779   5  enter     36
#> 69   1.379101681 0.9801391 0.67790537 0.001064969   7   exit     36
#> 73  -0.118789518 0.2620210 0.76193203 0.078500727   8  enter     36
#> 162  1.104106961 0.6142832 0.59138045 0.095027423  16   exit     36
#> 32   1.500000000 0.8444094 0.60565337 0.080357996   3   exit     37
#> 46  -0.090792160 0.9168547 0.78825826 0.063559779   5  enter     37
#> 70   1.500000000 0.9801391 0.67790537 0.001064969   7   exit     37
#> 74   0.214769653 0.2620210 0.76193203 0.078500727   8  enter     37
#> 163  1.500000000 0.6142832 0.59138045 0.095027423  16   exit     37
#> 47   0.267264700 0.9168547 0.78825826 0.063559779   5  enter     38
#> 75   0.262420964 0.2620210 0.76193203 0.078500727   8 static     38
#> 48   0.318415680 0.9168547 0.78825826 0.063559779   5 static     39
#> 76   0.262420964 0.2620210 0.76193203 0.078500727   8 static     39
#> 49   0.318415680 0.9168547 0.78825826 0.063559779   5 static     40
#> 77   0.262420964 0.2620210 0.76193203 0.078500727   8 static     40
#> 50   0.318415680 0.9168547 0.78825826 0.063559779   5 static     41
#> 78   0.262420964 0.2620210 0.76193203 0.078500727   8 static     41
#> 51   0.318415680 0.9168547 0.78825826 0.063559779   5    raw     42
#> 79   0.262420964 0.2620210 0.76193203 0.078500727   8    raw     42
#> 52   0.493465209 0.9168547 0.78825826 0.063559779   5   exit     43
#> 80   0.881210482 0.2620210 0.76193203 0.078500727   8   exit     43
#> 164 -0.500000000 0.6769289 0.96845764 0.003685942  17  enter     43
#> 53   1.324950471 0.9168547 0.78825826 0.063559779   5   exit     44
#> 81   1.500000000 0.2620210 0.76193203 0.078500727   8   exit     44
#> 165 -0.482984017 0.6769289 0.96845764 0.003685942  17  enter     44
#> 54   1.500000000 0.9168547 0.78825826 0.063559779   5   exit     45
#> 166 -0.363872139 0.6769289 0.96845764 0.003685942  17  enter     45
#> 167 -0.104378403 0.6769289 0.96845764 0.003685942  17  enter     46
#> 168  0.014733475 0.6769289 0.96845764 0.003685942  17  enter     47
#> 169  0.031749458 0.6769289 0.96845764 0.003685942  17    raw     48
#> 170  0.765874729 0.6769289 0.96845764 0.003685942  17   exit     49
#> 171  1.500000000 0.6769289 0.96845764 0.003685942  17   exit     50