Transition in and out of events

This tweening function is a more powerful version of tween_appear(), with support for newer features such as enter/exits and tween phase identification. The tweener treats each row in the data as unique events in time, and creates frames with the correct events present at any given time.

tween_events(
  .data,
  ease,
  nframes,
  start,
  end = NULL,
  range = NULL,
  enter = NULL,
  exit = NULL,
  enter_length = 0,
  exit_length = 0
)

Arguments

.data

A data.frame with components at different stages

ease

The easing function to use. Either a single string or one for each column in the data set.

nframes

The number of frames to calculate for the tween

start, end

The start (and potential end) of the event encoded in the row, as unquoted expressions. Will be evaluated in the context of .data so can refer to columns in it. If end = NULL the event will be without extend and only visible in a single frame, unless enter and/or exit is given.

range

The range of time points to include in the tween. If NULL it will use the range of time

enter

functions that calculate a start state for new observations that appear in to or an end state for observations that are not present in to. If NULL the new/old observations will not be part of the tween. The function gets a data.frame with either the start state of the exiting observations, or the end state of the entering observations and must return a modified version of that data.frame. See the Match, Enter, and Exit section for more information.

exit

functions that calculate a start state for new observations that appear in to or an end state for observations that are not present in to. If NULL the new/old observations will not be part of the tween. The function gets a data.frame with either the start state of the exiting observations, or the end state of the entering observations and must return a modified version of that data.frame. See the Match, Enter, and Exit section for more information.

enter_length

The lenght of the opening and closing transitions if enter and/or exit is given. Measured in the same units as time

exit_length

The lenght of the opening and closing transitions if enter and/or exit is given. Measured in the same units as time

Value

A data.frame with the same columns as .data along with .id giving the component id, .phase giving the state of each component in each frame, and .frame giving the frame membership of each row.

See also

Examples

d <- data.frame(
  x = runif(20),
  y = runif(20),
  time = runif(20),
  duration = runif(20, max = 0.1)
)
from_left <- function(x) {
  x$x <- -0.5
  x
}
to_right <- function(x) {
  x$x <- 1.5
  x
}

tween_events(d, 'cubic-in-out', 50, start = time, end = time + duration,
             enter = from_left, exit = to_right, enter_length = 0.1,
             exit_length = 0.05)
#>                x          y       time    duration .id .phase .frame
#> 68  -0.500000000 0.77157612 0.01048622 0.020730833   8  enter      1
#> 69  -0.413919248 0.77157612 0.01048622 0.020730833   8  enter      2
#> 70   0.188646016 0.77157612 0.01048622 0.020730833   8  enter      3
#> 132 -0.500000000 0.59016440 0.04884169 0.050818820  15  enter      3
#> 183 -0.500000000 0.85465563 0.05491495 0.021579102  20  enter      3
#> 71   0.791211279 0.77157612 0.01048622 0.020730833   8  enter      4
#> 133 -0.436723257 0.59016440 0.04884169 0.050818820  15  enter      4
#> 153 -0.500000000 0.67320255 0.09238641 0.077105321  17  enter      4
#> 184 -0.458582438 0.85465563 0.05491495 0.021579102  20  enter      4
#> 72   0.877292031 0.77157612 0.01048622 0.020730833   8 static      5
#> 134  0.006213941 0.59016440 0.04884169 0.050818820  15  enter      5
#> 154 -0.480135757 0.67320255 0.09238641 0.077105321  17  enter      5
#> 185 -0.168659506 0.85465563 0.05491495 0.021579102  20  enter      5
#> 73   0.877292031 0.77157612 0.01048622 0.020730833   8    raw      6
#> 135  0.449151140 0.59016440 0.04884169 0.050818820  15  enter      6
#> 155 -0.341086058 0.67320255 0.09238641 0.077105321  17  enter      6
#> 186  0.121263426 0.85465563 0.05491495 0.021579102  20  enter      6
#> 74   1.188646016 0.77157612 0.01048622 0.020730833   8   exit      7
#> 136  0.512427883 0.59016440 0.04884169 0.050818820  15 static      7
#> 156 -0.038156358 0.67320255 0.09238641 0.077105321  17  enter      7
#> 187  0.162680987 0.85465563 0.05491495 0.021579102  20 static      7
#> 56  -0.500000000 0.01655776 0.18114083 0.098269810   7  enter      8
#> 75   1.500000000 0.77157612 0.01048622 0.020730833   8   exit      8
#> 115 -0.500000000 0.04596846 0.18384493 0.042486853  13  enter      8
#> 137  0.512427883 0.59016440 0.04884169 0.050818820  15 static      8
#> 157  0.100893341 0.67320255 0.09238641 0.077105321  17  enter      8
#> 188  0.162680987 0.85465563 0.05491495 0.021579102  20    raw      8
#> 57  -0.451254091 0.01655776 0.18114083 0.098269810   7  enter      9
#> 116 -0.426148985 0.04596846 0.18384493 0.042486853  13  enter      9
#> 138  0.512427883 0.59016440 0.04884169 0.050818820  15    raw      9
#> 158  0.120757584 0.67320255 0.09238641 0.077105321  17 static      9
#> 189  0.831340494 0.85465563 0.05491495 0.021579102  20   exit      9
#> 58  -0.110032732 0.01655776 0.18114083 0.098269810   7  enter     10
#> 117  0.090808122 0.04596846 0.18384493 0.042486853  13  enter     10
#> 139  1.006213941 0.59016440 0.04884169 0.050818820  15   exit     10
#> 159  0.120757584 0.67320255 0.09238641 0.077105321  17 static     10
#> 190  1.500000000 0.85465563 0.05491495 0.021579102  20   exit     10
#> 59   0.231188628 0.01655776 0.18114083 0.098269810   7  enter     11
#> 106 -0.500000000 0.56650920 0.25784571 0.022162351  12  enter     11
#> 118  0.607765228 0.04596846 0.18384493 0.042486853  13  enter     11
#> 140  1.500000000 0.59016440 0.04884169 0.050818820  15   exit     11
#> 160  0.120757584 0.67320255 0.09238641 0.077105321  17 static     11
#> 60   0.279934537 0.01655776 0.18114083 0.098269810   7 static     12
#> 107 -0.472145562 0.56650920 0.25784571 0.022162351  12  enter     12
#> 119  0.681616243 0.04596846 0.18384493 0.042486853  13 static     12
#> 161  0.120757584 0.67320255 0.09238641 0.077105321  17    raw     12
#> 61   0.279934537 0.01655776 0.18114083 0.098269810   7 static     13
#> 108 -0.277164500 0.56650920 0.25784571 0.022162351  12  enter     13
#> 120  0.681616243 0.04596846 0.18384493 0.042486853  13 static     13
#> 162  0.810378792 0.67320255 0.09238641 0.077105321  17   exit     13
#> 62   0.279934537 0.01655776 0.18114083 0.098269810   7 static     14
#> 109  0.147615672 0.56650920 0.25784571 0.022162351  12  enter     14
#> 121  0.681616243 0.04596846 0.18384493 0.042486853  13    raw     14
#> 163  1.500000000 0.67320255 0.09238641 0.077105321  17   exit     14
#> 63   0.279934537 0.01655776 0.18114083 0.098269810   7 static     15
#> 110  0.342596735 0.56650920 0.25784571 0.022162351  12  enter     15
#> 122  1.090808122 0.04596846 0.18384493 0.042486853  13   exit     15
#> 21  -0.500000000 0.56035216 0.35980947 0.002969193   3  enter     16
#> 47  -0.500000000 0.04993308 0.37305577 0.047656770   6  enter     16
#> 64   0.279934537 0.01655776 0.18114083 0.098269810   7    raw     16
#> 111  0.370451173 0.56650920 0.25784571 0.022162351  12    raw     16
#> 123  1.500000000 0.04596846 0.18384493 0.042486853  13   exit     16
#> 22  -0.407498771 0.56035216 0.35980947 0.002969193   3  enter     17
#> 48  -0.462821213 0.04993308 0.37305577 0.047656770   6  enter     17
#> 65   0.460684976 0.01655776 0.18114083 0.098269810   7   exit     17
#> 88  -0.500000000 0.15397599 0.39153884 0.055245794  10  enter     17
#> 98  -0.500000000 0.99632063 0.39473791 0.018669197  11  enter     17
#> 112  0.537791740 0.56650920 0.25784571 0.022162351  12   exit     17
#> 23   0.240009832 0.56035216 0.35980947 0.002969193   3  enter     18
#> 49  -0.202569703 0.04993308 0.37305577 0.047656770   6  enter     18
#> 66   1.319249561 0.01655776 0.18114083 0.098269810   7   exit     18
#> 89  -0.439025863 0.15397599 0.39153884 0.055245794  10  enter     18
#> 99  -0.458535567 0.99632063 0.39473791 0.018669197  11  enter     18
#> 113  1.332659433 0.56650920 0.25784571 0.022162351  12   exit     18
#> 24   0.887518436 0.56035216 0.35980947 0.002969193   3  enter     19
#> 50   0.057681806 0.04993308 0.37305577 0.047656770   6  enter     19
#> 67   1.500000000 0.01655776 0.18114083 0.098269810   7   exit     19
#> 90  -0.012206903 0.15397599 0.39153884 0.055245794  10  enter     19
#> 100 -0.168284539 0.99632063 0.39473791 0.018669197  11  enter     19
#> 114  1.500000000 0.56650920 0.25784571 0.022162351  12   exit     19
#> 25   0.980019665 0.56035216 0.35980947 0.002969193   3    raw     20
#> 51   0.094860594 0.04993308 0.37305577 0.047656770   6 static     20
#> 91   0.414612057 0.15397599 0.39153884 0.055245794  10  enter     20
#> 101  0.121966490 0.99632063 0.39473791 0.018669197  11  enter     20
#> 26   1.240009832 0.56035216 0.35980947 0.002969193   3   exit     21
#> 39  -0.500000000 0.83495039 0.47926965 0.035764342   5  enter     21
#> 52   0.094860594 0.04993308 0.37305577 0.047656770   6 static     21
#> 92   0.475586194 0.15397599 0.39153884 0.055245794  10 static     21
#> 102  0.163430922 0.99632063 0.39473791 0.018669197  11 static     21
#> 27   1.500000000 0.56035216 0.35980947 0.002969193   3   exit     22
#> 40  -0.409968107 0.83495039 0.47926965 0.035764342   5  enter     22
#> 53   0.094860594 0.04993308 0.37305577 0.047656770   6    raw     22
#> 93   0.475586194 0.15397599 0.39153884 0.055245794  10 static     22
#> 103  0.163430922 0.99632063 0.39473791 0.018669197  11    raw     22
#> 41   0.220255143 0.83495039 0.47926965 0.035764342   5  enter     23
#> 54   0.797430297 0.04993308 0.37305577 0.047656770   6   exit     23
#> 94   0.475586194 0.15397599 0.39153884 0.055245794  10    raw     23
#> 104  0.831715461 0.99632063 0.39473791 0.018669197  11   exit     23
#> 164 -0.500000000 0.78136729 0.53422108 0.060867664  18  enter     23
#> 42   0.850478392 0.83495039 0.47926965 0.035764342   5  enter     24
#> 55   1.500000000 0.04993308 0.37305577 0.047656770   6   exit     24
#> 95   0.627351202 0.15397599 0.39153884 0.055245794  10   exit     24
#> 105  1.500000000 0.99632063 0.39473791 0.018669197  11   exit     24
#> 165 -0.433513104 0.78136729 0.53422108 0.060867664  18  enter     24
#> 28  -0.500000000 0.49616272 0.58636867 0.076239916   4  enter     25
#> 43   0.940510285 0.83495039 0.47926965 0.035764342   5 static     25
#> 96   1.348234992 0.15397599 0.39153884 0.055245794  10   exit     25
#> 166  0.031895165 0.78136729 0.53422108 0.060867664  18  enter     25
#> 29  -0.462649562 0.49616272 0.58636867 0.076239916   4  enter     26
#> 44   0.940510285 0.83495039 0.47926965 0.035764342   5    raw     26
#> 76  -0.500000000 0.31242737 0.61394738 0.094312062   9  enter     26
#> 97   1.500000000 0.15397599 0.39153884 0.055245794  10   exit     26
#> 167  0.497303433 0.78136729 0.53422108 0.060867664  18  enter     26
#> 30  -0.201196492 0.49616272 0.58636867 0.076239916   4  enter     27
#> 45   1.220255143 0.83495039 0.47926965 0.035764342   5   exit     27
#> 77  -0.476027642 0.31242737 0.61394738 0.094312062   9  enter     27
#> 168  0.563790329 0.78136729 0.53422108 0.060867664  18 static     27
#> 31   0.060256577 0.49616272 0.58636867 0.076239916   4  enter     28
#> 46   1.500000000 0.83495039 0.47926965 0.035764342   5   exit     28
#> 78  -0.308221137 0.31242737 0.61394738 0.094312062   9  enter     28
#> 169  0.563790329 0.78136729 0.53422108 0.060867664  18 static     28
#> 32   0.097607016 0.49616272 0.58636867 0.076239916   4 static     29
#> 79   0.057357320 0.31242737 0.61394738 0.094312062   9  enter     29
#> 141 -0.500000000 0.92187139 0.68471637 0.099245520  16  enter     29
#> 170  0.563790329 0.78136729 0.53422108 0.060867664  18 static     29
#> 33   0.097607016 0.49616272 0.58636867 0.076239916   4 static     30
#> 80   0.225163825 0.31242737 0.61394738 0.094312062   9  enter     30
#> 142 -0.435058660 0.92187139 0.68471637 0.099245520  16  enter     30
#> 171  0.563790329 0.78136729 0.53422108 0.060867664  18    raw     30
#> 12  -0.500000000 0.98232691 0.73100715 0.023959307   2  enter     31
#> 34   0.097607016 0.49616272 0.58636867 0.076239916   4 static     31
#> 81   0.249136183 0.31242737 0.61394738 0.094312062   9 static     31
#> 143  0.019530717 0.92187139 0.68471637 0.099245520  16  enter     31
#> 172  1.031895165 0.78136729 0.53422108 0.060867664  18   exit     31
#> 13  -0.417514218 0.98232691 0.73100715 0.023959307   2  enter     32
#> 35   0.097607016 0.49616272 0.58636867 0.076239916   4 static     32
#> 82   0.249136183 0.31242737 0.61394738 0.094312062   9 static     32
#> 144  0.474120094 0.92187139 0.68471637 0.099245520  16  enter     32
#> 173  1.500000000 0.78136729 0.53422108 0.060867664  18   exit     32
#> 14   0.159886260 0.98232691 0.73100715 0.023959307   2  enter     33
#> 36   0.097607016 0.49616272 0.58636867 0.076239916   4    raw     33
#> 83   0.249136183 0.31242737 0.61394738 0.094312062   9 static     33
#> 145  0.539061434 0.92187139 0.68471637 0.099245520  16 static     33
#> 15   0.737286737 0.98232691 0.73100715 0.023959307   2  enter     34
#> 37   0.798803508 0.49616272 0.58636867 0.076239916   4   exit     34
#> 84   0.249136183 0.31242737 0.61394738 0.094312062   9    raw     34
#> 146  0.539061434 0.92187139 0.68471637 0.099245520  16 static     34
#> 16   0.819772519 0.98232691 0.73100715 0.023959307   2 static     35
#> 38   1.500000000 0.49616272 0.58636867 0.076239916   4   exit     35
#> 85   0.434449341 0.31242737 0.61394738 0.094312062   9   exit     35
#> 147  0.539061434 0.92187139 0.68471637 0.099245520  16 static     35
#> 17   0.819772519 0.98232691 0.73100715 0.023959307   2    raw     36
#> 86   1.314686842 0.31242737 0.61394738 0.094312062   9   exit     36
#> 148  0.539061434 0.92187139 0.68471637 0.099245520  16 static     36
#> 18   0.920546961 0.98232691 0.73100715 0.023959307   2   exit     37
#> 87   1.500000000 0.31242737 0.61394738 0.094312062   9   exit     37
#> 124 -0.500000000 0.54182177 0.85804149 0.025032961  14  enter     37
#> 149  0.539061434 0.92187139 0.68471637 0.099245520  16 static     37
#> 19   1.399225558 0.98232691 0.73100715 0.023959307   2   exit     38
#> 125 -0.415014103 0.54182177 0.85804149 0.025032961  14  enter     38
#> 150  0.539061434 0.92187139 0.68471637 0.099245520  16    raw     38
#> 174 -0.500000000 0.34356280 0.90383746 0.026875205  19  enter     38
#> 20   1.500000000 0.98232691 0.73100715 0.023959307   2   exit     39
#> 126  0.179887177 0.54182177 0.85804149 0.025032961  14  enter     39
#> 151  1.019530717 0.92187139 0.68471637 0.099245520  16   exit     39
#> 175 -0.455195041 0.34356280 0.90383746 0.026875205  19  enter     39
#> 1   -0.500000000 0.93420653 0.93202862 0.096682407   1  enter     40
#> 127  0.774788457 0.54182177 0.85804149 0.025032961  14  enter     40
#> 152  1.500000000 0.92187139 0.68471637 0.099245520  16   exit     40
#> 176 -0.141560326 0.34356280 0.90383746 0.026875205  19  enter     40
#> 2   -0.435446905 0.93420653 0.93202862 0.096682407   1  enter     41
#> 128  0.859774354 0.54182177 0.85804149 0.025032961  14 static     41
#> 177  0.541715302 0.34356280 0.90383746 0.026875205  19  enter     41
#> 3    0.016424760 0.93420653 0.93202862 0.096682407   1  enter     42
#> 129  0.859774354 0.54182177 0.85804149 0.025032961  14    raw     42
#> 178  0.855350016 0.34356280 0.90383746 0.026875205  19  enter     42
#> 4    0.468296425 0.93420653 0.93202862 0.096682407   1  enter     43
#> 130  1.179887177 0.54182177 0.85804149 0.025032961  14   exit     43
#> 179  0.900154975 0.34356280 0.90383746 0.026875205  19 static     43
#> 5    0.532849520 0.93420653 0.93202862 0.096682407   1 static     44
#> 131  1.500000000 0.54182177 0.85804149 0.025032961  14   exit     44
#> 180  0.900154975 0.34356280 0.90383746 0.026875205  19    raw     44
#> 6    0.532849520 0.93420653 0.93202862 0.096682407   1 static     45
#> 181  1.200077488 0.34356280 0.90383746 0.026875205  19   exit     45
#> 7    0.532849520 0.93420653 0.93202862 0.096682407   1 static     46
#> 182  1.500000000 0.34356280 0.90383746 0.026875205  19   exit     46
#> 8    0.532849520 0.93420653 0.93202862 0.096682407   1 static     47
#> 9    0.532849520 0.93420653 0.93202862 0.096682407   1    raw     48
#> 10   1.016424760 0.93420653 0.93202862 0.096682407   1   exit     49
#> 11   1.500000000 0.93420653 0.93202862 0.096682407   1   exit     50