Transition in and out of events

This tweening function is a more powerful version of tween_appear(), with support for newer features such as enter/exits and tween phase identification. The tweener treats each row in the data as unique events in time, and creates frames with the correct events present at any given time.

tween_events(
  .data,
  ease,
  nframes,
  start,
  end = NULL,
  range = NULL,
  enter = NULL,
  exit = NULL,
  enter_length = 0,
  exit_length = 0
)

Arguments

.data

A data.frame with components at different stages

ease

The easing function to use. Either a single string or one for each column in the data set.

nframes

The number of frames to calculate for the tween

start, end

The start (and potential end) of the event encoded in the row, as unquoted expressions. Will be evaluated in the context of .data so can refer to columns in it. If end = NULL the event will be without extend and only visible in a single frame, unless enter and/or exit is given.

range

The range of time points to include in the tween. If NULL it will use the range of time

enter, exit

functions that calculate a start state for new observations that appear in to or an end state for observations that are not present in to. If NULL the new/old observations will not be part of the tween. The function gets a data.frame with either the start state of the exiting observations, or the end state of the entering observations and must return a modified version of that data.frame. See the Match, Enter, and Exit section for more information.

enter_length, exit_length

The lenght of the opening and closing transitions if enter and/or exit is given. Measured in the same units as time

Value

A data.frame with the same columns as .data along with .id giving the component id, .phase giving the state of each component in each frame, and .frame giving the frame membership of each row.

See also

Examples

d <- data.frame(
  x = runif(20),
  y = runif(20),
  time = runif(20),
  duration = runif(20, max = 0.1)
)
from_left <- function(x) {
  x$x <- -0.5
  x
}
to_right <- function(x) {
  x$x <- 1.5
  x
}

tween_events(d, 'cubic-in-out', 50, start = time, end = time + duration,
             enter = from_left, exit = to_right, enter_length = 0.1,
             exit_length = 0.05)
#>               x          y      time    duration .id .phase .frame
#> 72  -0.50000000 0.05295420 0.1435463 0.083388875   8  enter      1
#> 73  -0.47525548 0.05295420 0.1435463 0.083388875   8  enter      2
#> 74  -0.30204386 0.05295420 0.1435463 0.083388875   8  enter      3
#> 75   0.07531003 0.05295420 0.1435463 0.083388875   8  enter      4
#> 76   0.24852165 0.05295420 0.1435463 0.083388875   8  enter      5
#> 77   0.27326617 0.05295420 0.1435463 0.083388875   8 static      6
#> 78   0.27326617 0.05295420 0.1435463 0.083388875   8 static      7
#> 10  -0.50000000 0.76600812 0.2920822 0.023493670   2  enter      8
#> 79   0.27326617 0.05295420 0.1435463 0.083388875   8 static      8
#> 186 -0.50000000 0.83134397 0.3036425 0.092153998  19  enter      8
#> 11  -0.40772549 0.76600812 0.2920822 0.023493670   2  enter      9
#> 50  -0.50000000 0.49827238 0.3288646 0.077920516   6  enter      9
#> 80   0.27326617 0.05295420 0.1435463 0.083388875   8    raw      9
#> 187 -0.45645382 0.83134397 0.3036425 0.092153998  19  enter      9
#> 12   0.23819610 0.76600812 0.2920822 0.023493670   2  enter     10
#> 51  -0.46069681 0.49827238 0.3288646 0.077920516   6  enter     10
#> 81   0.45500451 0.05295420 0.1435463 0.083388875   8   exit     10
#> 188 -0.15163060 0.83134397 0.3036425 0.092153998  19  enter     10
#> 13   0.88411768 0.76600812 0.2920822 0.023493670   2  enter     11
#> 52  -0.18557450 0.49827238 0.3288646 0.077920516   6  enter     11
#> 82   1.31826165 0.05295420 0.1435463 0.083388875   8   exit     11
#> 189  0.51244858 0.83134397 0.3036425 0.092153998  19  enter     11
#> 14   0.97639219 0.76600812 0.2920822 0.023493670   2 static     12
#> 53   0.41379911 0.49827238 0.3288646 0.077920516   6  enter     12
#> 83   1.50000000 0.05295420 0.1435463 0.083388875   8   exit     12
#> 190  0.81727181 0.83134397 0.3036425 0.092153998  19  enter     12
#> 15   0.97639219 0.76600812 0.2920822 0.023493670   2    raw     13
#> 54   0.68892142 0.49827238 0.3288646 0.077920516   6  enter     13
#> 191  0.86081799 0.83134397 0.3036425 0.092153998  19 static     13
#> 16   1.05396372 0.76600812 0.2920822 0.023493670   2   exit     14
#> 55   0.72822460 0.49827238 0.3288646 0.077920516   6 static     14
#> 192  0.86081799 0.83134397 0.3036425 0.092153998  19 static     14
#> 17   1.42242847 0.76600812 0.2920822 0.023493670   2   exit     15
#> 56   0.72822460 0.49827238 0.3288646 0.077920516   6 static     15
#> 193  0.86081799 0.83134397 0.3036425 0.092153998  19 static     15
#> 18   1.50000000 0.76600812 0.2920822 0.023493670   2   exit     16
#> 57   0.72822460 0.49827238 0.3288646 0.077920516   6 static     16
#> 194  0.86081799 0.83134397 0.3036425 0.092153998  19 static     16
#> 58   0.72822460 0.49827238 0.3288646 0.077920516   6    raw     17
#> 118 -0.50000000 0.98531935 0.5074513 0.086542905  12  enter     17
#> 195  0.86081799 0.83134397 0.3036425 0.092153998  19    raw     17
#> 59   0.84256170 0.49827238 0.3288646 0.077920516   6   exit     18
#> 119 -0.47465433 0.98531935 0.5074513 0.086542905  12  enter     18
#> 196  1.18040899 0.83134397 0.3036425 0.092153998  19   exit     18
#> 1   -0.50000000 0.69706730 0.5360203 0.030728375   1  enter     19
#> 60   1.38566290 0.49827238 0.3288646 0.077920516   6   exit     19
#> 120 -0.29723461 0.98531935 0.5074513 0.086542905  12  enter     19
#> 160 -0.50000000 0.56341624 0.5505045 0.030916693  16  enter     19
#> 197  1.50000000 0.83134397 0.3036425 0.092153998  19   exit     19
#> 2   -0.45420098 0.69706730 0.5360203 0.030728375   1  enter     20
#> 61   1.50000000 0.49827238 0.3288646 0.077920516   6   exit     20
#> 121  0.08928692 0.98531935 0.5074513 0.086542905  12  enter     20
#> 161 -0.45640462 0.56341624 0.5505045 0.030916693  16  enter     20
#> 3   -0.13360786 0.69706730 0.5360203 0.030728375   1  enter     21
#> 30  -0.50000000 0.95926021 0.5957018 0.067480219   4  enter     21
#> 41  -0.50000000 0.98996515 0.5774142 0.024935029   5  enter     21
#> 122  0.26670664 0.98531935 0.5074513 0.086542905  12  enter     21
#> 162 -0.15123694 0.56341624 0.5505045 0.030916693  16  enter     21
#> 4    0.18698527 0.69706730 0.5360203 0.030728375   1  enter     22
#> 31  -0.46438209 0.95926021 0.5957018 0.067480219   4  enter     22
#> 42  -0.43312422 0.98996515 0.5774142 0.024935029   5  enter     22
#> 123  0.29205232 0.98531935 0.5074513 0.086542905  12 static     22
#> 163  0.51359266 0.56341624 0.5505045 0.030916693  16  enter     22
#> 5    0.23278428 0.69706730 0.5360203 0.030728375   1 static     23
#> 32  -0.21505673 0.95926021 0.5957018 0.067480219   4  enter     23
#> 43   0.03500626 0.98996515 0.5774142 0.024935029   5  enter     23
#> 124  0.29205232 0.98531935 0.5074513 0.086542905  12 static     23
#> 164  0.81876034 0.56341624 0.5505045 0.030916693  16  enter     23
#> 6    0.23278428 0.69706730 0.5360203 0.030728375   1 static     24
#> 33   0.32811637 0.95926021 0.5957018 0.067480219   4  enter     24
#> 44   0.50313673 0.98996515 0.5774142 0.024935029   5  enter     24
#> 125  0.29205232 0.98531935 0.5074513 0.086542905  12 static     24
#> 165  0.86235572 0.56341624 0.5505045 0.030916693  16 static     24
#> 7    0.23278428 0.69706730 0.5360203 0.030728375   1    raw     25
#> 34   0.57744173 0.95926021 0.5957018 0.067480219   4  enter     25
#> 45   0.57001252 0.98996515 0.5774142 0.024935029   5 static     25
#> 126  0.29205232 0.98531935 0.5074513 0.086542905  12 static     25
#> 166  0.86235572 0.56341624 0.5505045 0.030916693  16    raw     25
#> 8    0.86639214 0.69706730 0.5360203 0.030728375   1   exit     26
#> 35   0.61305964 0.95926021 0.5957018 0.067480219   4 static     26
#> 46   0.57001252 0.98996515 0.5774142 0.024935029   5    raw     26
#> 127  0.29205232 0.98531935 0.5074513 0.086542905  12    raw     26
#> 167  0.95682154 0.56341624 0.5505045 0.030916693  16   exit     26
#> 9    1.50000000 0.69706730 0.5360203 0.030728375   1   exit     27
#> 36   0.61305964 0.95926021 0.5957018 0.067480219   4 static     27
#> 47   0.70778844 0.98996515 0.5774142 0.024935029   5   exit     27
#> 128  0.89602616 0.98531935 0.5074513 0.086542905  12   exit     27
#> 168  1.40553418 0.56341624 0.5505045 0.030916693  16   exit     27
#> 37   0.61305964 0.95926021 0.5957018 0.067480219   4 static     28
#> 48   1.36222408 0.98996515 0.5774142 0.024935029   5   exit     28
#> 62  -0.50000000 0.15807564 0.7493248 0.031960261   7  enter     28
#> 129  1.50000000 0.98531935 0.5074513 0.086542905  12   exit     28
#> 169  1.50000000 0.56341624 0.5505045 0.030916693  16   exit     28
#> 38   0.61305964 0.95926021 0.5957018 0.067480219   4    raw     29
#> 49   1.50000000 0.98996515 0.5774142 0.024935029   5   exit     29
#> 63  -0.46105955 0.15807564 0.7493248 0.031960261   7  enter     29
#> 39   1.05652982 0.95926021 0.5957018 0.067480219   4   exit     30
#> 64  -0.18847642 0.15807564 0.7493248 0.031960261   7  enter     30
#> 149 -0.50000000 0.36000480 0.7825906 0.067830024  15  enter     30
#> 40   1.50000000 0.95926021 0.5957018 0.067480219   4   exit     31
#> 65   0.40536542 0.15807564 0.7493248 0.031960261   7  enter     31
#> 139 -0.50000000 0.04569286 0.8084343 0.050363983  14  enter     31
#> 150 -0.44477325 0.36000480 0.7825906 0.067830024  15  enter     31
#> 66   0.67794856 0.15807564 0.7493248 0.031960261   7  enter     32
#> 140 -0.47802142 0.04569286 0.8084343 0.050363983  14  enter     32
#> 151 -0.05818599 0.36000480 0.7825906 0.067830024  15  enter     32
#> 67   0.71688900 0.15807564 0.7493248 0.031960261   7 static     33
#> 96  -0.50000000 0.89354048 0.8444818 0.086035455  10  enter     33
#> 141 -0.32417136 0.04569286 0.8084343 0.050363983  14  enter     33
#> 152  0.32840126 0.36000480 0.7825906 0.067830024  15  enter     33
#> 19  -0.50000000 0.63024605 0.8786954 0.057384947   3  enter     34
#> 68   0.71688900 0.15807564 0.7493248 0.031960261   7    raw     34
#> 97  -0.43910499 0.89354048 0.8444818 0.086035455  10  enter     34
#> 142  0.01100199 0.04569286 0.8084343 0.050363983  14  enter     34
#> 153  0.38362801 0.36000480 0.7825906 0.067830024  15 static     34
#> 20  -0.45832964 0.63024605 0.8786954 0.057384947   3  enter     35
#> 69   0.83290545 0.15807564 0.7493248 0.031960261   7   exit     35
#> 98  -0.01283993 0.89354048 0.8444818 0.086035455  10  enter     35
#> 143  0.16485205 0.04569286 0.8084343 0.050363983  14  enter     35
#> 154  0.38362801 0.36000480 0.7825906 0.067830024  15 static     35
#> 21  -0.16663711 0.63024605 0.8786954 0.057384947   3  enter     36
#> 70   1.38398356 0.15807564 0.7493248 0.031960261   7   exit     36
#> 99   0.41342512 0.89354048 0.8444818 0.086035455  10  enter     36
#> 107 -0.50000000 0.60799961 0.9166153 0.079105662  11  enter     36
#> 130 -0.50000000 0.76315811 0.9218061 0.023550280  13  enter     36
#> 144  0.18683063 0.04569286 0.8084343 0.050363983  14 static     36
#> 155  0.38362801 0.36000480 0.7825906 0.067830024  15 static     36
#> 198 -0.50000000 0.35132613 0.9060738 0.075009677  20  enter     36
#> 22   0.46883589 0.63024605 0.8786954 0.057384947   3  enter     37
#> 71   1.50000000 0.15807564 0.7493248 0.031960261   7   exit     37
#> 100  0.47432013 0.89354048 0.8444818 0.086035455  10 static     37
#> 108 -0.47753735 0.60799961 0.9166153 0.079105662  11  enter     37
#> 131 -0.45854972 0.76315811 0.9218061 0.023550280  13  enter     37
#> 145  0.18683063 0.04569286 0.8084343 0.050363983  14 static     37
#> 156  0.38362801 0.36000480 0.7825906 0.067830024  15 static     37
#> 199 -0.44229200 0.35132613 0.9060738 0.075009677  20  enter     37
#> 23   0.76052842 0.63024605 0.8786954 0.057384947   3  enter     38
#> 101  0.47432013 0.89354048 0.8444818 0.086035455  10 static     38
#> 109 -0.32029883 0.60799961 0.9166153 0.079105662  11  enter     38
#> 132 -0.16839777 0.76315811 0.9218061 0.023550280  13  enter     38
#> 146  0.18683063 0.04569286 0.8084343 0.050363983  14    raw     38
#> 157  0.38362801 0.36000480 0.7825906 0.067830024  15    raw     38
#> 200 -0.03833603 0.35132613 0.9060738 0.075009677  20  enter     38
#> 24   0.80219878 0.63024605 0.8786954 0.057384947   3 static     39
#> 84  -0.50000000 0.26342196 0.9837050 0.092546809   9  enter     39
#> 102  0.47432013 0.89354048 0.8444818 0.086035455  10 static     39
#> 110  0.02225651 0.60799961 0.9166153 0.079105662  11  enter     39
#> 133  0.46371899 0.76315811 0.9218061 0.023550280  13  enter     39
#> 147  0.84341532 0.04569286 0.8084343 0.050363983  14   exit     39
#> 158  0.94181401 0.36000480 0.7825906 0.067830024  15   exit     39
#> 170 -0.50000000 0.16481774 0.9817456 0.013041831  17  enter     39
#> 178 -0.50000000 0.49589428 0.9805887 0.001357167  18  enter     39
#> 201  0.36561995 0.35132613 0.9060738 0.075009677  20  enter     39
#> 25   0.80219878 0.63024605 0.8786954 0.057384947   3 static     40
#> 85  -0.47361847 0.26342196 0.9837050 0.092546809   9  enter     40
#> 103  0.47432013 0.89354048 0.8444818 0.086035455  10 static     40
#> 111  0.17949503 0.60799961 0.9166153 0.079105662  11  enter     40
#> 134  0.75387094 0.76315811 0.9218061 0.023550280  13  enter     40
#> 148  1.50000000 0.04569286 0.8084343 0.050363983  14   exit     40
#> 159  1.50000000 0.36000480 0.7825906 0.067830024  15   exit     40
#> 171 -0.42532534 0.16481774 0.9817456 0.013041831  17  enter     40
#> 179 -0.41737058 0.49589428 0.9805887 0.001357167  18  enter     40
#> 202  0.42332795 0.35132613 0.9060738 0.075009677  20 static     40
#> 26   0.80219878 0.63024605 0.8786954 0.057384947   3    raw     41
#> 86  -0.28894777 0.26342196 0.9837050 0.092546809   9  enter     41
#> 104  0.47432013 0.89354048 0.8444818 0.086035455  10    raw     41
#> 112  0.20195768 0.60799961 0.9166153 0.079105662  11 static     41
#> 135  0.79532122 0.76315811 0.9218061 0.023550280  13 static     41
#> 172  0.09739731 0.16481774 0.9817456 0.013041831  17  enter     41
#> 180  0.16103535 0.49589428 0.9805887 0.001357167  18  enter     41
#> 203  0.42332795 0.35132613 0.9060738 0.075009677  20 static     41
#> 27   0.90557674 0.63024605 0.8786954 0.057384947   3   exit     42
#> 87   0.11337055 0.26342196 0.9837050 0.092546809   9  enter     42
#> 105  0.98716007 0.89354048 0.8444818 0.086035455  10   exit     42
#> 113  0.20195768 0.60799961 0.9166153 0.079105662  11 static     42
#> 136  0.79532122 0.76315811 0.9218061 0.023550280  13    raw     42
#> 173  0.62011996 0.16481774 0.9817456 0.013041831  17  enter     42
#> 181  0.73944129 0.49589428 0.9805887 0.001357167  18  enter     42
#> 204  0.42332795 0.35132613 0.9060738 0.075009677  20 static     42
#> 28   1.39662204 0.63024605 0.8786954 0.057384947   3   exit     43
#> 88   0.29804125 0.26342196 0.9837050 0.092546809   9  enter     43
#> 106  1.50000000 0.89354048 0.8444818 0.086035455  10   exit     43
#> 114  0.20195768 0.60799961 0.9166153 0.079105662  11 static     43
#> 137  1.14766061 0.76315811 0.9218061 0.023550280  13   exit     43
#> 174  0.69479462 0.16481774 0.9817456 0.013041831  17 static     43
#> 182  0.82207071 0.49589428 0.9805887 0.001357167  18    raw     43
#> 205  0.42332795 0.35132613 0.9060738 0.075009677  20    raw     43
#> 29   1.50000000 0.63024605 0.8786954 0.057384947   3   exit     44
#> 89   0.32442278 0.26342196 0.9837050 0.092546809   9 static     44
#> 115  0.20195768 0.60799961 0.9166153 0.079105662  11    raw     44
#> 138  1.50000000 0.76315811 0.9218061 0.023550280  13   exit     44
#> 175  0.69479462 0.16481774 0.9817456 0.013041831  17    raw     44
#> 183  0.92250468 0.49589428 0.9805887 0.001357167  18   exit     44
#> 206  0.58283492 0.35132613 0.9060738 0.075009677  20   exit     44
#> 90   0.32442278 0.26342196 0.9837050 0.092546809   9 static     45
#> 116  0.85097884 0.60799961 0.9166153 0.079105662  11   exit     45
#> 176  1.09739731 0.16481774 0.9817456 0.013041831  17   exit     45
#> 184  1.39956603 0.49589428 0.9805887 0.001357167  18   exit     45
#> 207  1.34049303 0.35132613 0.9060738 0.075009677  20   exit     45
#> 91   0.32442278 0.26342196 0.9837050 0.092546809   9 static     46
#> 117  1.50000000 0.60799961 0.9166153 0.079105662  11   exit     46
#> 177  1.50000000 0.16481774 0.9817456 0.013041831  17   exit     46
#> 185  1.50000000 0.49589428 0.9805887 0.001357167  18   exit     46
#> 208  1.50000000 0.35132613 0.9060738 0.075009677  20   exit     46
#> 92   0.32442278 0.26342196 0.9837050 0.092546809   9 static     47
#> 93   0.32442278 0.26342196 0.9837050 0.092546809   9    raw     48
#> 94   0.91221139 0.26342196 0.9837050 0.092546809   9   exit     49
#> 95   1.50000000 0.26342196 0.9837050 0.092546809   9   exit     50