Transition in and out of events

This tweening function is a more powerful version of tween_appear(), with support for newer features such as enter/exits and tween phase identification. The tweener treats each row in the data as unique events in time, and creates frames with the correct events present at any given time.

tween_events(.data, ease, nframes, start, end = NULL, range = NULL,
  enter = NULL, exit = NULL, enter_length = 0, exit_length = 0)

Arguments

.data
A data.frame with components at different stages

ease
The easing function to use. Either a single string or one for each column in the data set.

nframes
The number of frames to calculate for the tween

start, end
The start (and potential end) of the event encoded in the row, as unquoted expressions. Will be evaluated in the context of .data so can refer to columns in it. If end = NULL the event will be without extend and only visible in a single frame, unless enter and/or exit is given.

range
The range of time points to include in the tween. If NULL it will use the range of time

enter
functions that calculate a start state for new observations that appear in to or an end state for observations that are not present in to. If NULL the new/old observations will not be part of the tween. The function gets a data.frame with either the start state of the exiting observations, or the end state of the entering observations and must return a modified version of that data.frame. See the Match, Enter, and Exit section for more information.

exit
functions that calculate a start state for new observations that appear in to or an end state for observations that are not present in to. If NULL the new/old observations will not be part of the tween. The function gets a data.frame with either the start state of the exiting observations, or the end state of the entering observations and must return a modified version of that data.frame. See the Match, Enter, and Exit section for more information.

enter_length
The lenght of the opening and closing transitions if enter and/or exit is given. Measured in the same units as time

exit_length
The lenght of the opening and closing transitions if enter and/or exit is given. Measured in the same units as time

Value

A data.frame with the same columns as .data along with .id giving the component id, .phase giving the state of each component in each frame, and .frame giving the frame membership of each row.

See also

Examples

d <- data.frame(
  x = runif(20),
  y = runif(20),
  time = runif(20),
  duration = runif(20, max = 0.1)
)
from_left <- function(x) {
  x$x <- -0.5
  x
}
to_right <- function(x) {
  x$x <- 1.5
  x
}

tween_events(d, 'cubic-in-out', 50, start = time, end = time + duration,
             enter = from_left, exit = to_right, enter_length = 0.1,
             exit_length = 0.05)

#>                x          y        time    duration .id .phase .frame
#> 68  -0.500000000 0.35953067 0.008658075 0.040885388   8  enter      1
#> 126 -0.500000000 0.02488603 0.043001293 0.083862999  14  enter      2
#> 69  -0.422311360 0.35953067 0.008658075 0.040885388   8  enter      2
#> 127 -0.452590981 0.02488603 0.043001293 0.083862999  14  enter      3
#> 177 -0.500000000 0.87072715 0.064883093 0.034442903  19  enter      3
#> 70   0.121509119 0.35953067 0.008658075 0.040885388   8  enter      3
#> 128 -0.120727847 0.02488603 0.043001293 0.083862999  14  enter      4
#> 178 -0.452966660 0.87072715 0.064883093 0.034442903  19  enter      4
#> 71   0.665329598 0.35953067 0.008658075 0.040885388   8  enter      4
#> 129  0.602259696 0.02488603 0.043001293 0.083862999  14  enter      5
#> 179 -0.123733276 0.87072715 0.064883093 0.034442903  19  enter      5
#> 9   -0.500000000 0.53613326 0.109670019 0.013350766   2  enter      5
#> 72   0.743018237 0.35953067 0.008658075 0.040885388   8 static      5
#> 130  0.934122830 0.02488603 0.043001293 0.083862999  14  enter      6
#> 180  0.593525166 0.87072715 0.064883093 0.034442903  19  enter      6
#> 10  -0.466873037 0.53613326 0.109670019 0.013350766   2  enter      6
#> 73   0.743018237 0.35953067 0.008658075 0.040885388   8 static      6
#> 115 -0.500000000 0.71341195 0.156360570 0.085928589  13  enter      7
#> 131  0.981531849 0.02488603 0.043001293 0.083862999  14 static      7
#> 181  0.922758550 0.87072715 0.064883093 0.034442903  19  enter      7
#> 11  -0.234984297 0.53613326 0.109670019 0.013350766   2  enter      7
#> 74   0.743018237 0.35953067 0.008658075 0.040885388   8    raw      7
#> 1   -0.500000000 0.98456547 0.173224842 0.007838425   1  enter      8
#> 116 -0.469501144 0.71341195 0.156360570 0.085928589  13  enter      8
#> 132  0.981531849 0.02488603 0.043001293 0.083862999  14 static      8
#> 182  0.969791890 0.87072715 0.064883093 0.034442903  19 static      8
#> 12   0.270201886 0.53613326 0.109670019 0.013350766   2  enter      8
#> 75   1.121509119 0.35953067 0.008658075 0.040885388   8   exit      8
#> 2   -0.408072186 0.98456547 0.173224842 0.007838425   1  enter      9
#> 117 -0.256009156 0.71341195 0.156360570 0.085928589  13  enter      9
#> 133  0.981531849 0.02488603 0.043001293 0.083862999  14 static      9
#> 183  0.969791890 0.87072715 0.064883093 0.034442903  19    raw      9
#> 13   0.502090625 0.53613326 0.109670019 0.013350766   2  enter      9
#> 76   1.500000000 0.35953067 0.008658075 0.040885388   8   exit      9
#> 3    0.235422512 0.98456547 0.173224842 0.007838425   1  enter     10
#> 118  0.209098391 0.71341195 0.156360570 0.085928589  13  enter     10
#> 134  0.981531849 0.02488603 0.043001293 0.083862999  14    raw     10
#> 184  1.234895945 0.87072715 0.064883093 0.034442903  19   exit     10
#> 14   0.535217588 0.53613326 0.109670019 0.013350766   2    raw     10
#> 4    0.878917211 0.98456547 0.173224842 0.007838425   1  enter     11
#> 119  0.422590380 0.71341195 0.156360570 0.085928589  13  enter     11
#> 135  1.240765925 0.02488603 0.043001293 0.083862999  14   exit     11
#> 185  1.500000000 0.87072715 0.064883093 0.034442903  19   exit     11
#> 15   1.017608794 0.53613326 0.109670019 0.013350766   2   exit     11
#> 5    0.970845025 0.98456547 0.173224842 0.007838425   1 static     12
#> 120  0.453089235 0.71341195 0.156360570 0.085928589  13 static     12
#> 136  1.500000000 0.02488603 0.043001293 0.083862999  14   exit     12
#> 16   1.500000000 0.53613326 0.109670019 0.013350766   2   exit     12
#> 6    0.970845025 0.98456547 0.173224842 0.007838425   1    raw     13
#> 121  0.453089235 0.71341195 0.156360570 0.085928589  13 static     13
#> 7    1.235422512 0.98456547 0.173224842 0.007838425   1   exit     14
#> 122  0.453089235 0.71341195 0.156360570 0.085928589  13 static     14
#> 8    1.500000000 0.98456547 0.173224842 0.007838425   1   exit     15
#> 123  0.453089235 0.71341195 0.156360570 0.085928589  13    raw     15
#> 124  0.976544618 0.71341195 0.156360570 0.085928589  13   exit     16
#> 57  -0.500000000 0.37896074 0.350588162 0.098266867   7  enter     16
#> 125  1.500000000 0.71341195 0.156360570 0.085928589  13   exit     17
#> 58  -0.431573800 0.37896074 0.350588162 0.098266867   7  enter     17
#> 96  -0.500000000 0.17903429 0.412256387 0.014355923  11  enter     18
#> 59   0.047409601 0.37896074 0.350588162 0.098266867   7  enter     18
#> 97  -0.483528998 0.17903429 0.412256387 0.014355923  11  enter     19
#> 60   0.526393001 0.37896074 0.350588162 0.098266867   7  enter     19
#> 77  -0.500000000 0.84423142 0.430898793 0.036087369   9  enter     19
#> 98  -0.368231987 0.17903429 0.412256387 0.014355923  11  enter     20
#> 61   0.594819201 0.37896074 0.350588162 0.098266867   7 static     20
#> 78  -0.435982492 0.84423142 0.430898793 0.036087369   9  enter     20
#> 99  -0.117049211 0.17903429 0.412256387 0.014355923  11  enter     21
#> 62   0.594819201 0.37896074 0.350588162 0.098266867   7 static     21
#> 79   0.012140067 0.84423142 0.430898793 0.036087369   9  enter     21
#> 100 -0.001752199 0.17903429 0.412256387 0.014355923  11  enter     22
#> 63   0.594819201 0.37896074 0.350588162 0.098266867   7 static     22
#> 80   0.460262626 0.84423142 0.430898793 0.036087369   9  enter     22
#> 101  0.014718803 0.17903429 0.412256387 0.014355923  11    raw     23
#> 148 -0.500000000 0.82792132 0.514402497 0.023769774  16  enter     23
#> 64   0.594819201 0.37896074 0.350588162 0.098266867   7 static     23
#> 81   0.524280134 0.84423142 0.430898793 0.036087369   9 static     23
#> 102  0.757359401 0.17903429 0.412256387 0.014355923  11   exit     24
#> 104 -0.500000000 0.96620920 0.550862395 0.068079862  12  enter     24
#> 149 -0.446247350 0.82792132 0.514402497 0.023769774  16  enter     24
#> 65   0.594819201 0.37896074 0.350588162 0.098266867   7    raw     24
#> 82   0.524280134 0.84423142 0.430898793 0.036087369   9 static     24
#> 103  1.500000000 0.17903429 0.412256387 0.014355923  11   exit     25
#> 105 -0.408880055 0.96620920 0.550862395 0.068079862  12  enter     25
#> 150 -0.069978799 0.82792132 0.514402497 0.023769774  16  enter     25
#> 66   1.047409601 0.37896074 0.350588162 0.098266867   7   exit     25
#> 83   0.524280134 0.84423142 0.430898793 0.036087369   9    raw     25
#> 106  0.228959562 0.96620920 0.550862395 0.068079862  12  enter     26
#> 151  0.306289751 0.82792132 0.514402497 0.023769774  16  enter     26
#> 17  -0.500000000 0.76105941 0.586222192 0.080969620   3  enter     26
#> 67   1.500000000 0.37896074 0.350588162 0.098266867   7   exit     26
#> 84   1.012140067 0.84423142 0.430898793 0.036087369   9   exit     26
#> 107  0.866799179 0.96620920 0.550862395 0.068079862  12  enter     27
#> 152  0.360042401 0.82792132 0.514402497 0.023769774  16 static     27
#> 18  -0.458344327 0.76105941 0.586222192 0.080969620   3  enter     27
#> 85   1.500000000 0.84423142 0.430898793 0.036087369   9   exit     27
#> 108  0.957919125 0.96620920 0.550862395 0.068079862  12 static     28
#> 153  0.360042401 0.82792132 0.514402497 0.023769774  16    raw     28
#> 19  -0.166754616 0.76105941 0.586222192 0.080969620   3  enter     28
#> 109  0.957919125 0.96620920 0.550862395 0.068079862  12 static     29
#> 154  0.930021201 0.82792132 0.514402497 0.023769774  16   exit     29
#> 20   0.124835095 0.76105941 0.586222192 0.080969620   3  enter     29
#> 110  0.957919125 0.96620920 0.550862395 0.068079862  12 static     30
#> 137 -0.500000000 0.62645649 0.684644113 0.088131292  15  enter     30
#> 155  1.500000000 0.82792132 0.514402497 0.023769774  16   exit     30
#> 21   0.166490768 0.76105941 0.586222192 0.080969620   3 static     30
#> 49  -0.500000000 0.63906099 0.675661902 0.034969548   6  enter     30
#> 111  0.957919125 0.96620920 0.550862395 0.068079862  12    raw     31
#> 138 -0.434705927 0.62645649 0.684644113 0.088131292  15  enter     31
#> 22   0.166490768 0.76105941 0.586222192 0.080969620   3 static     31
#> 50  -0.429825074 0.63906099 0.675661902 0.034969548   6  enter     31
#> 86  -0.500000000 0.45674918 0.724935097 0.065251004  10  enter     32
#> 112  1.038227402 0.96620920 0.550862395 0.068079862  12   exit     32
#> 139  0.022352583 0.62645649 0.684644113 0.088131292  15  enter     32
#> 23   0.166490768 0.76105941 0.586222192 0.080969620   3 static     32
#> 51   0.061399411 0.63906099 0.675661902 0.034969548   6  enter     32
#> 87  -0.437698422 0.45674918 0.724935097 0.065251004  10  enter     33
#> 113  1.419691722 0.96620920 0.550862395 0.068079862  12   exit     33
#> 140  0.479411093 0.62645649 0.684644113 0.088131292  15  enter     33
#> 186 -0.500000000 0.28181258 0.762110915 0.076085823  20  enter     33
#> 24   0.166490768 0.76105941 0.586222192 0.080969620   3    raw     33
#> 52   0.552623895 0.63906099 0.675661902 0.034969548   6  enter     33
#> 88  -0.001587375 0.45674918 0.724935097 0.065251004  10  enter     34
#> 114  1.500000000 0.96620920 0.550862395 0.068079862  12   exit     34
#> 141  0.544705166 0.62645649 0.684644113 0.088131292  15 static     34
#> 187 -0.463247958 0.28181258 0.762110915 0.076085823  20  enter     34
#> 25   0.364047691 0.76105941 0.586222192 0.080969620   3   exit     34
#> 53   0.622798821 0.63906099 0.675661902 0.034969548   6 static     34
#> 89   0.434523672 0.45674918 0.724935097 0.065251004  10  enter     35
#> 142  0.544705166 0.62645649 0.684644113 0.088131292  15 static     35
#> 188 -0.205983664 0.28181258 0.762110915 0.076085823  20  enter     35
#> 26   1.302443077 0.76105941 0.586222192 0.080969620   3   exit     35
#> 54   0.622798821 0.63906099 0.675661902 0.034969548   6    raw     35
#> 90   0.496825250 0.45674918 0.724935097 0.065251004  10 static     36
#> 143  0.544705166 0.62645649 0.684644113 0.088131292  15 static     36
#> 189  0.354484977 0.28181258 0.762110915 0.076085823  20  enter     36
#> 27   1.500000000 0.76105941 0.586222192 0.080969620   3   exit     36
#> 55   1.061399411 0.63906099 0.675661902 0.034969548   6   exit     36
#> 91   0.496825250 0.45674918 0.724935097 0.065251004  10 static     37
#> 144  0.544705166 0.62645649 0.684644113 0.088131292  15 static     37
#> 166 -0.500000000 0.70848426 0.843096942 0.089790251  18  enter     37
#> 190  0.611749272 0.28181258 0.762110915 0.076085823  20  enter     37
#> 28  -0.500000000 0.56438573 0.850721327 0.055938324   4  enter     37
#> 56   1.500000000 0.63906099 0.675661902 0.034969548   6   exit     37
#> 92   0.496825250 0.45674918 0.724935097 0.065251004  10 static     38
#> 145  0.544705166 0.62645649 0.684644113 0.088131292  15    raw     38
#> 167 -0.464085398 0.70848426 0.843096942 0.089790251  18  enter     38
#> 191  0.648501314 0.28181258 0.762110915 0.076085823  20 static     38
#> 29  -0.465114022 0.56438573 0.850721327 0.055938324   4  enter     38
#> 93   0.496825250 0.45674918 0.724935097 0.065251004  10    raw     39
#> 146  1.022352583 0.62645649 0.684644113 0.088131292  15   exit     39
#> 168 -0.212683184 0.70848426 0.843096942 0.089790251  18  enter     39
#> 192  0.648501314 0.28181258 0.762110915 0.076085823  20 static     39
#> 30  -0.220912175 0.56438573 0.850721327 0.055938324   4  enter     39
#> 94   0.998412625 0.45674918 0.724935097 0.065251004  10   exit     40
#> 147  1.500000000 0.62645649 0.684644113 0.088131292  15   exit     40
#> 169  0.038719030 0.70848426 0.843096942 0.089790251  18  enter     40
#> 193  0.648501314 0.28181258 0.762110915 0.076085823  20 static     40
#> 31   0.023289672 0.56438573 0.850721327 0.055938324   4  enter     40
#> 38  -0.500000000 0.77845430 0.921526176 0.073372969   5  enter     40
#> 95   1.500000000 0.45674918 0.724935097 0.065251004  10   exit     41
#> 156 -0.500000000 0.64306903 0.945687848 0.057863085  17  enter     41
#> 170  0.074633632 0.70848426 0.843096942 0.089790251  18 static     41
#> 194  0.648501314 0.28181258 0.762110915 0.076085823  20    raw     41
#> 32   0.058175651 0.56438573 0.850721327 0.055938324   4 static     41
#> 39  -0.440053811 0.77845430 0.921526176 0.073372969   5  enter     41
#> 157 -0.437303368 0.64306903 0.945687848 0.057863085  17  enter     42
#> 171  0.074633632 0.70848426 0.843096942 0.089790251  18 static     42
#> 195  1.074250657 0.28181258 0.762110915 0.076085823  20   exit     42
#> 33   0.058175651 0.56438573 0.850721327 0.055938324   4 static     42
#> 40  -0.020430490 0.77845430 0.921526176 0.073372969   5  enter     42
#> 158  0.001573052 0.64306903 0.945687848 0.057863085  17  enter     43
#> 172  0.074633632 0.70848426 0.843096942 0.089790251  18 static     43
#> 196  1.500000000 0.28181258 0.762110915 0.076085823  20   exit     43
#> 34   0.058175651 0.56438573 0.850721327 0.055938324   4 static     43
#> 41   0.399192830 0.77845430 0.921526176 0.073372969   5  enter     43
#> 159  0.440449473 0.64306903 0.945687848 0.057863085  17  enter     44
#> 173  0.074633632 0.70848426 0.843096942 0.089790251  18 static     44
#> 35   0.058175651 0.56438573 0.850721327 0.055938324   4    raw     44
#> 42   0.459139019 0.77845430 0.921526176 0.073372969   5 static     44
#> 160  0.503146105 0.64306903 0.945687848 0.057863085  17 static     45
#> 174  0.074633632 0.70848426 0.843096942 0.089790251  18    raw     45
#> 36   0.779087825 0.56438573 0.850721327 0.055938324   4   exit     45
#> 43   0.459139019 0.77845430 0.921526176 0.073372969   5 static     45
#> 161  0.503146105 0.64306903 0.945687848 0.057863085  17 static     46
#> 175  0.787316816 0.70848426 0.843096942 0.089790251  18   exit     46
#> 37   1.500000000 0.56438573 0.850721327 0.055938324   4   exit     46
#> 44   0.459139019 0.77845430 0.921526176 0.073372969   5 static     46
#> 162  0.503146105 0.64306903 0.945687848 0.057863085  17 static     47
#> 176  1.500000000 0.70848426 0.843096942 0.089790251  18   exit     47
#> 45   0.459139019 0.77845430 0.921526176 0.073372969   5    raw     47
#> 163  0.503146105 0.64306903 0.945687848 0.057863085  17    raw     48
#> 46   0.613340646 0.77845430 0.921526176 0.073372969   5   exit     48
#> 164  1.001573052 0.64306903 0.945687848 0.057863085  17   exit     49
#> 47   1.345798373 0.77845430 0.921526176 0.073372969   5   exit     49
#> 165  1.500000000 0.64306903 0.945687848 0.057863085  17   exit     50
#> 48   1.500000000 0.77845430 0.921526176 0.073372969   5   exit     50