Transition in and out of events

This tweening function is a more powerful version of tween_appear(), with support for newer features such as enter/exits and tween phase identification. The tweener treats each row in the data as unique events in time, and creates frames with the correct events present at any given time.

tween_events(
  .data,
  ease,
  nframes,
  start,
  end = NULL,
  range = NULL,
  enter = NULL,
  exit = NULL,
  enter_length = 0,
  exit_length = 0
)

Arguments

.data

A data.frame with components at different stages

ease

The easing function to use. Either a single string or one for each column in the data set.

nframes

The number of frames to calculate for the tween

start, end

The start (and potential end) of the event encoded in the row, as unquoted expressions. Will be evaluated in the context of .data so can refer to columns in it. If end = NULL the event will be without extend and only visible in a single frame, unless enter and/or exit is given.

range

The range of time points to include in the tween. If NULL it will use the range of time

enter, exit

functions that calculate a start state for new observations that appear in to or an end state for observations that are not present in to. If NULL the new/old observations will not be part of the tween. The function gets a data.frame with either the start state of the exiting observations, or the end state of the entering observations and must return a modified version of that data.frame. See the Match, Enter, and Exit section for more information.

enter_length, exit_length

The lenght of the opening and closing transitions if enter and/or exit is given. Measured in the same units as time

Value

A data.frame with the same columns as .data along with .id giving the component id, .phase giving the state of each component in each frame, and .frame giving the frame membership of each row.

See also

Examples

d <- data.frame(
  x = runif(20),
  y = runif(20),
  time = runif(20),
  duration = runif(20, max = 0.1)
)
from_left <- function(x) {
  x$x <- -0.5
  x
}
to_right <- function(x) {
  x$x <- 1.5
  x
}

tween_events(d, 'cubic-in-out', 50, start = time, end = time + duration,
             enter = from_left, exit = to_right, enter_length = 0.1,
             exit_length = 0.05)
#>                 x          y      time    duration .id .phase .frame
#> 192 -0.5000000000 0.74210768 0.1060838 0.076233523  20  enter      1
#> 130 -0.5000000000 0.84186199 0.1355643 0.015531769  14  enter      2
#> 193 -0.4376195293 0.74210768 0.1060838 0.076233523  20  enter      2
#> 131 -0.4582698719 0.84186199 0.1355643 0.015531769  14  enter      3
#> 194 -0.0009562345 0.74210768 0.1060838 0.076233523  20  enter      3
#> 132 -0.1661589752 0.84186199 0.1355643 0.015531769  14  enter      4
#> 195  0.4357070603 0.74210768 0.1060838 0.076233523  20  enter      4
#> 133  0.4702254782 0.84186199 0.1355643 0.015531769  14  enter      5
#> 146 -0.5000000000 0.13193141 0.2029678 0.086073534  16  enter      5
#> 196  0.4980875310 0.74210768 0.1060838 0.076233523  20 static      5
#> 134  0.7623363749 0.84186199 0.1355643 0.015531769  14  enter      6
#> 147 -0.4522311136 0.13193141 0.2029678 0.086073534  16  enter      6
#> 197  0.4980875310 0.74210768 0.1060838 0.076233523  20 static      6
#> 135  0.8040665030 0.84186199 0.1355643 0.015531769  14    raw      7
#> 148 -0.1178489086 0.13193141 0.2029678 0.086073534  16  enter      7
#> 198  0.4980875310 0.74210768 0.1060838 0.076233523  20 static      7
#> 11  -0.5000000000 0.37074425 0.2573455 0.058254561   2  enter      8
#> 136  1.1520332515 0.84186199 0.1355643 0.015531769  14   exit      8
#> 149  0.6106266094 0.13193141 0.2029678 0.086073534  16  enter      8
#> 199  0.4980875310 0.74210768 0.1060838 0.076233523  20 static      8
#> 12  -0.4502811613 0.37074425 0.2573455 0.058254561   2  enter      9
#> 137  1.5000000000 0.84186199 0.1355643 0.015531769  14   exit      9
#> 150  0.9450088143 0.13193141 0.2029678 0.086073534  16  enter      9
#> 200  0.4980875310 0.74210768 0.1060838 0.076233523  20    raw      9
#> 1   -0.5000000000 0.02915020 0.3099438 0.071081272   1  enter     10
#> 13  -0.1022492908 0.37074425 0.2573455 0.058254561   2  enter     10
#> 151  0.9927777008 0.13193141 0.2029678 0.086073534  16 static     10
#> 201  0.9990437655 0.74210768 0.1060838 0.076233523  20   exit     10
#> 2   -0.4506494502 0.02915020 0.3099438 0.071081272   1  enter     11
#> 14   0.2457825798 0.37074425 0.2573455 0.058254561   2  enter     11
#> 152  0.9927777008 0.13193141 0.2029678 0.086073534  16 static     11
#> 202  1.5000000000 0.74210768 0.1060838 0.076233523  20   exit     11
#> 3   -0.1051956015 0.02915020 0.3099438 0.071081272   1  enter     12
#> 15   0.2955014184 0.37074425 0.2573455 0.058254561   2 static     12
#> 153  0.9927777008 0.13193141 0.2029678 0.086073534  16 static     12
#> 4    0.2402582472 0.02915020 0.3099438 0.071081272   1  enter     13
#> 16   0.2955014184 0.37074425 0.2573455 0.058254561   2 static     13
#> 154  0.9927777008 0.13193141 0.2029678 0.086073534  16    raw     13
#> 169 -0.5000000000 0.60957281 0.3823714 0.087586088  18  enter     13
#> 5    0.2896087971 0.02915020 0.3099438 0.071081272   1 static     14
#> 17   0.2955014184 0.37074425 0.2573455 0.058254561   2    raw     14
#> 155  1.2463888504 0.13193141 0.2029678 0.086073534  16   exit     14
#> 170 -0.4377403734 0.60957281 0.3823714 0.087586088  18  enter     14
#> 6    0.2896087971 0.02915020 0.3099438 0.071081272   1 static     15
#> 18   0.4739456527 0.37074425 0.2573455 0.058254561   2   exit     15
#> 156  1.5000000000 0.13193141 0.2029678 0.086073534  16   exit     15
#> 171 -0.0019229875 0.60957281 0.3823714 0.087586088  18  enter     15
#> 7    0.2896087971 0.02915020 0.3099438 0.071081272   1 static     16
#> 19   1.3215557657 0.37074425 0.2573455 0.058254561   2   exit     16
#> 21  -0.5000000000 0.01310343 0.4568395 0.061947737   3  enter     16
#> 49  -0.5000000000 0.02235420 0.4591535 0.070469290   6  enter     16
#> 172  0.4338943984 0.60957281 0.3823714 0.087586088  18  enter     16
#> 8    0.2896087971 0.02915020 0.3099438 0.071081272   1    raw     17
#> 20   1.5000000000 0.37074425 0.2573455 0.058254561   2   exit     17
#> 22  -0.4782737690 0.01310343 0.4568395 0.061947737   3  enter     17
#> 50  -0.4675566354 0.02235420 0.4591535 0.070469290   6  enter     17
#> 60  -0.5000000000 0.61546859 0.4794142 0.072976026   7  enter     17
#> 173  0.4961540250 0.60957281 0.3823714 0.087586088  18 static     17
#> 9    0.8948043985 0.02915020 0.3099438 0.071081272   1   exit     18
#> 23  -0.3261901523 0.01310343 0.4568395 0.061947737   3  enter     18
#> 39  -0.5000000000 0.78916052 0.4908285 0.064142768   5  enter     18
#> 51  -0.2404530832 0.02235420 0.4591535 0.070469290   6  enter     18
#> 61  -0.4591988085 0.61546859 0.4794142 0.072976026   7  enter     18
#> 174  0.4961540250 0.60957281 0.3823714 0.087586088  18 static     18
#> 10   1.5000000000 0.02915020 0.3099438 0.071081272   1   exit     19
#> 24   0.0051348699 0.01310343 0.4568395 0.061947737   3  enter     19
#> 40  -0.4194975921 0.78916052 0.4908285 0.064142768   5  enter     19
#> 52   0.2543082271 0.02235420 0.4591535 0.070469290   6  enter     19
#> 62  -0.1735904678 0.61546859 0.4794142 0.072976026   7  enter     19
#> 80  -0.5000000000 0.57099034 0.5123498 0.078517180   9  enter     19
#> 175  0.4961540250 0.60957281 0.3823714 0.087586088  18 static     19
#> 25   0.1572184867 0.01310343 0.4568395 0.061947737   3  enter     20
#> 41   0.1440192636 0.78916052 0.4908285 0.064142768   5  enter     20
#> 53   0.4814117793 0.02235420 0.4591535 0.070469290   6  enter     20
#> 63   0.4486277031 0.61546859 0.4794142 0.072976026   7  enter     20
#> 81  -0.4371257407 0.57099034 0.5123498 0.078517180   9  enter     20
#> 157 -0.5000000000 0.43182609 0.5471362 0.095980578  17  enter     20
#> 176  0.4961540250 0.60957281 0.3823714 0.087586088  18 static     20
#> 26   0.1789447176 0.01310343 0.4568395 0.061947737   3 static     21
#> 42   0.7075361192 0.78916052 0.4908285 0.064142768   5  enter     21
#> 54   0.5138551439 0.02235420 0.4591535 0.070469290   6 static     21
#> 64   0.7342360438 0.61546859 0.4794142 0.072976026   7  enter     21
#> 71  -0.5000000000 0.56272414 0.5784501 0.027349744   8  enter     21
#> 82   0.0029940743 0.57099034 0.5123498 0.078517180   9  enter     21
#> 138 -0.5000000000 0.81434363 0.5694733 0.006569973  15  enter     21
#> 158 -0.4513895776 0.43182609 0.5471362 0.095980578  17  enter     21
#> 177  0.4961540250 0.60957281 0.3823714 0.087586088  18    raw     21
#> 27   0.1789447176 0.01310343 0.4568395 0.061947737   3 static     22
#> 43   0.7880385271 0.78916052 0.4908285 0.064142768   5 static     22
#> 55   0.5138551439 0.02235420 0.4591535 0.070469290   6 static     22
#> 65   0.7750372353 0.61546859 0.4794142 0.072976026   7 static     22
#> 72  -0.4526630827 0.56272414 0.5784501 0.027349744   8  enter     22
#> 83   0.4431138892 0.57099034 0.5123498 0.078517180   9  enter     22
#> 139 -0.4212076831 0.81434363 0.5694733 0.006569973  15  enter     22
#> 159 -0.1111166212 0.43182609 0.5471362 0.095980578  17  enter     22
#> 178  0.9980770125 0.60957281 0.3823714 0.087586088  18   exit     22
#> 28   0.1789447176 0.01310343 0.4568395 0.061947737   3    raw     23
#> 44   0.7880385271 0.78916052 0.4908285 0.064142768   5 static     23
#> 56   0.5138551439 0.02235420 0.4591535 0.070469290   6 static     23
#> 66   0.7750372353 0.61546859 0.4794142 0.072976026   7 static     23
#> 73  -0.1213046619 0.56272414 0.5784501 0.027349744   8  enter     23
#> 84   0.5059881485 0.57099034 0.5123498 0.078517180   9 static     23
#> 140  0.1303385349 0.81434363 0.5694733 0.006569973  15  enter     23
#> 160  0.2291563353 0.43182609 0.5471362 0.095980578  17  enter     23
#> 179  1.5000000000 0.60957281 0.3823714 0.087586088  18   exit     23
#> 29   0.8394723588 0.01310343 0.4568395 0.061947737   3   exit     24
#> 45   0.7880385271 0.78916052 0.4908285 0.064142768   5 static     24
#> 57   0.5138551439 0.02235420 0.4591535 0.070469290   6    raw     24
#> 67   0.7750372353 0.61546859 0.4794142 0.072976026   7 static     24
#> 74   0.6005833263 0.56272414 0.5784501 0.027349744   8  enter     24
#> 85   0.5059881485 0.57099034 0.5123498 0.078517180   9 static     24
#> 141  0.6818847530 0.81434363 0.5694733 0.006569973  15  enter     24
#> 161  0.2777667576 0.43182609 0.5471362 0.095980578  17 static     24
#> 30   1.5000000000 0.01310343 0.4568395 0.061947737   3   exit     25
#> 46   0.7880385271 0.78916052 0.4908285 0.064142768   5    raw     25
#> 58   1.0069275720 0.02235420 0.4591535 0.070469290   6   exit     25
#> 68   0.7750372353 0.61546859 0.4794142 0.072976026   7    raw     25
#> 75   0.9319417471 0.56272414 0.5784501 0.027349744   8  enter     25
#> 86   0.5059881485 0.57099034 0.5123498 0.078517180   9 static     25
#> 142  0.7606770699 0.81434363 0.5694733 0.006569973  15 static     25
#> 162  0.2777667576 0.43182609 0.5471362 0.095980578  17 static     25
#> 47   1.1440192636 0.78916052 0.4908285 0.064142768   5   exit     26
#> 59   1.5000000000 0.02235420 0.4591535 0.070469290   6   exit     26
#> 69   1.1375186177 0.61546859 0.4794142 0.072976026   7   exit     26
#> 76   0.9792786643 0.56272414 0.5784501 0.027349744   8 static     26
#> 87   0.5059881485 0.57099034 0.5123498 0.078517180   9    raw     26
#> 143  0.7606770699 0.81434363 0.5694733 0.006569973  15    raw     26
#> 163  0.2777667576 0.43182609 0.5471362 0.095980578  17 static     26
#> 48   1.5000000000 0.78916052 0.4908285 0.064142768   5   exit     27
#> 70   1.5000000000 0.61546859 0.4794142 0.072976026   7   exit     27
#> 77   0.9792786643 0.56272414 0.5784501 0.027349744   8    raw     27
#> 88   0.6532491635 0.57099034 0.5123498 0.078517180   9   exit     27
#> 144  1.1303385349 0.81434363 0.5694733 0.006569973  15   exit     27
#> 164  0.2777667576 0.43182609 0.5471362 0.095980578  17 static     27
#> 78   1.2396393322 0.56272414 0.5784501 0.027349744   8   exit     28
#> 89   1.3527389850 0.57099034 0.5123498 0.078517180   9   exit     28
#> 145  1.5000000000 0.81434363 0.5694733 0.006569973  15   exit     28
#> 165  0.2777667576 0.43182609 0.5471362 0.095980578  17 static     28
#> 79   1.5000000000 0.56272414 0.5784501 0.027349744   8   exit     29
#> 90   1.5000000000 0.57099034 0.5123498 0.078517180   9   exit     29
#> 166  0.2777667576 0.43182609 0.5471362 0.095980578  17    raw     29
#> 167  0.8888833788 0.43182609 0.5471362 0.095980578  17   exit     30
#> 168  1.5000000000 0.43182609 0.5471362 0.095980578  17   exit     31
#> 101 -0.5000000000 0.53498550 0.8463406 0.054999120  11  enter     33
#> 102 -0.4563503074 0.53498550 0.8463406 0.054999120  11  enter     34
#> 119 -0.5000000000 0.11220753 0.8675427 0.087706898  13  enter     34
#> 103 -0.1508024596 0.53498550 0.8463406 0.054999120  11  enter     35
#> 120 -0.4491604895 0.11220753 0.8675427 0.087706898  13  enter     35
#> 31  -0.5000000000 0.80065954 0.9144442 0.009395654   4  enter     36
#> 104  0.1547453883 0.53498550 0.8463406 0.054999120  11  enter     36
#> 121 -0.0932839158 0.11220753 0.8675427 0.087706898  13  enter     36
#> 32  -0.4659311316 0.80065954 0.9144442 0.009395654   4  enter     37
#> 91  -0.5000000000 0.13207049 0.9326675 0.051440555  10  enter     37
#> 105  0.1983950809 0.53498550 0.8463406 0.054999120  11 static     37
#> 111 -0.5000000000 0.89896792 0.9391653 0.018461245  12  enter     37
#> 122  0.2625926579 0.11220753 0.8675427 0.087706898  13  enter     37
#> 33  -0.2274490525 0.80065954 0.9144442 0.009395654   4  enter     38
#> 92  -0.4247907031 0.13207049 0.9326675 0.051440555  10  enter     38
#> 106  0.1983950809 0.53498550 0.8463406 0.054999120  11 static     38
#> 112 -0.4612876245 0.89896792 0.9391653 0.018461245  12  enter     38
#> 123  0.3134321684 0.11220753 0.8675427 0.087706898  13 static     38
#> 34   0.0110330266 0.80065954 0.9144442 0.009395654   4  enter     39
#> 93   0.1016743752 0.13207049 0.9326675 0.051440555  10  enter     39
#> 107  0.1983950809 0.53498550 0.8463406 0.054999120  11 static     39
#> 113 -0.1903009962 0.89896792 0.9391653 0.018461245  12  enter     39
#> 124  0.3134321684 0.11220753 0.8675427 0.087706898  13 static     39
#> 180 -0.5000000000 0.56353676 0.9857591 0.099979275  19  enter     39
#> 35   0.0451018950 0.80065954 0.9144442 0.009395654   4 static     40
#> 94   0.6281394536 0.13207049 0.9326675 0.051440555  10  enter     40
#> 108  0.1983950809 0.53498550 0.8463406 0.054999120  11    raw     40
#> 114  0.0806856322 0.89896792 0.9391653 0.018461245  12  enter     40
#> 125  0.3134321684 0.11220753 0.8675427 0.087706898  13 static     40
#> 181 -0.4106656460 0.56353676 0.9857591 0.099979275  19  enter     40
#> 36   0.0451018950 0.80065954 0.9144442 0.009395654   4    raw     41
#> 95   0.7033487505 0.13207049 0.9326675 0.051440555  10 static     41
#> 109  0.8491975404 0.53498550 0.8463406 0.054999120  11   exit     41
#> 115  0.1193980076 0.89896792 0.9391653 0.018461245  12 static     41
#> 126  0.3134321684 0.11220753 0.8675427 0.087706898  13 static     41
#> 182  0.2146748317 0.56353676 0.9857591 0.099979275  19  enter     41
#> 37   0.7725509475 0.80065954 0.9144442 0.009395654   4   exit     42
#> 96   0.7033487505 0.13207049 0.9326675 0.051440555  10 static     42
#> 110  1.5000000000 0.53498550 0.8463406 0.054999120  11   exit     42
#> 116  0.1193980076 0.89896792 0.9391653 0.018461245  12    raw     42
#> 127  0.3134321684 0.11220753 0.8675427 0.087706898  13    raw     42
#> 183  0.8400153095 0.56353676 0.9857591 0.099979275  19  enter     42
#> 38   1.5000000000 0.80065954 0.9144442 0.009395654   4   exit     43
#> 97   0.7033487505 0.13207049 0.9326675 0.051440555  10    raw     43
#> 117  0.8096990038 0.89896792 0.9391653 0.018461245  12   exit     43
#> 128  0.9067160842 0.11220753 0.8675427 0.087706898  13   exit     43
#> 184  0.9293496634 0.56353676 0.9857591 0.099979275  19 static     43
#> 98   0.8213711578 0.13207049 0.9326675 0.051440555  10   exit     44
#> 118  1.5000000000 0.89896792 0.9391653 0.018461245  12   exit     44
#> 129  1.5000000000 0.11220753 0.8675427 0.087706898  13   exit     44
#> 185  0.9293496634 0.56353676 0.9857591 0.099979275  19 static     44
#> 99   1.3819775927 0.13207049 0.9326675 0.051440555  10   exit     45
#> 186  0.9293496634 0.56353676 0.9857591 0.099979275  19 static     45
#> 100  1.5000000000 0.13207049 0.9326675 0.051440555  10   exit     46
#> 187  0.9293496634 0.56353676 0.9857591 0.099979275  19 static     46
#> 188  0.9293496634 0.56353676 0.9857591 0.099979275  19 static     47
#> 189  0.9293496634 0.56353676 0.9857591 0.099979275  19    raw     48
#> 190  1.2146748317 0.56353676 0.9857591 0.099979275  19   exit     49
#> 191  1.5000000000 0.56353676 0.9857591 0.099979275  19   exit     50